+++ pypet2bids/pypet2bids/dcm2niix4pet.py
@@ -34,7 +34,6 @@ import importlib
 import dotenv
 import logging
 
-
 try:
     import helper_functions
     import is_pet
@@ -43,7 +42,6 @@ except ModuleNotFoundError:
     import pypet2bids.is_pet as is_pet
 
 logger = logging.getLogger("pypet2bids")
-
 
 # fields to check for
 module_folder = Path(__file__).parent.resolve()
@@ -382,7 +380,7 @@ class Dcm2niix4PET:
                 logger.warning(f"Minimum version {minimum_version} of dcm2niix is recommended, found "
                                f"installed version {version[0]} at {self.dcm2niix_path}.")
 
-        #check if user provided a custom tempdir location
+        # check if user provided a custom tempdir location
         self.tempdir_location = tempdir_location
         self.image_folder = Path(image_folder)
         if destination_path:
@@ -474,7 +472,8 @@ class Dcm2niix4PET:
                             **self.additional_arguments))
 
             # check for any blood (tsv) data or otherwise in the given spreadsheet values
-            blood_tsv_columns = ['time', 'plasma_radioactivity', 'metabolite_parent_fraction', 'whole_blood_radioactivity']
+            blood_tsv_columns = ['time', 'plasma_radioactivity', 'metabolite_parent_fraction',
+                                 'whole_blood_radioactivity']
             blood_json_columns = ['PlasmaAvail', 'WholeBloodAvail', 'MetaboliteAvail', 'MetaboliteMethod',
                                   'MetaboliteRecoveryCorrectionApplied', 'DispersionCorrected']
 
@@ -667,7 +666,7 @@ class Dcm2niix4PET:
 
                     sidecar_json = JsonMAJ(json_path=str(created),
                                            bids_null=True,
-                                           update_values=self.additional_arguments) # load all supplied and now written sidecar data in
+                                           update_values=self.additional_arguments)  # load all supplied and now written sidecar data in
 
                     sidecar_json.update()
 
@@ -969,8 +968,8 @@ def check_meta_radio_inputs(kwargs: dict) -> dict:
             tmp = (InjectedRadioactivity * 10 ** 6) / (InjectedMass * 10 ** 6)
             if SpecificRadioactivity:
                 if SpecificRadioactivity != tmp:
-                    print(colored("WARNING inferred SpecificRadioactivity in Bq/g doesn't match InjectedRadioactivity "
-                                  "and InjectedMass, could be a unit issue", "yellow"))
+                    logger.warning("Inferred SpecificRadioactivity in Bq/g doesn't match InjectedRadioactivity "
+                                   "and InjectedMass, could be a unit issue", "yellow")
                 data_out['SpecificRadioactivity'] = SpecificRadioactivity
                 data_out['SpecificRadioactivityUnits'] = kwargs.get('SpecificRadioactivityUnityUnits', 'n/a')
             else:
@@ -991,8 +990,8 @@ def check_meta_radio_inputs(kwargs: dict) -> dict:
             tmp = ((InjectedRadioactivity * (10 ** 6) / SpecificRadioactivity) * (10 ** 6))
             if InjectedMass:
                 if InjectedMass != tmp:
-                    print(colored("WARNING Infered InjectedMass in ug doesn't match InjectedRadioactivity and "
-                                  "InjectedMass, could be a unit issue", "yellow"))
+                    logger.warning("Inferred InjectedMass in ug doesn't match InjectedRadioactivity and "
+                                   "InjectedMass, could be a unit issue", "yellow")
                 data_out['InjectedMass'] = InjectedMass
                 data_out['InjectedMassUnits'] = kwargs.get('InjectedMassUnits', 'n/a')
             else:
[…]
