### 2023-09-11 78c15a7
added schema, updated warnings for inferred to use logging (#225)
