+++ pypet2bids/pyproject.toml
@@ -1,18 +1,12 @@
 [tool.poetry]
 name = "pypet2bids"
-version = "1.2.4"
+version = "1.2.5"
 description = "A python implementation of an ECAT to BIDS converter."
 authors = ["anthony galassi <28850131+bendhouseart@users.noreply.github.com>"]
 license = "MIT"
 include = [
-    'pypet2bids/metadata/blood_metadata.json',
-    'pypet2bids/metadata/definitions.json',
-    'pypet2bids/metadata/dicom2bids.json',
-    'pypet2bids/metadata/PET_metadata.json',
-    'pypet2bids/metadata/PET_Radionuclide.mkd',
-    'pypet2bids/metadata/PET_reconstruction_methods.json',
+    'pypet2bids/metadata/*',
     'pypet2bids/pyproject.toml',
-    'pypet2bids/metadata/README',
     ]
 
 [tool.poetry.dependencies]
