### 2023-09-11 #5822 @emdupre
Thanks for the invitation, @Kevin-Mattheus-Moerman !

I have a direct COI with one of the coauthors (MN and I were in the same lab until Dec 2022), and I have upcoming publications with several other authors---but no direct involvement---in large consortium papers re: BIDS.

If you think this can be waived, I would need a week to process one of my current submissions that's approaching end of review. But given the recent (< 1 year) shared affiliation with coauthor MN, it might make more sense for me to decline. Let me know what you think, and thanks again !
