### 2023-09-12 5451c0d
Update get_recon_method.m
PET_reconstruction_methods.json location defined relative to ecat2nii.m location, no longer with respect to "matlab" folder level, which is problematic if "matlab" occurs twice in PET2BIDS folder structure (e.g., /usr/matlab/PET2BIDS/matlab/ecat2nii.m)
