+++ matlab/get_recon_method.m
@@ -21,8 +21,8 @@ subset    = [];
 
 %% load metadata from the library
 current    = which('ecat2nii.m');
-root       = current(1:strfind(current,'matlab')-1);
-jsontoload = fullfile(root,['metadata' filesep 'PET_reconstruction_methods.json']);
+pn_parts   = strsplit(fileparts(current), '/');
+jsontoload = fullfile(strjoin(pn_parts(1:end-1), '/'), 'metadata', 'PET_reconstruction_methods.json');
 if exist(jsontoload,'file')
     reconstruction_metadata = jsondecode(fileread(jsontoload));
     reconstruction_metadata = reconstruction_metadata.reconstruction_names;
