### 2023-09-21 f3bba9f
more robust (and dumb) way to get relative path
