+++ matlab/get_recon_method.m
@@ -20,9 +20,9 @@ iteration = [];
 subset    = [];
 
 %% load metadata from the library
-current    = which('ecat2nii.m');
-pn_parts   = strsplit(fileparts(current), filesep);
-jsontoload = fullfile(strjoin(pn_parts(1:end-1), '/'), 'metadata', 'PET_reconstruction_methods.json');
+current    = which('get_recon_method.m');
+root       = fileparts(fileparts(current));
+jsontoload = fullfile(root,['metadata' filesep 'PET_reconstruction_methods.json']);
 if exist(jsontoload,'file')
     reconstruction_metadata = jsondecode(fileread(jsontoload));
     reconstruction_metadata = reconstruction_metadata.reconstruction_names;
