### 2023-09-26 #5822 @britta-wstnr
@CPernet - hi!

First of all, sorry for the delay - I was sick last week.
I am the handling editor for your submission. The first step is to identify reviewers - if you have any ideas for potential reviewers, you can suggest them here --- please read the guidelines in the editorialbot's [comment](https://github.com/openjournals/joss-reviews/issues/5822#issue-1889030096) on how to search for reviewers and how to mention them here (_do not_ "@" them, please).

I will be back shortly! 🙂 🌱
