### 2023-09-26 #5822 @CPernet
No worries - academic life :-)

I have searched but haven't seen someone with the right background in the list -- note that it will be hard to find someone with no COI, apologies in advance.
- You could look for 1 reviewer among regular contributors of the whole BIDS project (if that doesn't conflict with journal policy, re coi --> https://github.com/bids-standard/bids-specification/graphs/contributors) as our software builds on that.
-  for PET expertise - most co-authors from the pre-BIDS consensus paper (2007 - https://journals.sagepub.com/doi/10.1038/sj.jcbfm.9600493#tab-contributors) haven't collaborated in a long time with any of us which should work  (authors 1, bob and 15, gitte - to exclude as they are co-author here)
