### 2023-09-29 f65858d
silenced extra warnings, checking spreadsheet inputs too
