### 2023-09-29 35a9be4
silenced extra warnings, checking spreadsheet inputs too (#228)
* silenced extra warnings, checking spreadsheet inputs too

* bump version

* update poetry.lock
