### 2023-10-18 #5822 @britta-wstnr
Hi @CPernet,
quick update: I am still searching for reviewers.
In the meanwhile, could you please double check the (potentially) missing DOIs and update them (see https://github.com/openjournals/joss-reviews/issues/5822#issuecomment-1768217530)? Thanks! 🙏
