### 2023-10-19 81c77f7
Update paper.bib
add missing doi
