+++ paper.bib
@@ -2,9 +2,10 @@
 	title = {The brain imaging data structure, a format for organizing and describing outputs of neuroimaging experiments},
 	volume = {3},
 	url = {https://doi.org/10.1038/sdata.2016.44},
+        doi = {10.1038/sdata.2016.44},
 	pages = {160044},
 	journaltitle = {Scientific Data},
-	shortjournal = {Scientific Data},
+	shortjournal = {Sci Data},
 	author = {Gorgolewski, Krzysztof J. and Auer, Tibor and Calhoun, Vince D. and Craddock, R. Cameron and Das, Samir and Duff, Eugene P. and Flandin, Guillaume and Ghosh, Satrajit S. and Glatard, Tristan and Halchenko, Yaroslav O. and Handwerker, Daniel A. and Hanke, Michael and Keator, David and Li, Xiangrui and Michael, Zachary and Maumet, Camille and Nichols, B. Nolan and Nichols, Thomas E. and Pellman, John and Poline, Jean-Baptiste and Rokem, Ariel and Schaefer, Gunnar and Sochat, Vanessa and Triplett, William and Turner, Jessica A. and Varoquaux, Gaël and Poldrack, Russell A.},
 	date = {2016-06-21},
 }
@@ -116,6 +117,7 @@ Publisher: Nature Publishing Group},
 	title = {{BIDScoin}: A User-Friendly Application to Convert Source Data to Brain Imaging Data Structure},
 	volume = {15},
 	issn = {1662-5196},
+	doi = {10.3389/fninf.2021.770608},
 	url = {https://www.frontiersin.org/articles/10.3389/fninf.2021.770608},
 	shorttitle = {{BIDScoin}},
 	journaltitle = {Frontiers in Neuroinformatics},
