### 2023-10-19 #5822 @CPernet
thx robot for finding dois :-) added now to the .bib
