### 2023-11-13 5cf48ea
added check for a default config file and template (#229)
* added check for a default config file and template

The template file is sourced by default from
pypet2bids/pypet2bids/template_json.json

This file sets default units for site/user constant
values, it's intended to most commonly be used to set the units
for PET blood TSV's

* bump version
