### 2023-11-13 e8e142f
Convert msec to sec for ECAT to BIDS json (#230)
* convert msec to sec when reading in from ecat
* Update contributors.md
* bump version
