### 2023-11-20 #6067 @editorialbot
[REVIEW]: PET2BIDS: a library for converting Positron Emission Tomography data to BIDS
**Submitting author:** @CPernet (<a href="http://orcid.org/0000-0003-4010-4632">Cyril Pernet</a>)
**Repository:** https://github.com/openneuropet/PET2BIDS/
**Branch with paper.md** (empty if default branch): JOSS_paper
**Version:** v1.3.20240502
