### 2023-11-20 #6067 @britta-wstnr
@nbeliy @adswa @pjtoussaint 👋
FYI @CPernet

This is the review thread for the paper. All of our higher-level communications will happen here from now on, review comments and discussion can happen in the repository of the project (details below).

📓 Please read the "Reviewer instructions & questions" in the [comment from our editorialbot](https://github.com/openjournals/joss-reviews/issues/6067#issue-2001869366).

✅ All reviewers get their own checklist with the JOSS requirements - you generate them as per the details in the editorialbot comment. As you go over the submission, please check any items that you feel have been satisfied.

💻 The JOSS review is different from most other journals: The reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention the link this issue (so that a link is created to this thread). That will also help me to keep track!

❓ Please also feel free to comment and ask questions on this thread.

🎯 We aim for the review process to be completed within about 4-6 weeks but please make a start well ahead of this as JOSS reviews are by their nature iterative and any early feedback you may be able to provide to the author will be very helpful in meeting this schedule.
