### 2023-11-21 openneuropet/PET2BIDS#236 @nbeliy
[JOSS review] Article comments
Dear pet2bids authors, this is in reference to the review of the paper that I'm currently conducting for JOSS [here](https://github.com/openjournals/joss-reviews/issues/6067).

First thanks to put an effort for developing the bidsifier for PET data -- the BIDS specification for PET is quite complex, and your tool is for sure is needed!

I will put my comments in the checklist below:

- [ ] `Statement of needs` do not state the need of software, you may add a small trivial phrase why pet2bids exists (not what it does).
- [ ] JOSS asks to compare the software with similar tools
- [ ] `File conversation`/`PET Metadata`/`Spreadsheet conversation` may be worth to be put in the dedicated (sub)sections
- [ ] Are conversation, metadata and spreadsheets export are extensive list of pet2bids tools?
- [ ] In Metadata/Spreadsheet section you need to change `JSON file` to `sidecar JSON file`, first one is an any generic JSON file, second is BIDS specific one that supplements image data with metadata.
- [ ] In spreadsheets section, I don't really like the usage of points (`(a)` ,  `(i)`, etc.) -- you never reference them. Maybe replace the first `and` at line 54 by `as well as`, and make the iteration in lines 55-58 to a list. I think it would be easier to read.

### Inline comments

- [ ] On line 27, missing references to ECAT and DICOM formats
- [ ] On line 27, by time measurements do you mean the timing of each PET frame? Just to be sure.
- [ ] Line 30 is a little bit confusing, do you mean by `library code, allowing conversion ... using the command line`, that you provide a library **and** a (command line) script?
- [ ] Line 30, I may be wrong here, but `library code` reads as a code that do something with library. Library of tools/functions sounds better for me.
- [ ]  Line 31, `it can be integrated into software` because it's library, not because it modular. Need to be rephrased.
- [ ]  Line 39, needed references for BT Christian and RF Muzic, are those people or tools or functions?
- [ ]  Line 43, What do you mean by `the further testing of data reading`?
- [ ]  Line 46, It's not immediately clear that `JSON files created from reading PET` refers to JSON sidecar files from conversion using pet2bids.
- [ ]  Line 46, `data are always missing...`, `always` seems too strong, maybe better to replace by `often` or `usually`.
- [ ] Line 48, as original PET data (DICOM/ECAT) don't contain JSON, there no `original JSON file`, I guess you meant sidecar JSON file?
- [ ] Line 48, before you provided the names of function (for ex. ecat2nii for conversion), but here you use `PET JSON updater`, can you change it into function name(s)?
- [ ] Line 52, is `xls, xlsx, csv, tsv, bld` it extensive list of tabular data formats in existence, or list of data  that bids2pet supports? If it's first, than an `for example` may be needed.
- [ ] Line 57, what is `blood.tsv` file, you never mentioned it before. Maybe rephrase it to `exports blood data from PMOD tabular file`?
