+++ scripts/python_conversions.sh
@@ -139,7 +139,7 @@ echo "${SOURCE_FOLDER}/PhilipsGeminiPETMR-Unimedizin/reqCTAC"
 dcm2niix4pet $SOURCE_FOLDER/PhilipsGeminiPETMR-Unimedizin/reqCTAC --destination-path $DESTINATION/sub-PhilipsGeminiUnimedizinMainz/pet \
 --kwargs \
 Manufacturer="Philips Medical Systems" \
-ManufacturersModelName="PET/CT Gemini TF16"
+ManufacturersModelName="PET/CT Gemini TF16" \
 InstitutionName="Unimedizin, Mainz, DE" \
 BodyPart="Phantom" \
 Units="Bq/mL" \
@@ -156,14 +156,17 @@ ReconFilterSize=0 \
 AttenuationCorrection="CTAC-SG" \
 ScatterCorrectionMethod="SS-SIMUL" \
 ReconstructionMethod="LOR-RAMLA" \
-ReconMethodParameterValues="[1,1]"
- FrameDuration=[1798] \
+ReconMethodParameterValues="[1,1]" \
+FrameDuration=[1798] \
+ReconMethodParameterLabels="[none, none]" \
+ReconMethodParameterUnits="[none, none]" \
+FrameTimesStart=[0] \
 
 echo "${SOURCE_FOLDER}/PhilipsGeminiPETMR-Unimedizin/reqNAC"
 dcm2niix4pet $SOURCE_FOLDER/PhilipsGeminiPETMR-Unimedizin/reqNAC --destination-path $DESTINATION/sub-PhilipsGeminiNACUnimedizinMainz/pet \
 --kwargs \
 Manufacturer="Philips Medical Systems" \
-ManufacturersModelName="PET/CT Gemini TF16"
+ManufacturersModelName="PET/CT Gemini TF16" \
 InstitutionName="Unimedizin, Mainz, DE" \
 BodyPart="Phantom" \
 Units="Bq/mL" \
@@ -179,8 +182,13 @@ ReconFilterType=None \
 ReconFilterSize=0 \
 AttenuationCorrection="None" \
 ScatterCorrectionMethod="None" \
-ReconstructionMethod="3D-RAMLA"
+ReconstructionMethod="3D-RAMLA" \
 FrameDuration=[1798] \
+ReconMethodParameterLabels="[none, none]" \
+ReconMethodParameterUnits="[none, none]" \
+ReconMethodParameterValues="[1,1]" \
+FrameTimesStart=[0] \
+
 
 # Amsterdam UMC
 # ---------------------------
