### 2023-11-28 05abe4e
Tested previously failing phantoms, pass now. (#237)
* Tested previously failing phantoms, pass now.
Version v1.0.20220720 of dcm2niix fixed some issues with Phillips
dicoms. Uncommented those previously failing phantoms in
python_conversions.sh and ran `make phantoms` they convert and
pass the validator.
