### 2023-11-29 9ff7332
uncomment philips in matlab script too
