### 2023-11-29 b043bab
touching up conversion inputs, added in philips for matlab
