+++ scripts/python_conversions.sh
@@ -72,7 +72,7 @@ FrameTimesStart="[0]" \
 FrameDuration=[300] \
 ImageDecayCorrected="true" \
 ImageDecayCorrectionTime=0 \
-DecayCorrectionFactor=[1] \
+#DecayCorrectionFactor=[1] \
 AttenuationCorrection="MR-corrected" \
 InjectionStart=0
 
@@ -124,10 +124,10 @@ ImageDecayCorrected="true" \
 ImageDecayCorrectionTime=0 \
 AttenuationCorrection="MR-corrected" \
 ReconFilterType='unknown' \
-ReconFilterSize=1 \
-ReconMethodParameterLabels="[none, none]" \
-ReconMethodParameterUnits="[none, none]" \
-ReconMethodParameterValues="[0, 0]"
+ReconFilterSize=0 \
+ReconMethodParameterLabels="[none]" \
+ReconMethodParameterUnits="[none]" \
+ReconMethodParameterValues="[0]"
 
 # Johannes Gutenberg University of Mainz
 # --------------------------------------
@@ -152,14 +152,14 @@ AcquisitionMode="list mode" \
 ImageDecayCorrected=True \
 ImageDecayCorrectionTime=0 \
 ReconFilterType='n/a' \
-ReconFilterSize=0 \
+ReconFilterSize=1 \
 AttenuationCorrection="CTAC-SG" \
 ScatterCorrectionMethod="SS-SIMUL" \
 ReconstructionMethod="LOR-RAMLA" \
-ReconMethodParameterValues="[1,1]" \
+ReconMethodParameterValues="[0]" \
 FrameDuration=[1798] \
-ReconMethodParameterLabels="[none, none]" \
-ReconMethodParameterUnits="[none, none]" \
+ReconMethodParameterLabels="[none]" \
+ReconMethodParameterUnits="[none]" \
 FrameTimesStart=[0] \
 
 echo "${SOURCE_FOLDER}/PhilipsGeminiPETMR-Unimedizin/reqNAC"
@@ -179,14 +179,14 @@ AcquisitionMode="list mode" \
 ImageDecayCorrected=True \
 ImageDecayCorrectionTime=0 \
 ReconFilterType=None \
-ReconFilterSize=0 \
+ReconFilterSize=1 \
 AttenuationCorrection="None" \
 ScatterCorrectionMethod="None" \
 ReconstructionMethod="3D-RAMLA" \
 FrameDuration=[1798] \
-ReconMethodParameterLabels="[none, none]" \
-ReconMethodParameterUnits="[none, none]" \
-ReconMethodParameterValues="[1,1]" \
+ReconMethodParameterLabels="[none]" \
+ReconMethodParameterUnits="[none]" \
+ReconMethodParameterValues="[0]" \
 FrameTimesStart=[0] \
 
 
@@ -212,16 +212,16 @@ AcquisitionMode="list mode" \
 ImageDecayCorrected="true" \
 ImageDecayCorrectionTime=0 \
 DecayCorrectionFactor=[1] \
-ReconFilterSize=0 \
+ReconFilterSize=1 \
 ReconMethodParameterValues=[1] \
 AttenuationCorrection="CTAC-SG" \
 RandomsCorrectionMethod="DLYD" \
 ScatterCorrectionMethod="SS-SIMUL" \
 ReconstructionMethod="BLOB-OS-TF" \
-ReconMethodParameterLabels="[none, none]" \
-ReconMethodParameterUnits="[none, none]" \
-ReconMethodParameterValues="[0, 0]" \
-ReconFilterType="none" \
+ReconMethodParameterLabels="[none]" \
+ReconMethodParameterUnits="[none]" \
[…]
