### 2023-11-30 c0b464a
update conversion scripts
