### 2023-11-30 e66165a
bumped up version of pytest for ezBIDS
