### 2023-11-30 dc5de03
bumped up version of pytest for ezBIDS (#240)
* bumped up version of pytest for ezBIDS
* poetry lock removed "toml" manually added back in
