### 2023-11-30 b6a77c3
Bug fix: enable json sidecar generation for older dcm2niix versions
