+++ scripts/matlab_conversions.m
@@ -2,11 +2,11 @@
 % note that usually we use the *parameter.txt file facilitating the metadata
 % information but here it is done fully for completeness
 % 
-% cyril pernet - may 2022
+% cyril pernet - Nov 2023
 
-%% set paths where the repo is
-%source      = 'D:\BIDS\ONP\OpenNeuroPET-Phantoms\sourcedata\';
-%destination = 'D:\BIDS\ONP\OpenNeuroPET-Phantoms\';
+%% set paths where the repo is to bypass the path/env parts below
+% source      = 'D:\BIDS\ONP\OpenNeuroPET-Phantoms\sourcedata\';
+% destination = 'D:\BIDS\ONP\OpenNeuroPET-Phantoms\';
 
 %% get the path to this script
 matlab_conversions_script_path = mfilename('fullpath');
@@ -20,7 +20,9 @@ phantoms_folder_path = strjoin(parts(1:end-2),filesep);
 source      = strjoin({phantoms_folder_path, 'sourcedata'}, filesep);
 destination = strjoin({phantoms_folder_path, 'matlab'}, filesep);
 
-mkdir(destination)
+if ~exist(destination,'dir')
+    mkdir(destination)
+end
 copyfile(strjoin({phantoms_folder_path, 'dataset_description.json'}, filesep), strjoin({destination, 'dataset_description.json'}, filesep))
 
 %% sets the environment variable so that matlab can reach dcm2niix on posix
@@ -211,10 +213,9 @@ try
     meta.ReconFilterType            = "none";
     meta.ReconFilterSize            = 1;
 
-    %dcm2niix4pet(fullfile(source,['PhilipsGeminiPETMR-Unimedizin' filesep 'reqCTAC']),meta,...
-    %    'o',fullfile(destination,['sub-PhilipsGeminiUnimedizinMainz' filesep 'pet']));
-    
-    
+    dcm2niix4pet(fullfile(source,['PhilipsGeminiPETMR-Unimedizin' filesep 'reqCTAC']),meta,...
+       'o',fullfile(destination,['sub-PhilipsGeminiUnimedizinMainz' filesep 'pet']));
+        
     % meta.AttenuationCorrection      = 'NONE';
     % meta.ScatterCorrectionMethod    = 'NONE';%GEG
     % meta.ReconstructionMethod       = '3D-RAMLA';%GEG
@@ -436,13 +437,14 @@ catch
 end
 
 try
-meta.AttenuationCorrection = 'measured' ; % some how the field is not there
-meta.ReconMethodParameterLabels = ["none", "none"];
-meta.ReconMethodParameterUnits = ["none", "none"];
-meta.ReconMethodParameterValues = [0, 0];
-meta.FrameDuration = 98000
-dcm2niix4pet(fullfile(source,['GeneralElectricAdvance-NIMH' filesep ...
-    'long_trans']),meta,'o',fullfile(destination,['sub-GeneralElectricAdvanceLongNIMH' filesep 'pet'])); 
+    meta.AttenuationCorrection = 'measured' ; % some how the field is not there
+    meta.ReconMethodParameterLabels = ["none", "none"];
+    meta.ReconMethodParameterUnits = ["none", "none"];
+    meta.ReconMethodParameterValues = [0, 0];
+    meta.FrameDuration = 98000;
+
+    dcm2niix4pet(fullfile(source,['GeneralElectricAdvance-NIMH' filesep ...
+        'long_trans']),meta,'o',fullfile(destination,['sub-GeneralElectricAdvanceLongNIMH' filesep 'pet']));
 catch
     disp(message);
 end
@@ -453,3 +455,109 @@ end
[…]
