### 2023-12-08 52aaae1
update python conversion script for newer phantoms
