+++ scripts/python_conversions.sh
@@ -48,7 +48,7 @@ SpecificRadioactivityUnits="Bq" \
 ScanStart=0 \
 InjectionStart=0 \
 InjectedRadioactivityUnits='Bq' \
-ReconFilterType="none"
+ReconFilterType="none" \
 #TimeZero="10:10:10" \
 #InjectedMass=1 \
 #InjectedMassUnits=g \
@@ -74,7 +74,8 @@ ImageDecayCorrected="true" \
 ImageDecayCorrectionTime=0 \
 DecayCorrectionFactor=[1] \
 AttenuationCorrection="MR-corrected" \
-InjectionStart=0
+InjectionStart=0 \
+--silent
 
 
 # Århus University Hospital
@@ -101,7 +102,8 @@ ReconFilterType='none' \
 FrameTimesStart=[0] \
 ReconMethodParameterLabels="[none]" \
 ReconMethodParameterUnits="[none]" \
-ReconMethodParameterValues="[0]"
+ReconMethodParameterValues="[0]" \
+--silent
 #DecayCorrectionFactor="[1]" \
 
 echo "${SOURCE_FOLDER}/GeneralElectricSignaPETMR-Aarhus"
@@ -127,7 +129,8 @@ ReconFilterType='unknown' \
 ReconFilterSize=1 \
 ReconMethodParameterLabels="[none, none]" \
 ReconMethodParameterUnits="[none, none]" \
-ReconMethodParameterValues="[0, 0]"
+ReconMethodParameterValues="[0, 0]" \
+--silent
 
 # Johannes Gutenberg University of Mainz
 # --------------------------------------
@@ -161,6 +164,7 @@ FrameDuration=[1798] \
 ReconMethodParameterLabels="[none, none]" \
 ReconMethodParameterUnits="[none, none]" \
 FrameTimesStart=[0] \
+--silent
 
 echo "${SOURCE_FOLDER}/PhilipsGeminiPETMR-Unimedizin/reqNAC"
 dcm2niix4pet $SOURCE_FOLDER/PhilipsGeminiPETMR-Unimedizin/reqNAC --destination-path $DESTINATION/sub-PhilipsGeminiNACUnimedizinMainz/pet \
@@ -188,6 +192,7 @@ ReconMethodParameterLabels="[none, none]" \
 ReconMethodParameterUnits="[none, none]" \
 ReconMethodParameterValues="[1,1]" \
 FrameTimesStart=[0] \
+--silent
 
 
 # Amsterdam UMC
@@ -222,6 +227,7 @@ ReconMethodParameterLabels="[none, none]" \
 ReconMethodParameterUnits="[none, none]" \
 ReconMethodParameterValues="[0, 0]" \
 ReconFilterType="none" \
+--silent
 #FrameTimesStart=[0]
 #TimeZero="12:12:12" \
 
@@ -255,7 +261,8 @@ ReconFilterType="['n/a', 'n/a']" \
 DecayCorrectionFactor=[1] \
 ReconMethodParameterLabels="[none, none]" \
 ReconMethodParameterUnits="[none, none]" \
-ReconMethodParameterValues="[0, 0]"
+ReconMethodParameterValues="[0, 0]" \
+--silent
 
 # philipsVereosPET-CT
 # -------------------
@@ -281,7 +288,8 @@ AttenuationCorrection="CTAC-SG" \
 ScatterCorrectionMethod="SS-SIMUL" \
 RandomsCorrectionMethod="DLYD" \
 ReconstructionMethod="OSEMi3s15" \
-TimeZero="11:40:24"
+TimeZero="11:40:24" \
+--silent
 #FrameDuration=[1221]
 
 
[…]
