### 2023-12-26 b498671
updated readme to emphasize PyPi install, clarified other options too
