### 2023-12-26 6da0ae7
update docstrings and rtd pages
