+++ pypet2bids/pypet2bids/sidecar.py
@@ -1,6 +1,6 @@
 """
 This is a lazy way to avoid opening a json, simply import this file to collect your BIDS sidecar templates instead. This
-is not a function or a `true` module. It's just two python dictionaries with keys and empty value pairs.
+is not a function or a true module. This is just two python dictionaries with keys and empty value pairs.
 
 :param sidecar_template_full: a dictionary containing every field specified in the BIDS standard for PET imaging data
 :param sidecar_template_short: a dictionary containing only the required fields in the BIDS standard for PET
@@ -9,6 +9,7 @@ is not a function or a `true` module. It's just two python dictionaries with key
 
 |*Anthony Galassi*
 |*Copyright OpenNeuroPET team*
+
 """
 
 
