+++ pypet2bids/pyproject.toml
@@ -18,7 +18,6 @@ python-dateutil = "^2.8.2"
 python-dotenv = "^0.19.1"
 scipy = "^1.7.1"
 six = "^1.16.0"
-pytest = ">=6.2.5, <8.0.0"
 pydicom = "^2.2.2"
 openpyxl = "^3.0.9"
 xlrd = "^2.0.1"
@@ -31,6 +30,7 @@ toml = ">=0.10.2"
 
 
 [tool.poetry.dev-dependencies]
+pytest = ">=6.2.5, <8.0.0"
 sphinx = "<=4.5.0"
 sphinx-rtd-theme = "^1.0.0"
 sphinxcontrib-matlabdomain = "^0.13.0"
