### 2023-12-29 c7abd22
fixed minor bugs with sphinx
