### 2024-01-08 5a9919a
file conversion and spreadsheet txt clarification
