+++ paper.bib
@@ -160,6 +160,19 @@ year = {1997}
 	date = {2023},
 }
 
+@online{dicm_ref_2020,
+	title = {DICOM PS3.3 2020b - Information Object Definitions},
+	url = {https://dicom.nema.org/medical/dicom/2020b/output/chtml/part03/PS3.3.html},
+	urldate = {2024-01-08},
+	date = {2020},
+}
+
+@software{@Li_2016,
+	title = {DICOM to NIfTI conversion, DICOM and NIfTI tools, NIfTI visualization},
+	url = {https://github.com/xiangruili/dicm2nii},
+	urldate = {2024-01-08},
+	date = {2016},
+}
 @software{noauthor_ezbids,
 	title = {{ezBIDS}},
 	url = {https://brainlife.io/ezbids/},
