+++ paper.bib
@@ -167,14 +167,14 @@ year = {1997}
 	date = {2020},
 }
 
-@software{@Li_2016,
+@software{Li_2016,
 	title = {DICOM to NIfTI conversion, DICOM and NIfTI tools, NIfTI visualization},
 	url = {https://github.com/xiangruili/dicm2nii},
 	urldate = {2024-01-08},
 	date = {2016},
 }
 @software{noauthor_ezbids,
-	title = {{ezBIDS}},
+	title = {ezBIDS},
 	url = {https://brainlife.io/ezbids/},
 	urldate = {2023-08-22},
 	date = {2023},
