+++ paper.bib
@@ -168,8 +168,9 @@ year = {1997}
 }
 
 @software{Li_2016,
-	title = {DICOM to NIfTI conversion, DICOM and NIfTI tools, NIfTI visualization},
+	title = {dicm2nii},
 	url = {https://github.com/xiangruili/dicm2nii},
+	author = {Li, X.},
 	urldate = {2024-01-08},
 	date = {2016},
 }
