### 2024-01-09 74c6b7b
Updating Python Installation Instructions (#253)
* updated readme to emphasize PyPi install, clarified other options too

* update docstrings and rtd pages

* fixed minor bugs with sphinx
