### 2024-01-09 1926cfc
update to poetry based install of requirements
