+++ pypet2bids/pyproject.toml
@@ -10,7 +10,7 @@ include = [
     ]
 
 [tool.poetry.dependencies]
-python = ">=3.8,<=3.11"
+python = ">=3.8,<=3.11.6"
 nibabel = ">=3.2.1"
 numpy = "^1.21.3"
 pyparsing = "^3.0.4"
