+++ .readthedocs.yaml
@@ -20,20 +20,14 @@ build:
   jobs:
     post_create_environment:
       # install poetry
-      # https://python-poetry.org/docs/#osx--linux--bashonwindows-install-instructions
-      - curl -sSL https://install.python-poetry.org | python3 -
-      # Tell poetry to not use a virtual environment
-      - $HOME/.local/bin/poetry config virtualenvs.create false
+      - pip install poetry
+      - poetry config virtualenvs.create false
       # just use poetry to export a requirements.txt as that worked much better than the previous attempts
-      - cd pypet2bids && $HOME/.local/bin/poetry export --with dev --without-hashes -o requirements.txt
+      - cd pypet2bids && poetry export --without-hashes --with dev --format=requirements.txt > requirements.txt
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
    configuration: docs/conf.py
-
-# If using Sphinx, optionally build your docs in additional formats such as PDF
-# formats:
-#    - pdf
 
 # Optionally declare the Python requirements required to build your docs
 python:
