### 2024-01-09 dda33a8
remove unused dep, more fighting with rtd
