+++ .readthedocs.yaml
@@ -16,14 +16,14 @@ submodules:
 build:
   os: ubuntu-20.04
   tools:
-    python: "3.9"
+    python: "3.11"
   jobs:
     post_create_environment:
       # install poetry
       - pip install poetry
       - poetry config virtualenvs.create false
       # just use poetry to export a requirements.txt as that worked much better than the previous attempts
-      - cd pypet2bids && poetry export --without-hashes --with dev --format=requirements.txt > requirements.txt
+      - cd pypet2bids && poetry lock && poetry export --without-hashes --with dev --format=requirements.txt > requirements.txt
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
