### 2024-01-10 dd91243
Update README.md (#254)
Remove duplication of links to readthedocs site. Fix small spelling and or typos
