### 2024-01-10 e75f7b3
added entrypoint for updating a json from a pet dicom file
