### 2024-01-15 openneuropet/PET2BIDS#236 @CPernet
ready to go once #255 is fixed
