### 2024-01-17 1eefe1f
update scripts to work on us hosted phantoms
