+++ scripts/matlab_conversions.m
@@ -49,285 +49,285 @@ end
 
 message = 'Failed to convert subject, moving onto next.';
 
-%% Neurobiology Research Unit - Copenhagen
-% ----------------------------------------
-
-% Siemens HRRT
-% ------------
-try
-    clear meta
-    meta.TimeZero                   = 'ScanStart';
-    meta.Manufacturer               = 'Siemens';
-    meta.ManufacturersModelName     = 'HRRT';
-    meta.InstitutionName            = 'Rigshospitalet, NRU, DK';
-    meta.BodyPart                   = 'Phantom';
-    meta.Units                      = 'Bq/mL';
-    meta.TracerName                 = 'FDG';
-    meta.TracerRadionuclide         = 'F18';
-    meta.InjectedRadioactivity      = 81.24; % Mbq
-    meta.SpecificRadioactivity      = 1.3019e+04; % ~ 81240000 Bq/ 6240 g
-    meta.ModeOfAdministration       = 'infusion';   
-    meta.AcquisitionMode            = 'list mode';
-    meta.ImageDecayCorrected        = true; % when passing this as string it fails validation
-    meta.ImageDecayCorrectionTime   = 0;
-    meta.ReconFilterType            = 'none';
-    meta.ReconFilterSize            = 0;
-    meta.AttenuationCorrection      = '10-min transmission scan';
-    % meta.FrameDuration            = 1200; % not needed, encoded in ecat even for 4D images
-    % meta.FrameTimesStart          = 0;
-    
-    out = ecat2nii(fullfile(source,['SiemensHRRT-NRU' filesep 'XCal-Hrrt-2022.04.21.15.43.05_EM_3D.v']),...
-        meta,'gz',true,'FileListOut',fullfile(destination,['sub-SiemensHRRTNRU'  filesep 'pet' filesep 'sub-SiemensHRRTNRU.nii']));
-
-catch
-    disp(message);
-end
-
-% Siemens Biograph
-% ---------------------------
-try
-    clear meta
-    meta.TimeZero                   = 'ScanStart';
-    meta.Manufacturer               = 'Siemens';
-    meta.ManufacturersModelName     = 'Biograph';
-    meta.InstitutionName            = 'Rigshospitalet, NRU, DK';
-    meta.BodyPart                   = 'Phantom';
-    meta.Units                      = 'Bq/mL';
-    meta.TracerName                 = 'FDG';
-    meta.TracerRadionuclide         = 'F18';
-    meta.InjectedRadioactivity      = 81.24; % Mbq
-    meta.SpecificRadioactivity      = 1.3019e+04; % ~ 81240000 Bq/ 6240 g
-    meta.ModeOfAdministration       = 'infusion';
-    meta.FrameTimesStart            = 0;
-    meta.AcquisitionMode            = 'list mode';
-    meta.ImageDecayCorrected        = 'true';
-    meta.ImageDecayCorrectionTime   = 0;
-    meta.AttenuationCorrection      = 'MR-corrected';
-    meta.FrameDuration              = 300;
-    meta.FrameTimesStart            = 0;
-    meta.ReconFilterType            = "none";
-    meta.ReconFilterSize            = 1;
-
-    dcm2niix4pet(fullfile(source,'SiemensBiographPETMR-NRU'),meta,...
-        'o',fullfile(destination,['sub-SiemensBiographNRU' filesep 'pet'])); % note we only need to use folders here
-catch
-    disp(message);
-end
-
-%% Århus University Hospital
-% ---------------------------
-try
-    clear meta
-    meta.TimeZero                   = 'ScanStart';
[…]
