+++ .github/workflows/python.yaml
@@ -53,7 +53,7 @@ jobs:
         if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
         run: |
           cd pypet2bids
-          python -m pip install -upgrade pip
+          python -m pip install --upgrade pip
           pip install poetry
           poetry install --with dev
 
