### 2024-01-17 d4cc324
Switch phantoms to s3, update CI, fix bug with updatepetjson.m on GE dicoms (#259)
* switch scripts to work on us s3 hosted phantoms
* update poetry, switch to poetry run
* fix updatejsonpetfile.m to handle missing Recon Data
