+++ scripts/python_conversions.sh
@@ -22,268 +22,258 @@ echo "SOURCE_FOLDER ${SOURCE_FOLDER}, DESTINATION ${DESTINATION}"
 mkdir $DESTINATION
 cp $repo_path/dataset_description.json $DESTINATION/dataset_description.json
 
-# Neurobiology Research Unit - Copenhagen
-# ----------------------------------------
-
-# Siemens HRRT
-# ------------
-echo "${SOURCE_FOLDER}/SiemensHRRT-NRU/XCal-Hrrt-2022.04.21.15.43.05_EM_3D.v"
-ecatpet2bids $SOURCE_FOLDER/SiemensHRRT-NRU/XCal-Hrrt-2022.04.21.15.43.05_EM_3D.v --nifti $DESTINATION/sub-SiemensHRRTNRU/pet/sub-SiemensHRRTNRU_pet.nii --convert --kwargs \
-Manufacturer=Siemens \
-ManufacturersModelName=HRRT \
-InstitutionName="Rigshospitalet, NRU, DK" \
-BodyPart="Phantom" \
-Units="Bq/mL" \
-TracerName=FDG \
-TracerRadionuclide=F18 \
-InjectedRadioactivity=81.24 \
-SpecificRadioactivity="1.3019e+04" \
-ModeOfAdministration=infusion \
-AcquisitionMode="list mode" \
-ImageDecayCorrected="true" \
-ImageDecayCorrectionTime=0 \
-ReconFilterSize=0 \
-AttenuationCorrection="10-min transmission scan" \
-SpecificRadioactivityUnits="Bq" \
-ScanStart=0 \
-InjectionStart=0 \
-InjectedRadioactivityUnits='Bq' \
-ReconFilterType="none"
-#TimeZero="10:10:10" \
-#InjectedMass=1 \
-#InjectedMassUnits=g \
-
-# Siemens Biograph
-# ---------------------------
-echo "${SOURCE_FOLDER}/SiemensBiographPETMR-NRU"
-dcm2niix4pet $SOURCE_FOLDER/SiemensBiographPETMR-NRU --destination-path $DESTINATION/sub-SiemensBiographNRU/pet --kwargs \
-Manufacturer=Siemens \
-ManufacturersModelName=Biograph \
-InstitutionName="Rigshospitalet, NRU, DK" \
-BodyPart=Phantom \
-Units="Bq/mL" \
-TracerName="FDG" \
-TracerRadionuclide="F18" \
-InjectedRadioactivity=81.24 \
-SpecificRadioactivity=1.3019e+04 \
-ModeOfAdministration="infusion" \
-AcquisitionMode="list mode" \
-FrameTimesStart="[0]" \
-FrameDuration=[300] \
-ImageDecayCorrected="true" \
-ImageDecayCorrectionTime=0 \
-DecayCorrectionFactor=[1] \
-AttenuationCorrection="MR-corrected" \
-InjectionStart=0
-
-
-# Århus University Hospital
-# ---------------------------
-echo "${SOURCE_FOLDER}/GeneralElectricDiscoveryPETCT-Aarhus"
-dcm2niix4pet $SOURCE_FOLDER/GeneralElectricDiscoveryPETCT-Aarhus  --destination-path $DESTINATION/sub-GeneralElectricDiscoveryAarhus/pet --kwargs \
-Manufacturer="General Electric" \
-ManufacturersModelName="Discovery" \
-InstitutionName="Århus University Hospital, DK" \
-BodyPart="Phantom" \
-Units="Bq/mL" \
-TracerName="FDG" \
-TracerRadionuclide="F18" \
-InjectedRadioactivity=25.5 \
[…]
