+++ .github/workflows/python.yaml
@@ -4,6 +4,7 @@ on:
   pull_request:
     branches:
       - main
+      - switch-phantoms-to-s3
   workflow_call:
   workflow_dispatch:
     inputs:
@@ -18,8 +19,7 @@ jobs:
     runs-on: ${{ matrix.os }}
 
     env:
-      ECAT_TEST_FOLDER: "cimbi36"
-      REAL_TEST_ECAT_PATH: cimbi36/Gris_102_19_2skan-2019.04.30.13.04.41_em_3d.v
+      REAL_ECAT_TEST_PATH: ${{ github.workspace }}/OpenNeuroPET-Phantoms/sourcedata/SiemensHRRT-JHU/Hoffman.v
       SMALLER_ECAT_PATH: ${{ github.workspace }}/ecat_validation/ECAT7_multiframe.v.gz
       TEST_ECAT_PATH: ${{ github.workspace }}/ecat_validation/ECAT7_multiframe.v
       OUTPUT_NIFTI_PATH: ${{ github.workspace}}/pypet2bids/tests/ECAT7_multiframe.nii
@@ -36,15 +36,15 @@ jobs:
       - name: Checkout Repo
         uses: actions/checkout@v2
 
-      - name: Cache ECAT
-        id: cache-ecat
+      - name: Cache Phantoms
+        id: cache-phantoms
         uses: actions/cache@v2
         with:
-          path: cimbi36
-          key: ${{ runner.os }}-ecats
+          path: US-sourced-OpenNeuroPET-Phantoms.zip
+          key: ${{ runner.os }}-phantoms
 
       - name: Set up python
-        uses: actions/setup-python@v2
+        uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python }}
 
@@ -52,25 +52,33 @@ jobs:
         if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
         run: |
           cd pypet2bids
-          pip3 install .
+          python -m pip install --upgrade pip
+          pip install poetry
+          poetry install --with dev
 
-      - name: Install gdown and collect Ecat from Google Drive
-        if: steps.cache-ecat.outputs.cache-hit != 'true'
-        run: "python3 -m pip install gdown && gdown ${{ secrets.CIMBI_ECAT_ON_GOOGLE_DRIVE }} -O ecat_test"
+      - name: Collect ECAT and other phantoms
+        if: steps.cache-phantoms.outputs.cache-hit != 'true'
+        run: "wget https://openneuropet.s3.amazonaws.com/US-sourced-OpenNeuroPET-Phantoms.zip"
 
-      - name: Decompress dataset
-        if: steps.cache-ecat.outputs.cache-hit != 'true'
-        run: "tar xvzf ecat_test && rm ecat_test"
+      - name: Decompress ECAT and other phantoms
+        run: "unzip US-sourced-OpenNeuroPET-Phantoms.zip"
+
+      - name: Debug
+        uses: mxschmitt/action-tmate@v3
+        if: ${{ github.event_name == 'workflow_dispatch' && inputs.debug_enabled }}
+        timeout-minutes: 15
+        with:
+          limit-access-to-actor: true
 
       - name: Test CLI --help
         run: |
           cd pypet2bids/
-          python3 -m pypet2bids.ecat_cli --help
+          poetry run python3 -m pypet2bids.ecat_cli --help
 
       - name: Test CLI Ecat Dump
         run: |
           cd pypet2bids/
-          python3 -m pypet2bids.ecat_cli ../${{ env.REAL_TEST_ECAT_PATH }} --dump
+          poetry run python3 -m pypet2bids.ecat_cli ${{ env.REAL_ECAT_TEST_PATH }} --dump
 
[…]
