### 2024-01-17 292ec00
try globing with bash
