+++ .readthedocs.yaml
@@ -25,7 +25,7 @@ build:
       # just use poetry to export a requirements.txt as that worked much better than the previous attempts
       - cd pypet2bids && poetry lock && poetry export --without-hashes --with dev --format=requirements.txt > requirements.txt
       - pip install -r pypet2bids/requirements.txt
-      - cd pypet2bids && pip install sdist/pypet2bid-*.whl
+      - cd pypet2bids && bash pip install sdist/pypet2bid-*.whl
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
