### 2024-01-18 bcac72d
remove leading pypet2bids?
