### 2024-01-18 94fcb04
copy over metadata to pypet2bids/pypet2bids folder
