+++ .readthedocs.yaml
@@ -19,7 +19,7 @@ build:
     python: "3.11"
   jobs:
     post_create_environment:
-      - ls .
+      - cp -r metadata pypet2bids/pypet2bids/
       # install poetry
       - pip install poetry
       - poetry config virtualenvs.create false
