### 2024-01-18 d4da7f3
fix-rtd-formatting (#257)
fixes failed rendering on rtd for matlab and python

Updates to the build system on readthedocs.org, changes to sphinx-contrib
for matlab, and now explicitly copying over some metadata (on this author)
so that it's available on the read the docs vm during build.

* refactoring and small formatting changes to docs
* update rtd yaml and conf.py
* copy over metadata to pypet2bids/pypet2bids folder
* update sphinx-contrib version to 0.21.4
