+++ pypet2bids/pypet2bids/sidecar.py
@@ -2,10 +2,12 @@
 This is a lazy way to avoid opening a json, simply import this file to collect your BIDS sidecar templates instead. This
 is not a function or a true module. This is just two python dictionaries with keys and empty value pairs.
 
-:param sidecar_template_full: a dictionary containing every field specified in the BIDS standard for PET imaging data
-:param sidecar_template_short: a dictionary containing only the required fields in the BIDS standard for PET
-    imaging data
-:return: sidecar_template_full, sidecar_template_short
+**Parameters**
+
+* sidecar_template_full: a dict containing every PET BIDS field
+* sidecar_template_short: a dict containing only required PET BIDS fields
+
+**Returns**     sidecar_template_full, sidecar_template_short
 
 *Anthony Galassi*
 *Copyright OpenNeuroPET team*
