### 2024-01-19 b19dbe3
added entity flags for trc, rec, run and full filename support
