+++ pypet2bids/pypet2bids/dcm2niix4pet.py
@@ -67,6 +67,11 @@ if pypet2bids_config.exists():
     if default_metadata_json and Path(default_metadata_json).exists():
         # do nothing
         pass
+    else:
+        try:
+            shutil.copy(Path(metadata_folder) / 'template_json.json', default_metadata_json)
+        except FileNotFoundError:
+            shutil.copy(module_folder / 'template_json.json', default_metadata_json)
 else:
     # if it doesn't exist use the default one included in this library
     helper_functions.modify_config_file('DEFAULT_METADATA_JSON', module_folder / 'template_json.json')
@@ -421,25 +426,29 @@ class Dcm2niix4PET:
         # check if user provided a custom tempdir location
         self.tempdir_location = tempdir_location
         self.image_folder = Path(image_folder)
-        if destination_path:
-            self.destination_path = Path(destination_path)
-        else:
-            self.destination_path = self.image_folder
+        self.destination_folder = None
 
         # if we're provided an entire file path just us that no matter what, we're assuming the user knows what they
         # are doing in that case
         self.full_file_path_given = False
-        for part in self.destination_path.parts:
+
+        for part in Path(destination_path).parts:
             if '.nii' in part or '.nii.gz' in part:
                 self.full_file_path_given = True
+                self.destination_folder = Path(destination_path).parent
                 # replace .nii and .nii.gz
-                self.destination_path = Path(str(self.destination_path).replace('.nii', '').replace('.gz', ''))
+                self.destination_path = Path(str(destination_path).replace('.nii', '').replace('.gz', ''))
                 break
 
         # replace the suffix in the destination path with '' if a non-nifti full file path is give
-        if Path(self.destination_path).suffix:
+        if Path(destination_path).suffix:
             self.full_file_path_given = True
-            self.destination_path = Path(self.destination_path).with_suffix('')
+            self.destination_folder = destination_path.parent
+            self.destination_path = Path(destination_path).with_suffix('')
+
+        if not self.full_file_path_given:
+            self.destination_folder = Path(destination_path)
+            self.destination_path = self.destination_folder
 
         # extract PET filename parts from destination path if given
         self.subject_id = helper_functions.collect_bids_part('sub', str(self.destination_path))
@@ -449,13 +458,11 @@ class Dcm2niix4PET:
         self.reconstruction_method = helper_functions.collect_bids_part('rec', str(self.destination_path))
         self.run_id = helper_functions.collect_bids_part('run', str(self.destination_path))
 
-        # once we've groked all the parts (entities) or ultimate path of the output files (blood, nifti, json, etc)
-        # we will save that to this variable.
-        self.new_file_name_with_entities = None
+        self.file_name_slug = None
 
         # we keep track of PET metadata in this spreadsheet metadata_dict, that includes nifti, _blood.json, and
         # _blood.tsv data
-        self.spreadsheet_metadata = {}
+        self.spreadsheet_metadata = {'nifti_json': {}, 'blood_json': {}, 'blood_tsv': {}}
         self.dicom_headers = self.extract_dicom_headers()
         # we consider values stored in a default JSON file to be additional arguments, we load those
         # values first and then overwrite them with any user supplied values
@@ -824,6 +831,7 @@ class Dcm2niix4PET:
 
                     if self.full_file_path_given:
                         new_path = self.destination_path.with_suffix(suffix)
+                        self.destination_folder = self.destination_path.parent
                     else:
                         new_path = self.destination_path / Path(self.subject_id + session_id + task + trc + rec +
                                                                 run + '_pet' + suffix)
@@ -861,13 +869,13 @@ class Dcm2niix4PET:
                 if type(blood_tsv_data) is dict:
[…]
