### 2024-01-22 6dd23d4
added tests run on github to Makefile
