### 2024-01-23 d750f95
update documentation on how to run tests locally
