### 2024-01-23 2499d72
Add Instructions for Users to Run Tests & Add Easily Retrievable Test Data (#261)
* added tests run on github to Makefile
* update documentation on how to run tests locally per issue #247
