### 2024-01-23 b8969c9
added entity flags for trc, rec, run and full filename support (#260)
