### 2024-01-23 6f8b812
begining refactoring to externalize json commands
