+++ pypet2bids/pypet2bids/update_json_pet_file.py
@@ -1,0 +1,543 @@
+from pathlib import Path
+from os.path import join
+from os import listdir
+import json
+from json_maj.main import JsonMAJ, load_json_or_dict
+import re
+from dateutil import parser
+import argparse
+import pydicom
+from typing import Union
+
+try:
+    import helper_functions
+    import is_pet
+except ModuleNotFoundError:
+    import pypet2bids.helper_functions as helper_functions
+    import pypet2bids.is_pet as is_pet
+
+# import logging
+logger = helper_functions.logger("pypet2bids")
+
+# load module and metadata_json paths from helper_functions
+module_folder, metadata_folder = helper_functions.module_folder, helper_functions.metadata_folder
+
+try:
+    # collect metadata jsons in dev mode
+    metadata_jsons = \
+        [Path(join(metadata_folder, metadata_json)) for metadata_json
+         in listdir(metadata_folder) if '.json' in metadata_json]
+except FileNotFoundError:
+    metadata_jsons = \
+        [Path(join(module_folder, 'metadata', metadata_json)) for metadata_json
+         in listdir(join(module_folder, 'metadata')) if '.json' in metadata_json]
+
+# create a dictionary to house the PET metadata files
+metadata_dictionaries = {}
+
+for metadata_json in metadata_jsons:
+    try:
+        with open(metadata_json, 'r') as infile:
+            dictionary = json.load(infile)
+
+        metadata_dictionaries[metadata_json.name] = dictionary
+    except FileNotFoundError as err:
+        raise Exception(f"Missing pet metadata file {metadata_json} in {metadata_folder}, unable to validate metadata.")
+    except json.decoder.JSONDecodeError as err:
+        raise IOError(f"Unable to read from {metadata_json}")
+
+
+def check_json(path_to_json, items_to_check=None, silent=False, spreadsheet_metadata={}, **additional_arguments):
+    """
+    This method opens a json and checks to see if a set of mandatory values is present within that json, optionally it
+    also checks for recommended key value pairs. If fields are not present a warning is raised to the user.
+
+    :param spreadsheet_metadata:
+    :type spreadsheet_metadata:
+    :param path_to_json: path to a json file e.g. a BIDS sidecar file created after running dcm2niix
+    :param items_to_check: a dictionary with items to check for within that json. If None is supplied defaults to the
+           PET_metadata.json contained in this repository
+    :param silent: Raises warnings or errors to stdout if this flag is set to True
+    :return: dictionary of items existence and value state, if key is True/False there exists/(does not exist) a
+            corresponding entry in the json the same can be said of value
+    """
+
+    if silent:
+        logger.disable = True
+    else:
+        logger.disabled = False
+
+    # check if path exists
+    path_to_json = Path(path_to_json)
+    if not path_to_json.exists():
+        raise FileNotFoundError(path_to_json)
+
+    # check for default argument for dictionary of items to check
+    if items_to_check is None:
+        items_to_check = metadata_dictionaries['PET_metadata.json']
+        # remove blood tsv data from items to check
+        if items_to_check.get('blood_recording_fields', None):
+            items_to_check.pop('blood_recording_fields')
+
+    # open the json
+    with open(path_to_json, 'r') as in_file:
+        json_to_check = json.load(in_file)
+
+    # initialize warning colors and warning storage dictionary
+    storage = {}
+    flattened_spreadsheet_metadata = {}
+    flattened_spreadsheet_metadata.update(spreadsheet_metadata.get('nifti_json', {}))
+    flattened_spreadsheet_metadata.update(spreadsheet_metadata.get('blood_json', {}))
+    flattened_spreadsheet_metadata.update(spreadsheet_metadata.get('blood_tsv', {}))
+
[…]
