+++ pypet2bids/pypet2bids/dcm2niix4pet.py
@@ -13,9 +13,8 @@ For more details see the CLI portion of this module or the documentation for the
 """
 import pathlib
 import sys
-import os
 import textwrap
-from json_maj.main import JsonMAJ, load_json_or_dict
+from json_maj.main import JsonMAJ
 from platform import system
 import subprocess
 import pandas as pd
@@ -27,36 +26,29 @@ import pydicom
 import re
 from tempfile import TemporaryDirectory
 import shutil
-from dateutil import parser
 import argparse
 import importlib
-import dotenv
-import logging
 
 try:
     import helper_functions
     import is_pet
+    from update_json_pet_file import (check_json, update_json_with_dicom_value, update_json_with_dicom_value_cli,
+                                      get_radionuclide, check_meta_radio_inputs, metadata_dictionaries,
+                                      get_metadata_from_spreadsheet)
 except ModuleNotFoundError:
     import pypet2bids.helper_functions as helper_functions
     import pypet2bids.is_pet as is_pet
+    from pypet2bids.update_json_pet_file import (check_json, update_json_with_dicom_value,
+                                                 update_json_with_dicom_value_cli, get_radionuclide,
+                                                 check_meta_radio_inputs, metadata_dictionaries,
+                                                 get_metadata_from_spreadsheet)
 
 logger = helper_functions.logger("pypet2bids")
 
-# fields to check for
 module_folder = Path(__file__).parent.resolve()
 python_folder = module_folder.parent
 pet2bids_folder = python_folder.parent
 metadata_folder = join(pet2bids_folder, 'metadata')
-
-try:
-    # collect metadata jsons in dev mode
-    metadata_jsons = \
-        [Path(join(metadata_folder, metadata_json)) for metadata_json
-         in listdir(metadata_folder) if '.json' in metadata_json]
-except FileNotFoundError:
-    metadata_jsons = \
-        [Path(join(module_folder, 'metadata', metadata_json)) for metadata_json
-         in listdir(join(module_folder, 'metadata')) if '.json' in metadata_json]
 
 # check to see if config file exists
 home_dir = Path.home()
@@ -75,235 +67,6 @@ if pypet2bids_config.exists():
 else:
     # if it doesn't exist use the default one included in this library
     helper_functions.modify_config_file('DEFAULT_METADATA_JSON', module_folder / 'template_json.json')
-
-# create a dictionary to house the PET metadata files
-metadata_dictionaries = {}
-
-for metadata_json in metadata_jsons:
-    try:
-        with open(metadata_json, 'r') as infile:
-            dictionary = json.load(infile)
-
-        metadata_dictionaries[metadata_json.name] = dictionary
-    except FileNotFoundError as err:
-        raise Exception(f"Missing pet metadata file {metadata_json} in {metadata_folder}, unable to validate metadata.")
-    except json.decoder.JSONDecodeError as err:
-        raise IOError(f"Unable to read from {metadata_json}")
-
-
-def check_json(path_to_json, items_to_check=None, silent=False, spreadsheet_metadata={}, **additional_arguments):
-    """
-    This method opens a json and checks to see if a set of mandatory values is present within that json, optionally it
-    also checks for recommended key value pairs. If fields are not present a warning is raised to the user.
-
-    :param spreadsheet_metadata:
-    :type spreadsheet_metadata:
-    :param path_to_json: path to a json file e.g. a BIDS sidecar file created after running dcm2niix
-    :param items_to_check: a dictionary with items to check for within that json. If None is supplied defaults to the
-           PET_metadata.json contained in this repository
-    :param silent: Raises warnings or errors to stdout if this flag is set to True
-    :return: dictionary of items existence and value state, if key is True/False there exists/(does not exist) a
[…]
