### 2024-01-24 8815594
added tests for metadata for ecat cli.
