+++ pypet2bids/tests/test_ecat.py
@@ -1,0 +1,200 @@
+import os
+import subprocess
+import tempfile
+import pathlib
+import json
+import pdb
+from pypet2bids.ecat import Ecat
+
+TESTS_DIR = pathlib.Path(__file__).resolve().parent
+PYPET2BIDS_DIR = TESTS_DIR.parent
+PET2BIDS_DIR = PYPET2BIDS_DIR.parent
+
+# obtain ecat file path
+ecat_file_path = PET2BIDS_DIR / 'ecat_validation' / 'ECAT7_multiframe.v.gz'
+ecatpet2bids = PYPET2BIDS_DIR / 'pypet2bids' / 'ecat_cli.py'
+
+dataset_description_dictionary = {
+  "_Comment": "This is a very basic example of a dataset description json",
+  "Name": "PET Brain phantoms",
+  "BIDSVersion": "1.7.0",
+  "DatasetType": "raw",
+  "License": "CC0",
+  "Authors": [
+    "Author1 Surname1",
+    "Author2 Surname2",
+    "Author3 Surname3",
+    "Author4 Middlename4 Surname4",
+    "Author5 Middlename5 Surname5"
+  ],
+  "HowToAcknowledge": "No worries this is fake.",
+  "ReferencesAndLinks": ["No you aren't getting any", "Don't bother to ask", "Fine, https://fake.fakelink.null"]
+}
+
+
+def test_kwargs_produce_valid_conversion(tmp_path):
+    # prepare a set of kwargs (stolen from a valid bids subject/dataset, mum's the word ;) )
+    full_set_of_kwargs = {
+            "Modality": "PT",
+            "Manufacturer": "Siemens",
+            "ManufacturersModelName": "Biograph 64_mCT",
+            "InstitutionName": "NIH",
+            "InstitutionalDepartmentName": "NIMH MIB",
+            "InstitutionAddress": "10 Center Drive, Bethesda, MD 20892",
+            "DeviceSerialNumber": "60005",
+            "StationName": "MIAWP60005",
+            "PatientPosition": "FFS",
+            "SoftwareVersions": "VG60A",
+            "SeriesDescription": "PET Brain Dyn TOF",
+            "ProtocolName": "PET Brain Dyn TOF",
+            "ImageType": [
+                "ORIGINAL",
+                "PRIMARY"
+            ],
+            "SeriesNumber": 6,
+            "ScanStart": 2,
+            "TimeZero": "10:39:46",
+            "InjectionStart": 0,
+            "AcquisitionNumber": 2001,
+            "ImageComments": "Frame 1 of 33^AC_CT_Brain",
+            "Radiopharmaceutical": "ps13",
+            "RadionuclidePositronFraction": 0.997669,
+            "RadionuclideTotalDose": 714840000.0,
+            "RadionuclideHalfLife": 1220.04,
+            "DoseCalibrationFactor": 30806700.0,
+            "ConvolutionKernel": "XYZ Gauss2.00",
+            "Units": "Bq/mL",
+            "ReconstructionName": "PSF+TOF",
+            "ReconstructionParameterUnits": [
+                "None",
+                "None"
+            ],
+            "ReconstructionParameterLabels": [
+                "subsets",
+                "iterations"
+            ],
+            "ReconstructionParameterValues": [
+                21,
+                3
+            ],
+            "ReconFilterType": "XYZ Gauss",
+            "ReconFilterSize": 2.0,
+            "AttenuationCorrection": "measured,AC_CT_Brain",
+            "DecayFactor": [
+                1.00971
+            ],
+            "FrameTimesStart": [
+                0
+            ],
+            "FrameDuration": [
+                30
+            ],
+            "SliceThickness": 2,
+            "ImageOrientationPatientDICOM": [
[…]
