### 2024-01-25 #6067 @CPernet
getting there, retesting the read the doc (that crashes) and our CI -- @bendhouseart has been tasked to do those
