### 2024-01-25 #6067 @bendhouseart
Yes, as Cyril mentioned I've been chipping away at the issues raised by @adswa and am currently wrapping up some changes holding @CPernet up with respect to [openneuropet/PET2BIDS#236](https://github.com/openneuropet/PET2BIDS/issues/236). I think that now we're all waiting on my completing and merging [openneuropet/PET2BIDS PR #262](https://github.com/openneuropet/PET2BIDS/pull/262).
