### 2024-01-25 e21c2ad
added spreadsheet support to ecat.py and ecat_cli.py
