+++ pypet2bids/pypet2bids/ecat.py
@@ -12,6 +12,7 @@ import nibabel
 import os
 import json
 import pathlib
+import pandas as pd
 
 try:
     import helper_functions
@@ -79,6 +80,16 @@ class Ecat:
         self.kwargs = kwargs
         self.output_path = None
         self.metadata_path = metadata_path
+
+        # load config file
+        default_json_path = helper_functions.check_pet2bids_config('DEFAULT_METADATA_JSON')
+        if default_json_path and pathlib.Path(default_json_path).exists():
+            with open(default_json_path, 'r') as json_file:
+                try:
+                    self.spreadsheet_metadata.update(json.load(json_file))
+                except json.decoder.JSONDecodeError:
+                    logger.warning(f"Unable to load default metadata json file at {default_json_path}, skipping.")
+
         if os.path.isfile(ecat_file):
             self.ecat_file = str(ecat_file)
         else:
@@ -367,6 +378,57 @@ class Ecat:
         else:
             print(json.dumps(helper_functions.replace_nones(self.sidecar_template), indent=4))
 
+    def write_out_blood_files(self, new_file_name_with_entities=None, destination_folder=None):
+        recording_entity = "_recording-manual"
+
+        if not new_file_name_with_entities:
+            new_file_name_with_entities = pathlib.Path(self.nifti_file)
+        if not destination_folder:
+            destination_folder = pathlib.Path(self.nifti_file).parent
+
+        if '_pet' in new_file_name_with_entities.name:
+            if new_file_name_with_entities.suffix == '.gz' and len(new_file_name_with_entities.suffixes) > 1:
+                new_file_name_with_entities = new_file_name_with_entities.with_suffix('').with_suffix('')
+
+            blood_file_name = new_file_name_with_entities.stem.replace('_pet', recording_entity + '_blood')
+        else:
+            blood_file_name = new_file_name_with_entities.stem + recording_entity + '_blood'
+
+        if self.spreadsheet_metadata.get('blood_tsv', {}) != {}:
+            blood_tsv_data = self.spreadsheet_metadata.get('blood_tsv')
+            if type(blood_tsv_data) is pd.DataFrame or type(blood_tsv_data) is dict:
+                if type(blood_tsv_data) is dict:
+                    blood_tsv_data = pd.DataFrame(blood_tsv_data)
+                # write out blood_tsv using pandas csv write
+                blood_tsv_data.to_csv(os.path.join(destination_folder, blood_file_name + ".tsv")
+                                      , sep='\t',
+                                      index=False)
+
+            elif type(blood_tsv_data) is str:
+                # write out with write
+                with open(os.path.join(destination_folder, blood_file_name + ".tsv"), 'w') as outfile:
+                    outfile.writelines(blood_tsv_data)
+            else:
+                raise (f"blood_tsv dictionary is incorrect type {type(blood_tsv_data)}, must be type: "
+                       f"pandas.DataFrame")
+
+        # if there's blood data in the tsv then write out the sidecar file too
+        if self.spreadsheet_metadata.get('blood_json', {}) != {} \
+                and self.spreadsheet_metadata.get('blood_tsv', {}) != {}:
+            blood_json_data = self.spreadsheet_metadata.get('blood_json')
+            if type(blood_json_data) is dict:
+                # write out to file with json dump
+                pass
+            elif type(blood_json_data) is str:
+                # write out to file with json dumps
+                blood_json_data = json.loads(blood_json_data)
+            else:
+                raise (f"blood_json dictionary is incorrect type {type(blood_json_data)}, must be type: dict or str"
+                       f"pandas.DataFrame")
+
[…]
