### 2024-01-25 96b42a3
need to test and then validate sidecar updating
