### 2024-01-26 0ad8e53
added entrypoints for json updating methods
