+++ pypet2bids/pyproject.toml
@@ -42,6 +42,7 @@ pet2bids-spreadsheet-template = 'pypet2bids.helper_functions:write_out_module'
 convert-pmod-to-blood = 'pypet2bids.convert_pmod_to_blood:main'
 ispet = 'pypet2bids.is_pet:main'
 updatepetjsonfromdicom = 'pypet2bids.dcm2niix4pet:update_json_with_dicom_value_cli'
+updatepetjsonfromecat = 'pypet2bids.ecat_cli:update_json_with_ecat_value_cli'
 
 [tool.poetry.group.dev.dependencies]
 pyinstaller = "^5.4.1"
