### 2024-01-29 5f4b343
Shape using number of subheaders rather than number frames from main header
Using the number of frames from the main header causes issues when that number differs from the real number of frames enumerated in the subheaders and in the image data. Instead using the number of subheaders (or alternatively, making an option to use the number of subheaders) instead ensures the data sizes are in agreement.
