### 2024-01-30 c042994
updated logging, running checks on updated jsons
