+++ pypet2bids/pypet2bids/update_json_pet_file.py
@@ -47,7 +47,8 @@ for metadata_json in metadata_jsons:
         raise IOError(f"Unable to read from {metadata_json}")
 
 
-def check_json(path_to_json, items_to_check=None, silent=False, spreadsheet_metadata={}, **additional_arguments):
+def check_json(path_to_json, items_to_check=None, silent=False, spreadsheet_metadata={}, mandatory=True,
+               recommended=True, logger_name='pypet2bids', **additional_arguments):
     """
     This method opens a json and checks to see if a set of mandatory values is present within that json, optionally it
     also checks for recommended key value pairs. If fields are not present a warning is raised to the user.
@@ -62,8 +63,10 @@ def check_json(path_to_json, items_to_check=None, silent=False, spreadsheet_meta
             corresponding entry in the json the same can be said of value
     """
 
+    logger = helper_functions.logger(logger_name)
+
     if silent:
-        logger.disable = True
+        logger.disabled = True
     else:
         logger.disabled = False
 
@@ -90,28 +93,50 @@ def check_json(path_to_json, items_to_check=None, silent=False, spreadsheet_meta
     flattened_spreadsheet_metadata.update(spreadsheet_metadata.get('blood_json', {}))
     flattened_spreadsheet_metadata.update(spreadsheet_metadata.get('blood_tsv', {}))
 
-    for requirement in items_to_check.keys():
-        for item in items_to_check[requirement]:
+    if mandatory:
+        for item in items_to_check['mandatory']:
             all_good = False
-            if item in json_to_check.keys() and json_to_check.get(item, None) or item in additional_arguments or item in flattened_spreadsheet_metadata.keys():
+            if (item in json_to_check.keys() and
+                    (json_to_check.get(item, None) is not None or json_to_check.get(item) != '')
+                    or item in additional_arguments or item in flattened_spreadsheet_metadata.keys()):
                 # this json has both the key and a non-blank value do nothing
                 all_good = True
                 pass
-            elif item in json_to_check.keys() and not json_to_check.get(item, None):
-                logger.warning(f"{item} present but has null value.")
+            elif (item in json_to_check.keys()
+                  and (json_to_check.get(item, None) is None or json_to_check.get(item, None) == '')):
+                logger.error(f"{item} present but has null value.")
                 storage[item] = {'key': True, 'value': False}
             elif not all_good:
-                logger.warning(f"{item} is not present in {path_to_json}. This will have to be "
+                logger.error(f"{item} is not present in {path_to_json}. This will have to be "
                                f"corrected post conversion.")
                 storage[item] = {'key': False, 'value': False}
 
+    if recommended:
+        for item in items_to_check['recommended']:
+            all_good = False
+            if (item in json_to_check.keys() and
+                    (json_to_check.get(item, None) is not None or json_to_check.get(item) != '')
+                    or item in additional_arguments or item in flattened_spreadsheet_metadata.keys()):
+                # this json has both the key and a non-blank value do nothing
+                all_good = True
+                pass
+            elif (item in json_to_check.keys()
+                  and (json_to_check.get(item, None) is None or json_to_check.get(item, None) == '')):
+                logger.info(f"{item} present but has null value.")
+                storage[item] = {'key': True, 'value': False}
+            elif not all_good:
+                logger.info(f"{item} is recommended but not present in {path_to_json}")
+                storage[item] = {'key': False, 'value': False}
+
     return storage
+
 
 def update_json_with_dicom_value(
         path_to_json,
[…]
