+++ pypet2bids/pypet2bids/ecat_cli.py
@@ -14,9 +14,11 @@ from os.path import join
 try:
     import helper_functions
     import Ecat
+    from update_json_pet_file import check_json, check_meta_radio_inputs
 except ModuleNotFoundError:
     import pypet2bids.helper_functions as helper_functions
     from pypet2bids.ecat import Ecat
+    from pypet2bids.update_json_pet_file import check_json, check_meta_radio_inputs
 
 epilog = textwrap.dedent('''
     
@@ -236,10 +238,11 @@ def main():
         ecat.update_pet_json(cli_args.update)
 
 
-def update_json_with_ecat_values_cli():
+def update_json_with_ecat_value_cli():
     json_update_cli = argparse.ArgumentParser()
     json_update_cli.add_argument("-j", "--json", help="Path to a json to update file.", required=True)
     json_update_cli.add_argument("-e", "--ecat", help="Path to an ecat file.", required=True)
+    json_update_cli.add_argument("-m", "--metadata-path", help="Path to a spreadsheet containing PET metadata.")
     json_update_cli.add_argument("-k", "--additional-arguments", nargs='*', action=helper_functions.ParseKwargs, default={},
                                  help="Include additional values in the sidecar json or override values extracted "
                                       "from the supplied ECAT or metadata spreadsheet. "
@@ -253,8 +256,11 @@ def update_json_with_ecat_values_cli():
     args = json_update_cli.parse_args()
 
     update_ecat = Ecat(ecat_file=args.ecat, nifti_file=None, collect_pixel_data=True,
-                       metadata_path=args.metadata_path, kwargs=args.additional_arguments
+                       metadata_path=args.metadata_path, kwargs=args.additional_arguments)
     update_ecat.update_pet_json(args.json)
+
+    # lastly check the json
+    check_json(args.json, logger='check_json', silent=False)
 
 
 if __name__ == "__main__":
