### 2024-01-30 10a5848
bump version, add help text
