### 2024-01-30 09a10bf
Externalizing update PET json's (#262)
* begining refactoring to externalize json commands
* added tests for metadata for ecat cli.
* added spreadsheet support to ecat.py and ecat_cli.py
* got blood sheets working for ecat
* need to test and then validate sidecar updating
* added entrypoints for json updating methods
