+++ pypet2bids/pypet2bids/ecat.py
@@ -12,6 +12,7 @@ import nibabel
 import os
 import json
 import pathlib
+import pandas as pd
 
 try:
     import helper_functions
@@ -19,16 +20,19 @@ try:
     import read_ecat
     import ecat2nii
     import dcm2niix4pet
+    from update_json_pet_file import get_metadata_from_spreadsheet, check_meta_radio_inputs
 except ModuleNotFoundError:
     import pypet2bids.helper_functions as helper_functions
     import pypet2bids.sidecar as sidecar
     import pypet2bids.read_ecat as read_ecat
     import pypet2bids.ecat2nii as ecat2nii
     import pypet2bids.dcm2niix4pet as dcm2niix4pet
+    from pypet2bids.update_json_pet_file import get_metadata_from_spreadsheet, check_meta_radio_inputs
 
 from dateutil import parser
 
 logger = helper_functions.logger('pypet2bids')
+
 
 def parse_this_date(date_like_object) -> str:
     """
@@ -51,7 +55,8 @@ class Ecat:
     viewing in stdout. Additionally, this class can be used to convert an ECAT7.X image into a nifti image.
     """
 
-    def __init__(self, ecat_file, nifti_file=None, decompress=True, collect_pixel_data=True):
+    def __init__(self, ecat_file, nifti_file=None, decompress=True, collect_pixel_data=True, metadata_path=None,
+                 kwargs={}):
         """
         Initialization of this class requires only a path to an ecat file.
 
@@ -69,9 +74,24 @@ class Ecat:
         self.decay_factors = []  # stored here
         self.sidecar_template = sidecar.sidecar_template_full  # bids approved sidecar file with ALL bids fields
         self.sidecar_template_short = sidecar.sidecar_template_short  # bids approved sidecar with only required bids fields
+        self.sidecar_path = None
         self.directory_table = None
+        self.spreadsheet_metadata = {'nifti_json': {}, 'blood_tsv': {}, 'blood_json': {}}
+        self.kwargs = kwargs
+        self.output_path = None
+        self.metadata_path = metadata_path
+
+        # load config file
+        default_json_path = helper_functions.check_pet2bids_config('DEFAULT_METADATA_JSON')
+        if default_json_path and pathlib.Path(default_json_path).exists():
+            with open(default_json_path, 'r') as json_file:
+                try:
+                    self.spreadsheet_metadata.update(json.load(json_file))
+                except json.decoder.JSONDecodeError:
+                    logger.warning(f"Unable to load default metadata json file at {default_json_path}, skipping.")
+
         if os.path.isfile(ecat_file):
-            self.ecat_file = ecat_file
+            self.ecat_file = str(ecat_file)
         else:
             raise FileNotFoundError(ecat_file)
 
@@ -113,6 +133,25 @@ class Ecat:
             self.nifti_file = os.path.splitext(self.ecat_file)[0] + ".nii"
         else:
             self.nifti_file = nifti_file
+
+        # que up metadata path for spreadsheet loading later
+        if self.metadata_path:
+            if pathlib.Path(metadata_path).is_file() and pathlib.Path(metadata_path).exists():
+                self.metadata_path = metadata_path
+        elif metadata_path == '':
+            self.metadata_path = pathlib.Path(self.ecat_file).parent
+        else:
+            self.metadata_path = None
+
+        if self.metadata_path:
+            load_spreadsheet_data = get_metadata_from_spreadsheet(metadata_path=self.metadata_path,
+                                                                  image_folder=pathlib.Path(self.ecat_file).parent,
+                                                                  image_header_dict={})
+
+            self.spreadsheet_metadata['nifti_json'].update(load_spreadsheet_data['nifti_json'])
[…]
