### 2024-01-30 8a727c1
added generic json updater for use without dicom or ecat
