### 2024-01-30 5705138
255 externalize function (#266)
* added generic json updater for use without dicom or ecat
