+++ paper.md
@@ -81,7 +81,7 @@ _File conversion_: The conversion for PET data stored in DICOM format to NIfTI i
 
 _PET Metadata_: JSON sidecar files created from reading PET scanner data are always missing some of the
 radiotracer and pharmaceutical information. To accommodate this, a dedicated PET JSON updater was
-created. The PET JSON updater functions (updatejsonpetfile.m and xxx.py) takes the existing Nifti sidecar JSON file and new metadata to add as input, checks that the full BIDS specification is respected (correct metadata but also consistency of metadata values for the different metadata keys) and updates the image sidecar JSON file. These functions are called by the dcm2niix4pet and ecat2nii (.m and .py) wrapper functions, but can just as easily be called directly allowing already converted image data to become BIDS compliant.
+created. The PET JSON updater functions (updatejsonpetfile.m and update_json_pet_file.py) takes the existing Nifti sidecar JSON file and new metadata to add as input, checks that the full BIDS specification is respected (correct metadata but also consistency of metadata values for the different metadata keys) and updates the image sidecar JSON file. These functions are called by the dcm2niix4pet and ecat2nii (.m and .py) wrapper functions, but can just as easily be called directly allowing already converted image data to become BIDS compliant.
 
 _Spreadsheet conversion_: tabular data formats are ubiquitous in the PET community in particular to keep track of radiotracer information injected per participant and recording of time and radiotracer concentration from the blood sampling. To facilitate conversion to BIDS, three dedicated functions were created to (i) convert pre-formatted tabular data to image sidecar JSON files, or(ii) use pre-formatted tabular data to update image sidecar JSON files, or (iii) convert a tabular PMOD file to a BIDS blood.tsv file with it's sidecar JSON. (PMOD being a popular commercial pharmacokinetic modelling software [@Burger1997]). Supported formats are .xls, .xlsx, .csv, .tsv and .bld.
 
