### 2024-01-31 #6067 @CPernet
@nbeliy response posted in https://github.com/openneuropet/PET2BIDS/issues/236 with the text updated - thx
