### 2024-01-31 #6067 @adswa
minor nits:
line 46: Fro -> For
line 52: takes -> take (or make "PET JSON updater functions" singular)
line 62: missing space
