### 2024-02-02 #6067 @pjtoussaint
Hello @CPernet : I had trouble with installation on one platform - maybe it would be good to add more details about the platform requirements.
