### 2024-02-02 2f00b74
catch typos from @adswa
solved typos reported in https://github.com/openjournals/joss-reviews/issues/6067
