+++ paper.md
@@ -77,13 +77,13 @@ The BIDS extension for Positron Emission Tomography (PET-BIDS) [@norgaard_pet-bi
 
 Conversion tools from native format to BIDS are essential to help researchers to use BIDS. While it exists tools for MRI and MEEG, PET2BIDS is the first tool to convert PET data to BIDS. Because PET2BIDS is a code library, conversion is done using the command line only. PET2BIDS can however also be integrated into software (with a graphical user interface) that aim at more general BIDS conversion, and current efforts are underway integrating PET2BIDS with ezBIDS [@noauthor_ezbids] and BIDSCoins [@zwiers_bidscoin_2022].
 
-_File conversion_: The conversion for PET data stored in DICOM format to NIfTI is performed using the dcm2niix4pet (.m and .py) functions which are wrapper functions around dcm2niix [@rorden_dcm2nii; @li_first_2016] that extend the image sidecar JSON file generated by dcm2niix with details that are not included in the source images but are required for BIDS. The conversion of PET data stored in ECAT format is performed using the newly created ecat2nii (.m .py) functions. The Matlab code relies on the readECAT7.m function writen by BT Christian (1998) and revised by RF Muzic (2002) to read the data, while writing Nifti files relies on nii_tool [@Li_2016]. The Python code was developed to mirror the Matlab code, with further testing of data reading (i.e., which file bits are read according to the PET data frames - see [ecat validation](https://github.com/openneuropet/PET2BIDS/tree/main/ecat_validation)) and writing, relying here on Nibabel [@brett_nipynibabel_2023]. Fro both DICOM and EACT file conversion, JSON sidecar files are created from image data but adding PET metadata provided by users. 
+_File conversion_: The conversion for PET data stored in DICOM format to NIfTI is performed using the dcm2niix4pet (.m and .py) functions which are wrapper functions around dcm2niix [@rorden_dcm2nii; @li_first_2016] that extend the image sidecar JSON file generated by dcm2niix with details that are not included in the source images but are required for BIDS. The conversion of PET data stored in ECAT format is performed using the newly created ecat2nii (.m .py) functions. The Matlab code relies on the readECAT7.m function writen by BT Christian (1998) and revised by RF Muzic (2002) to read the data, while writing Nifti files relies on nii_tool [@Li_2016]. The Python code was developed to mirror the Matlab code, with further testing of data reading (i.e., which file bits are read according to the PET data frames - see [ecat validation](https://github.com/openneuropet/PET2BIDS/tree/main/ecat_validation)) and writing, relying here on Nibabel [@brett_nipynibabel_2023]. For both DICOM and EACT file conversion, JSON sidecar files are created from image data but adding PET metadata provided by users. 
 
 _PET Metadata_: JSON sidecar files created from reading PET scanner data are always missing some of the
 radiotracer and pharmaceutical information. To accommodate this, a dedicated PET JSON updater was
-created. The PET JSON updater functions (updatejsonpetfile.m and update_json_pet_file.py) takes the existing Nifti sidecar JSON file and new metadata to add as input, checks that the full BIDS specification is respected (correct metadata but also consistency of metadata values for the different metadata keys) and updates the image sidecar JSON file. These functions are called by the dcm2niix4pet and ecat2nii (.m and .py) wrapper functions, but can just as easily be called directly allowing already converted image data to become BIDS compliant.
[…]
