### 2024-02-02 #6067 @CPernet
fixed typos from @adswa in https://github.com/openneuropet/PET2BIDS/commit/2f00b749b112e302ae2e3615e2eebcc5b2e9d7f9
