### 2024-02-03 #6067 @CPernet
ok I'm taking this is a default python install that failed - @bendhouseart uses mac everyday, maybe he has some insights about that one
