### 2024-02-05 #6067 @bendhouseart
> The issue happened on a mac (ventura) - I switched to Linux and install ran as described.

@pjtoussaint could you be more specific about the error you encountered? Was it an error with Poetry, Make, building the package, or installing directly from PyPi?

I've noticed "breakage" often enough with Mac updates on my own machine so I tend to rely on the Action's CI to be sure. Perhaps you need to use a brew installed version of Python?
