### 2024-02-05 #6067 @pjtoussaint
Below are my comments for the paper. I saw @nbeliy's comments (issues) and _agree with most of them_.

**Summary**

This is a commendable effort at developing a bidsifier for PET data and metadata. Few attempts at PET standardisation have been made in the past, and this initiative may well be what the community required to encourage the adoption of PET standards.

**General comments**:

- [ ] As indicated by the authors in the text (spelled out name), the CONP is a pan-Canadian collaborative neuroscience initiative and **not a repository**, and it is certainly not specific to brain imaging. I would suggest to remove (or replace) reference to this platform.
- [ ] I understand that the authors want to cite widely used and/or widely publicised data management tools, but as far as I know LORIS offers little (if any) support for BIDS whereas FlyWheel -for instance- has a well-documented BIDS client and provides user assistance. I am happy to discuss this more with the authors, but I would suggest to remove LORIS and cite FlyWheel instead.
- [ ] English (grammar, spelling) needs to be revised.

**Specific comments**:

- [ ] L17 - Perhaps these should not be referred to as ‘major brain imaging repositories’ but rather ‘commonly used brain imaging repositories’? Again, CONP is not a repository and should be removed here.
