### 2024-02-05 b649d8b
very small fix for tsv spreadsheets (#269)
includes fixes for bugs discovered with the help of ezBIDS

is_pet.py would error out on tsv files and recent updates added bug
when dcm2niix4pet had the same input and output directory.
* very small fix for tsv spreadsheets
* included path fix for dcm2niix4pet
