+++ paper.bib
@@ -10,24 +10,17 @@
 	date = {2016-06-21},
 }
 
-@article{norgaard_pet-bids_2022,
-	title = {{PET}-{BIDS}, an extension to the brain imaging data structure for positron emission tomography},
-	volume = {9},
-	rights = {2022 The Author(s)},
-	issn = {2052-4463},
-	url = {https://www.nature.com/articles/s41597-022-01164-1},
-	doi = {10.1038/s41597-022-01164-1},
-	pages = {65},
-	number = {1},
-	journaltitle = {Scientific Data},
-	shortjournal = {Sci Data},
-	author = {Norgaard, Martin and Matheson, Granville J. and Hansen, Hanne D. and Thomas, Adam and Searle, Graham and Rizzo, Gaia and Veronese, Mattia and Giacomel, Alessio and Yaqub, Maqsood and Tonietto, Matteo and Funck, Thomas and Gillman, Ashley and Boniface, Hugo and Routier, Alexandre and Dalenberg, Jelle R. and Betthauser, Tobey and Feingold, Franklin and Markiewicz, Christopher J. and Gorgolewski, Krzysztof J. and Blair, Ross W. and Appelhoff, Stefan and Gau, Remi and Salo, Taylor and Niso, Guiomar and Pernet, Cyril and Phillips, Christophe and Oostenveld, Robert and Gallezot, Jean-Dominique and Carson, Richard E. and Knudsen, Gitte M. and Innis, Robert B. and Ganz, Melanie},
-	urldate = {2023-08-22},
-	date = {2022-03-02},
+@article{Hayashi_2023,
+	title = {brainlife.io: A decentralized and open source cloud platform to support 
+      neuroscience research.},
+	url = {https://arxiv.org/abs/2306.02183v3},
+	doi = {arXiv:2306.02183v3},
+	journaltitle = {{ArXiv [Preprint]}},
+	shortjournal = {{ArXiv}},
+	author = {Hayashi S, Caron BA, Heinsfeld AS, Vinci-Booher S, McPherson B, Bullock DN, Bertò G, Niso G, Hanekamp S, Levitas D, Ray K, MacKenzie A, Kitchell L, Leong JK, Nascimento-Silva F, Koudoro S, Willis H, Jolly JK, Pisner D, Zuidema TR, Kurzawski JW, Mikellidou K, Bussalb A, Rorden C, Victory C, Bhatia D, Baran Aydogan D, Yeh FF, Delogu F, Guaje J, Veraart J, Bollman S, Stewart A, Fischer J, Faskowitz J, Chaumon M, Fabrega R, Hunt D, McKee S, Brown ST, Heyman S, Iacovella V, Mejia AF, Marinazzo D, Craddock RC, Olivetti E, Hanson JL, Avesani P, Garyfallidis E, Stanzione D, Carson J, Henschel R, Hancock DY, Stewart CA, Schnyer D, Eke DO, Poldrack RA, George N, Bridge H, Sani I, Freiwald WA, Puce A, Port NL, Pestilli F},
+	urldate = {2024-02-06},
+	date = {2023-08-11},
 	langid = {english},
-	note = {Number: 1
-Publisher: Nature Publishing Group},
-	keywords = {Bioinformatics, Magnetic resonance imaging, Neuroscience, Positron-emission tomography, Software},
 }
 
 @article{landis_coins_2016,
@@ -46,38 +39,6 @@ Publisher: Nature Publishing Group},
 	date = {2016-01-01},
 	langid = {english},
 	keywords = {{COINS}, Data Exchange, Data sharing, {DX}, Neuroinformatics},
-}
-
-@article{das_loris_2012,
-	title = {{LORIS}: a web-based data management system for multi-center studies},
-	volume = {5},
-	issn = {1662-5196},
[…]
