+++ paper.md
@@ -69,7 +69,7 @@ bibliography: paper.bib
 
 # Summary
 
-The Brain Imaging Data Structure [@gorgolewski_brain_2016] is a standard for organizing and naming neuroimaging data which has quickly become successful and popular in the community with adoption by brain imaging repositories (e.g. OpenNeuro [@noauthor_openneuro], PublicnEUro [@noauthor_public]), data management tools (e.g. COINS [@landis_coins_2016], XNAT [@marcus_extensible_2007]) and computational platforms (e.g. BrainLife []). BIDS not only allows data to be shared much more easily, but also enables the development of automated data analysis pipelines, and together improves reproducibility.  
+The Brain Imaging Data Structure [@gorgolewski_brain_2016] is a standard for organizing and naming neuroimaging data which has quickly become successful and popular in the community with adoption by brain imaging repositories (e.g. OpenNeuro [@noauthor_openneuro], PublicnEUro [@noauthor_public]), data management tools (e.g. COINS [@landis_coins_2016], XNAT [@marcus_extensible_2007]) and computational platforms (e.g. BrainLife [@Hayashi_2023]). BIDS not only allows data to be shared much more easily, but also enables the development of automated data analysis pipelines, and together improves reproducibility.  
 
 The BIDS extension for Positron Emission Tomography (PET-BIDS) [@norgaard_pet-bids_2022] provides a structured data and metadata nomenclature, including all the necessary information to share and report on PET blood and metabolite [@knudsen_guidelines_2020]. Here we present a new code library, developed in both Matlab and Python, allowing the conversion of DICOM [@dicm_ref_2020] and ECAT (CTI/Siemens proprietary data format) PET imaging data and metadata (e.g., timing information such as e.g. 'time zero' or blood measurements) into the BIDS specification.
 
