+++ paper.md
@@ -69,21 +69,20 @@ bibliography: paper.bib
 
 # Summary
 
-The Brain Imaging Data Structure [@gorgolewski_brain_2016] is a standard for organizing and naming neuroimaging data which has quickly become successful and popular in the community with adoption by brain imaging repositories (e.g. OpenNeuro [@noauthor_openneuro], PublicnEUro [@noauthor_public]), data management tools (e.g. COINS [@landis_coins_2016], XNAT [@marcus_extensible_2007]) and computational platforms (e.g. BrainLife [@Hayashi_2023]). BIDS not only allows data to be shared much more easily, but also enables the development of automated data analysis pipelines, and together improves reproducibility.  
+The Brain Imaging Data Structure [@gorgolewski_brain_2016] is a standard for organizing and naming neuroimaging data, which has quickly become successful and popular in the community with adoption by brain imaging repositories (e.g., OpenNeuro [@noauthor_openneuro], PublicnEUro [@noauthor_public]), data management tools (e.g., COINS [@landis_coins_2016], XNAT [@marcus_extensible_2007]) and computational platforms (e.g. BrainLife [@Hayashi_2023]). BIDS allows data to be shared much more efficiently, enables the development of automated data analysis pipelines, and improves reproducibility.  
 
-The BIDS extension for Positron Emission Tomography (PET-BIDS) [@norgaard_pet-bids_2022] provides a structured data and metadata nomenclature, including all the necessary information to share and report on PET blood and metabolite [@knudsen_guidelines_2020]. Here we present a new code library, developed in both Matlab and Python, allowing the conversion of DICOM [@dicm_ref_2020] and ECAT (CTI/Siemens proprietary data format) PET imaging data and metadata (e.g., timing information such as e.g. 'time zero' or blood measurements) into the BIDS specification.
+The BIDS extension for Positron Emission Tomography (PET-BIDS) [@norgaard_pet-bids_2022] provides nomenclature for structured data and metadata, including all the necessary information to share and report on PET blood and metabolite [@knudsen_guidelines_2020]. Here we present a new code library, developed in both Matlab and Python, allowing the conversion of DICOM [@dicm_ref_2020] and ECAT (CTI/Siemens proprietary data format) PET imaging data and metadata (e.g., timing information such as e.g. 'time zero' or blood measurements) into the BIDS specification.
 
 # Statement of need
 
-Conversion tools from native format to BIDS are essential to help researchers to use BIDS. While it exists tools for MRI and MEEG, PET2BIDS is the first tool to convert PET data to BIDS. Because PET2BIDS is a code library, conversion is done using the command line only. PET2BIDS can however also be integrated into software (with a graphical user interface) that aim at more general BIDS conversion, and current efforts are underway integrating PET2BIDS with ezBIDS [@noauthor_ezbids] and BIDSCoins [@zwiers_bidscoin_2022].
+Conversion tools from native format to BIDS are essential to help researchers to use BIDS. While tools for MRI and MEEG exist, PET2BIDS is the first tool to convert PET data to BIDS. Because PET2BIDS is a code library, conversion is done using the command line only. PET2BIDS can, however, also be integrated into software (with a graphical user interface) that aims at more general BIDS conversion, and current efforts are underway integrating PET2BIDS with ezBIDS [@noauthor_ezbids] and BIDSCoins [@zwiers_bidscoin_2022].
 
[…]
