### 2024-02-06 #6067 @CPernet
summary fixed by removing 'major',  'CONP', 'loris' but adding the computational platform brainlife  (https://github.com/openneuropet/PET2BIDS/commit/f85cf19bfd131738c5f7462c228d7f2df5225b40)

(re CONP, I was thinking about CBRAIN that uses CONP on BIDS data, but it's mixing computation and repo, so I just removed them all for cliarity)
