### 2024-02-06 #6067 @CPernet
English fixed with Grammarly (https://github.com/openneuropet/PET2BIDS/commit/28fc3f6b6d39caa67e7905625ca4d045a962b230)
