### 2024-02-09 cf21bad
added small script to update ecat mainheader with, bump version
