+++ pypet2bids/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pypet2bids"
-version = "1.3.2"
+version = "1.3.3"
 description = "A python implementation of an ECAT to BIDS converter."
 authors = ["anthony galassi <28850131+bendhouseart@users.noreply.github.com>"]
 license = "MIT"
@@ -44,6 +44,7 @@ ispet = 'pypet2bids.is_pet:main'
 updatepetjsonfromdicom = 'pypet2bids.dcm2niix4pet:update_json_with_dicom_value_cli'
 updatepetjsonfromecat = 'pypet2bids.ecat_cli:update_json_with_ecat_value_cli'
 updatepetjson = 'pypet2bids.update_json:update_json_cli'
+ecatheaderupdate = 'pypet2bids.ecat_header_update:main'
 
 [tool.poetry.group.dev.dependencies]
 pyinstaller = "^5.4.1"
