### 2024-02-10 openneuropet/PET2BIDS#236 @CPernet
updated are per other reviewer request as well
