### 2024-02-10 #6067 @CPernet
I have posted the requested correction :-) the only thing left is a special install issue on @pjtoussaint mac (while other mac users have not add issues) -- so it's idiosyncratic but we need more details to fix that (as requested by @bendhouseart)

@adswa issues all fixed
@nbeliy posted in https://github.com/openneuropet/PET2BIDS/issues/236 to be validated when back from winter holidays
