+++ paper.md
@@ -75,7 +75,7 @@ The BIDS extension for Positron Emission Tomography (PET-BIDS) [@norgaard_pet-bi
 
 # Statement of need
 
-Conversion tools from native format to BIDS are essential to help researchers to use BIDS. While tools for MRI and MEEG exist, PET2BIDS is the first tool to convert PET data to BIDS. Because PET2BIDS is a code library, conversion is done using the command line only. PET2BIDS can, however, also be integrated into software (with a graphical user interface) that aims at more general BIDS conversion, and current efforts are underway integrating PET2BIDS with ezBIDS [@noauthor_ezbids] and BIDSCoins [@zwiers_bidscoin_2022].
+Conversion tools from native format to BIDS are essential to help researchers to use BIDS. While tools for MRI and MEEG exist, PET2BIDS is the first tool to convert PET data to BIDS. Because PET2BIDS is a code library, conversion is done using the command line only. PET2BIDS can, however, also be integrated into software (with a graphical user interface) that aims at more general BIDS conversion, and current efforts are underway integrating PET2BIDS with ezBIDS [ezBIDS2024] and BIDSCoins [@zwiers_bidscoin_2022].
 
 _File conversion_: The conversion for PET data stored in DICOM format to NIfTI is performed using the dcm2niix4pet (.m and .py) functions which are wrapper functions around dcm2niix [@rorden_dcm2nii; @li_first_2016] that extend the image sidecar JSON file generated by dcm2niix with details that are not included in the source images but are required for BIDS. The conversion of PET data stored in ECAT format is performed using the newly created ecat2nii (.m .py) functions. The Matlab code relies on the readECAT7.m function written by BT Christian (1998) and revised by RF Muzic (2002) to read the data while writing Nifti files relies on nii_tool [@Li_2016]. The Python code was developed to mirror the Matlab code, with further testing of data reading (i.e., which file bits are read according to the PET data frames - see [ecat validation](https://github.com/openneuropet/PET2BIDS/tree/main/ecat_validation)) and writing, relying here on Nibabel [@brett_nipynibabel_2023]. Mirroring the DICOM conversion, the ecat2nii functions generate JSON sidecar files from image data, and users must provide additional metadata to make files BIDS-compliant.  
 
