+++ paper.bib
@@ -136,6 +136,15 @@ volume = {38},
 year = {1997}
 }
 
+@article{ezBIDS2024,
+author = {Levitas, Daniel and Hayashi, Soichi and Vinci-Booher, Sophia and Heinsfeld, Anibal, and Bhatia, Dheeraj and Lee, Nicholas and Galassi, Anthony and Niso, Guiomar and Pestilli, Franco},
+title = {{ezBIDS: Guided standardization of neuroimaging data interoperable with major data archives and platforms.}},
+doi = {10.1038/s41597-024-02959-0},
+journal = {Scientifc Data},
+volume = {11},
+year = {2024}
+} 
+
 @online{noauthor_openneuro,
 	title = {{OpenNeuro}},
 	url = {https://openneuro.org/},
@@ -163,12 +172,6 @@ year = {1997}
 	author = {Li, X.},
 	urldate = {2024-01-08},
 	date = {2016},
-}
-@software{noauthor_ezbids,
-	title = {ezBIDS},
-	url = {https://brainlife.io/ezbids/},
-	urldate = {2023-08-22},
-	date = {2023},
 }
 
 @software{rorden_dcm2nii,
