### 2024-02-14 openneuropet/PET2BIDS#236 @nbeliy
I feel that the structure of the article can be better.

I would suggest to add a section `Description of functionality`, which you can start with saying what exactly pet2bids do (conversion, organization, metadata extension and blood export). Then the paragraphs with technical details will be placed in it's context.
