### 2024-02-14 openneuropet/PET2BIDS#236 @CPernet
that is not the JOSS format - I do not want to change their structure
