### 2024-02-14 openneuropet/PET2BIDS#236 @nbeliy
Paragraph on spreadsheet conversion is still quite heavy, you can rewrite it in much more concise way, espetially that there very little difference between (i) and (ii). One just create JSON, other updates JSON. You can say that pet2bids allow to extract blood data from pre-formatted tabular data (including PMOD one) and supplement the sidecar JSON together with blood.tsv.
