### 2024-02-14 openneuropet/PET2BIDS#236 @nbeliy
Are all technical details in file conversion really needed? As reader, I don't really care that you use dcm2niix or Nibabel. I feel that it can be replaced by how and what kind of metadata user can and must supply.
