### 2024-02-14 openneuropet/PET2BIDS#236 @CPernet
'Paragraph on spreadsheet conversion is still quite heavy, you can rewrite it in much more concise way, espetially that there very little difference between (i) and (ii). One just create JSON, other updates JSON. You can say that pet2bids allow to extract blood data from pre-formatted tabular data (including PMOD one) and supplement the sidecar JSON together with blood.tsv.'

sure - done https://github.com/openneuropet/PET2BIDS/commit/3e28b32178575ae9e9b218e86991744fcd26090f#diff-0403ef06adf405f7b310b4518bd6a3559854f54c61676f676ce9cbfee7172ab6
