### 2024-02-14 openneuropet/PET2BIDS#236 @CPernet
but you convert speadsheet to json, DICOM and ecst to Nifti
