### 2024-02-14 openneuropet/PET2BIDS#236 @nbeliy
The starting phrase of PET metadata is still awkward. `Some ... information are always missing` sounds like a truism. You can rephrase it into something like: Missing medata from conversions (e.g. radiotracer and pharmaceutical information) can be completed by PET JSON updater.

The important point is that JSON updator updates json and check for consistency and missing information, everything else is not important.
