### 2024-02-14 openneuropet/PET2BIDS#236 @CPernet
Some ... information are always missing sounds like a truis ?? ok you last comment was tat surely is not always
I do not want to change that -- it has to be stated
