### 2024-02-14 openneuropet/PET2BIDS#236 @CPernet
I would mention the CLI only at the end paragraph of L30.

I could --> 'Note that the python tools are command line tools, i.e. they can be called directly from a terminal. ' ?
