### 2024-02-14 openneuropet/PET2BIDS#236 @nbeliy
> that is not the JOSS format - I do not want to change their structure

If you want, but you still need a paragraph providing an overview of what exactly pet2bids do. It may be short, but without it, `File conversion`, `Metadata` and `spreadsheet` looks out of place and technical.

It will also resolve the issue that you never mention that pet2bids sort dataset into BIDS folder structure.

P.S. There was a way to start a tread to keep stuff organized. Can't find it( If you know how to do it, please help
