### 2024-02-14 openneuropet/PET2BIDS#236 @CPernet
overview of what exactly pet2bids do.  -- that what the statement of need does as per JOSSS structure -- 'Here we present a new code library, developed in both Matlab and Python, allowing the conversion of DICOM [@dicm_ref_2020] and ECAT (CTI/Siemens proprietary data format) PET imaging data and metadata (e.g., timing information such as e.g. 'time zero' or blood measurements) into the BIDS specification.'
