### 2024-02-14 openneuropet/PET2BIDS#236 @CPernet
pet2bids sort dataset into BIDS folder structure. ?? but it does not
