### 2024-02-14 openneuropet/PET2BIDS#236 @nbeliy
> it does, because we are talking quantitative data - not relative like MRI - that's why we have test counting the rounding errors and stuff like that - so for PET people it is essential IMO

I understand the importance of testing. But, why do you not test conversion from DICOM, or conversion from ECAT in matlab?

It's not the problem that you mention the tests, but how you do it.
