### 2024-02-14 openneuropet/PET2BIDS#236 @CPernet
I understand the importance of testing. But, why do you not test conversion from DICOM, or conversion from ECAT in matlab?

what ?? conversion form DICOMN is tested by dcm2niix - conversion from ecat is tested both matlab and python - I do not get your point
