### 2024-02-14 openneuropet/PET2BIDS#236 @CPernet
- [x]  Line 31: is the first tool to convert --> is the first tool, in our knowledge, to organize. I know at least few tools that can organize PET dataset at least to some extent. Either go with First PET dedicated, or stress that your tool do organization in the most complete way.

That's right, some tools can organize - none can convert

- [x]  Line 37, 41: If both in python and matlab, functions shares the same name, you don't need to precise extensions. Especially if in python, .py is a script, not a function.

fine

- [x]  Line 36-39: Phrase too long, difficult to read. On top of it, functions which are wrapper functions is ugly.

slightly updated, left wrapper - your opinion the name is 'ugly'

- [x]  Line 38: That as it is placed here, refers to dcm2niix, not to dcm2niix4pet. Also it is not clear how the details not from source images are retrieved, are they supplied by user?

added

- [x] Line 40: the newly created contradicts their reliance on code of Christian and Muzic

I don't think so, the new converter relies on a function that reads

- [x] Line 42: If you can't provide reference to library from which readECAT7 was taken, move Christian and Muzic to acknowledgments. And make their names less DJ, for ex. B. T. Christian.

why? there is no convention for that in the JOSS ref system

- [x] Line 44: I would omit testing part. First it implies that matlab functions are untested, second, testing is part of unit tests, not of description of a specific function.

requested by over reviewer

- [x] Line 47: change and  must provide to with user provided

?? doesn't work as a sentence

- [x] Line 50: Why the update_json_pet_file name is different in matlab and in python?

idiosyncrasies of the code

- [x] Line 54: The mention that dcm2niix4pet make a call to update_json_pet_file is pointless. You already stated that user can provide additional metadata to conversion, and the fact that you describe update_json_pet_file means that user can call it.

fixed

- [x] Line 62: The point (iii) implies that spreadsheet conversion extracts data for blood.tsv onlyfrom PMOD file? Why it can't extract blood information from tabular data?

PMOD is a very specific tabular type - what can it not do from others? because there is no standard for that

- [x] Line 62: Missing PMOD reference.

no it's there  - @Burger1997

https://github.com/openneuropet/PET2BIDS/commit/ac4a7354f5756e2157f45ec1e80b4ded6270640a#diff-0403ef06adf405f7b310b4518bd6a3559854f54c61676f676ce9cbfee7172ab6
-------------
