### 2024-02-14 openneuropet/PET2BIDS#236 @nbeliy
> pet2bids sort dataset into BIDS folder structure. ?? but it does not (hence not 'mentioned')

Exactly the point. BIDS imposes folder structure, specific naming, specific metadata and specific format. When you tell that you convert to BIDS, it is understood that you do everything. If you do only format and metadata, it must be told explicitly. And probably motivate why you don't do folder structure (which is the easiest part of bidsification).

For example you can say that the goal of your library is to manage the complex BIDS-PET metadata, and it is supposed to be used from inside of other bidsifiers like ezBIDS and BIDSCOIN.
