### 2024-02-14 c026ad4
BIDS files vs structure
