### 2024-02-14 openneuropet/PET2BIDS#236 @CPernet
got you now, makes sense --> https://github.com/openneuropet/PET2BIDS/commit/c026ad49ebf187211a2ddeb2ec65053901ba8c96#diff-0403ef06adf405f7b310b4518bd6a3559854f54c61676f676ce9cbfee7172ab6
