### 2024-02-14 openneuropet/PET2BIDS#236 @nbeliy
> what ?? conversion form DICOM is tested by dcm2niix - conversion from ecat is tested both matlab and python - I do not get your point

It's not what you write, it's how you write.

`The Python code was developed to mirror the Matlab code, with further testing of data reading`. Read literally, it tells 2 things: python code (ecat2nii) contains the test of data reading, and Matlab do not contain this test. It also imply that dcm2niix4pet do not contain such tests.

I'm still not sure what exactly you wanted to say, but  your point is that because ecat2nii is home made function, we did additional tests to ensure the quality of conversion. It could look something like:

In order to insure the quality of conversion (especially important in quantitative imagery, like PET), we developed a python tool <name> that compares the original ECAT image with converted one.
