### 2024-02-14 6a9c530
better explain ecat testing
