### 2024-02-14 openneuropet/PET2BIDS#236 @CPernet
again, got you know -- again makes sense!
https://github.com/openneuropet/PET2BIDS/commit/6a9c530427358543ff5060c2f1f5a8724f5f09a1#diff-0403ef06adf405f7b310b4518bd6a3559854f54c61676f676ce9cbfee7172ab6
