### 2024-02-14 openneuropet/PET2BIDS#236 @CPernet
latest requests https://github.com/openneuropet/PET2BIDS/commit/8e3953020fb0cf1c046cf32bee0e79f0fde43fcb#diff-0403ef06adf405f7b310b4518bd6a3559854f54c61676f676ce9cbfee7172ab6
