+++ paper.md
@@ -82,7 +82,7 @@ _File conversion_: The conversion for PET data stored in DICOM format to NIfTI i
 _PET Metadata_: Some radiotracer and pharmaceutical information are always missing in the JSON sidecar files created from reading PET scanner data. To accommodate this, a dedicated PET JSON updater was
 created. The PET JSON updater functions (updatejsonpetfile.m or update_json_pet_file.py) take the existing Nifti sidecar JSON file and new metadata to add as input, checks that the full BIDS specification is respected (correct metadata but also consistency of metadata values for the different metadata keys) and updates the image sidecar JSON file, allowing already converted image data to become BIDS compliant.  
 
-_Spreadsheet conversion_: Tabular data formats are ubiquitous in the PET community, mainly to keep track of radiotracer information injected per participant and record time and radiotracer concentration from the blood sampling. Functions were created to read pre-formatted tabular data to either create or update existing sidecar JSON files. In addition dedicated functions were created to convcert the tabular PMOD file(s) to a BIDS blood.tsv file with it's sidecar JSON. (PMOD being a popular commercial pharmacokinetic modeling software [@Burger1997]). Supported formats are .xls, .xlsx, .csv, .tsv and .bld.
+_Spreadsheet conversion_: Tabular data formats are ubiquitous in the PET community, mainly to keep track of radiotracer information injected per participant and record time and radiotracer concentration from the blood sampling. Functions were created to read pre-formatted tabular data to either create or update existing sidecar JSON files. In addition dedicated functions were created to convert either a preformated tabular file or PMOD file (PMOD being a popular commercial pharmacokinetic modeling software [@Burger1997]) to a BIDS blood.tsv file with it's sidecar JSON. Supported formats are .xls, .xlsx, .csv, .tsv and .bld.
 
 # Acknowledgements
 
