### 2024-02-14 openneuropet/PET2BIDS#236 @CPernet
1 - blood data are unusual and do not come with all imaging
2 - the image json does not contain blood information -- this is optional additional metadata
3 - you were right (reworded) we propose a preformated table, or PMOD (https://github.com/openneuropet/PET2BIDS/tree/main/spreadsheet_conversion/blood)
3 - PMOD is not one file but several files to be combined
