### 2024-02-14 #6067 @adswa
Hey, I'm concluding my review. :) I've installed and played with the Python-part of the software on a Debian bookworm workstation, and tested the functionality following the documented walk-throughs with phantom data. The issues I've raised were sufficiently addressed, so I'm happy to tick all boxes. If there is anything else I can do, please let me know within the next month - I will be unavailable from mid-March until July. :)
