### 2024-02-14 f1127f5
Update ecat2nii.py
Remove extra scaling applied again via nifti `scl_slope`.
