### 2024-02-15 a349d52
add conditional statement
ecat validation unchanged
