### 2024-02-16 bdad449
Update pypet2bids/pypet2bids/ecat2nii.py
