### 2024-02-20 d8502b5
updated to handle uneven rows in input spreadsheet blood data
