### 2024-02-20 f8f4ca5
[Ecat2Nii] Only apply ECAT_CALIBRATION_FACTOR if it has not already b… (#268)
* [Ecat2Nii] Only apply ECAT_CALIBRATION_FACTOR if it has not been applied;

check CALIBRATION_UNITS to see if calibration has been applied, if it has, return input image, otherwise multiply by calibration factor.

!!!!!!!!IMPORTANT NOTE!!!!!!!!!!!

CALIBRATION_UNITS=1 -> Multiply the image by the ECAT_CALIBRATION_FACTOR.
CALIBRATION_UNITS=0 -> Do not multiply the image by the ECAT_CALIBRATION_FACTOR.

This change will correctly account for whether the calibration factor has already been applied to Ecat data. In the 
[…]
