### 2024-02-21 9d76972
added unique SeriesDescription to outputs 4 ezBIDS
* Made use of the SeriesDescription field produced by dcm2niix
  to allow ezBIDS to differentiate between mulitple pet scans
  for the same subject/session even. The SeriesDescription field
  previously housed (observed at least) the reconstruction method.
  Now it gets some PET BIDS entities along with an MD5 hash of the
  TimeZero, InjectedMass, TracerName, ReconMethodName etc to ensure
  that a unique and human readable value is recorded in
  SeriesDescription field; added the same functionality and field
  to the ecat methods
* Sanitized d
[…]
