+++ pypet2bids/pypet2bids/helper_functions.py
@@ -20,6 +20,7 @@ import re
 import shutil
 import typing
 import json
+import hashlib
 import warnings
 import logging
 import dotenv
@@ -979,3 +980,20 @@ class CustomFormatter(logging.Formatter):
         log_fmt = self.FORMATS.get(record.levelno)
         formatter = logging.Formatter(log_fmt)
         return formatter.format(record)
+
+
+def hash_fields(**fields):
+    hash_return_string = ""
+    hash_string = ""
+    keys_we_want = ['ses', 'rec', 'trc']
+    for key, value in fields.items():
+        # sanitize values
+        regex = r"[^a-zA-Z0-9]"
+        value = re.sub(regex, "", str(value))
+        hash_string += f"{key}-{value}_"
+        if key in keys_we_want:
+            hash_return_string += f"{key}-{value}_"
+
+    hash_hex = hashlib.md5(hash_string.encode('utf-8')).hexdigest()
+
+    return f"{hash_return_string}{hash_hex}"
