### 2024-02-22 openneuropet/PET2BIDS#236 @nbeliy
Based on draft of [c8eaf08](https://github.com/openneuropet/PET2BIDS/commit/c8eaf08b55edb8b6522732a1b5b90d6039ed8888).

I collected here the points that paper is either misleading or straight incorrect.

Line 31:
> PET2BIDS is the first tool to convert ... to BIDS

Misleading, implying that you convert  and structure the data following BIDS. I asked several persons and they all reads it this way. To rephrase.

Line 32:
> conversion done using command line **only**

Factually incorrect. You provide CLI only for Python, for matlab you provide functions.

Line 41:
> with details that user **must** provide **because ** ...

Factually incorrect. The functions `dcm2niix4pet` allow user to not provide all required details. To rephrase.

Line 46:
> further testing data reading...

Incomprehensible. To rewrite.

Line 55:
> The PET JSON updater functions (... update_json_pet.py)

Factually incorrect. `.py` are not functions, but files. It can contain (and actually it does) several functions. To correct.
