### 2024-02-22 openneuropet/PET2BIDS#276 @nbeliy
[JOSS review] Documentation review
- [x] Lack of command description for online help (here example of dcm2niix4pet, but issue is present in other commands):

```bash
>dcm2niix4pet --help
usage: dcm2niix4pet [-h] [--metadata-path [METADATA_PATH]]
                    [--translation-script-path TRANSLATION_SCRIPT_PATH]
                    [--destination-path DESTINATION_PATH] [--tempdir TEMPDIR] [--kwargs [KWARGS ...]]
                    [--silent] [--show-examples] [--set-dcm2niix-path SET_DCM2NIIX_PATH]
                    [--set-default-metadata-json SET_DEFAULT_METADATA_JSON] [--trc TRC] [--run RUN]
                    [--rec REC]
                    [folder]

positional arguments:
  folder                Folder path containing imaging data
```

- [x] In readthesedocs, no mentioning of PET JSON updater, a function(s) that you explicitly describe in paper.
- [x] Lack of documentation CLI commands installed by pip, namely all `updatejson*` functions
- [x] Docstring for `update_json_pet_file.check_meta_radio_inputs` needs to be rewrite (`author does not recall everything it does`).
- [x] Repository don't have tags/releases
- [x] Repository do not follow PEP8, flake8 finds > 1000 errors. Outside purely stylistic ones, I saw few ` F541 f-string is missing placeholders`, and few `F405 'pathlib' may be undefined`, that looked dangerous.
