### 2024-02-22 openneuropet/PET2BIDS#236 @CPernet
[x] 'PET2BIDS is the first tool to convert ... to BIDS'  --> made the statement valid  --> .. convert to BIDS files.  (note we already state above that it's only files and not structure)

[x] command line only -- removed only

[x] because -- made it correct --> 'to be BIDS valid'

[x] functions - yeah, after so many changes requested, the whole paragraph lost its meaning -- rephrase entirely. 'Some radiotracer and pharmaceutical information are always missing in the JSON sidecar files created from reading PET scanner data which is why dcm2niix4pet and ecat2nii take additional metadata. It is also possible to update existing JSON sidecar files with new metadata directly using updatejsonpetfile.m or update_json_pet_file.py. '
