+++ paper.md
@@ -77,7 +77,7 @@ The BIDS extension for Positron Emission Tomography (PET-BIDS) [@norgaard_pet-bi
 
 Conversion tools from native format to BIDS are essential to help researchers to use BIDS. While tools for MRI and MEEG exist, PET2BIDS is the first tool to convert both DICOM and ECAT PET data and PET metadata to BIDS files. Because PET2BIDS is a code library, conversion is done using the command line. PET2BIDS can, however, also be integrated into software (with a graphical user interface) that aims at more general BIDS conversion, and current efforts are underway integrating PET2BIDS with ezBIDS [@ezBIDS2024] and BIDSCoins [@zwiers_bidscoin_2022]. While our library allows to convert and name files, these more generic tools also allow data to be structured following the BIDS specification.
 
-_File conversion_: The conversion for PET data stored in DICOM format to NIfTI is performed using the dcm2niix4pet functions, which are wrappers around dcm2niix [@rorden_dcm2nii; @li_first_2016]. Those functions extend the image sidecar JSON file generated by dcm2niix with details that the user must provide to be BIDS valid because they are not included in the source images but are required for BIDS. The conversion of PET data stored in ECAT format is performed using the dedicated ecat2nii functions. The Matlab code relies on the readECAT7.m function written by B.T. Christian (1998) and revised by R.F. Muzic (2002) to read the data while writing Nifti files relies on nii_tool [@Li_2016]. The Python code was developed to mirror the Matlab code, further testing data reading (i.e., which file bits are read according to the PET data frames) and writing, relying here on Nibabel [@brett_nipynibabel_2023]. For both Matlab and Python, the data conversion and writing are thoroughly tested (see [ecat validation](https://github.com/openneuropet/PET2BIDS/tree/main/ecat_validation)). Mirroring the DICOM conversion, the ecat2nii functions generate JSON sidecar files from image data, and users must provide additional metadata to make files BIDS-compliant. Note that the Python tools are command line tools, i.e. they can be called directly from a terminal.  
+_File conversion_: The conversion for PET data stored in DICOM format to NIfTI is performed using the dcm2niix4pet functions, which are wrappers around dcm2niix [@rorden_dcm2nii; @li_first_2016]. Those functions extend the image sidecar JSON file generated by dcm2niix with metadata from the users that are necessary to make files BIDS-compliant. The conversion of PET data stored in ECAT format is performed using the dedicated ecat2nii functions. The Matlab code relies on the readECAT7.m function written by B.T. Christian (1998) and revised by R.F. Muzic (2002) to read the data while writing Nifti files relies on nii_tool [@Li_2016]. The Python code was developed to mirror the Matlab code, further testing data reading (i.e., which file bits are read according to the PET data frames) and writing, relying here on Nibabel [@brett_nipynibabel_2023]. For both Matlab and Python, the data conversion and writing are thoroughly tested (see [ecat validation](https://github.com/openneuropet/PET2BIDS/tree/main/ecat_validation)). Mirroring the DICOM conversion, the ecat2nii functions generate JSON sidecar files from image data, and users must provide additional metadata to make files BIDS-compliant. Note that the Python tools are command line tools, i.e. they can be called directly from a terminal.  
 
 _PET Metadata_: Some radiotracer and pharmaceutical information are always missing in the JSON sidecar files created from reading PET scanner data which is why dcm2niix4pet and ecat2nii take additional metadata. It is also possible to update existing JSON sidecar files with new metadata directly using updatejsonpetfile.m or update_json_pet_file.py.
 
