### 2024-02-22 openneuropet/PET2BIDS#236 @CPernet
does really work either -- changed to 'Those functions extend the image sidecar JSON file generated by dcm2niix with metadata from the users that are necessary to make files BIDS-compliant'
