### 2024-02-22 openneuropet/PET2BIDS#236 @nbeliy
Check if it sounds better:

Those functions can extend the image sidecar JSON file generated by dcm2niix with user-provided metadata, making them BIDS-compliant.

Or:
In order to make image files generated by dcm2niix BIDS compliant, these function can extend image sidecar JSON file with user-provided metadata.

Or even:
Missing, but required by BIDS, metadata from image sidecar JSON file generated by dcm2niix can be completed by user.
