### 2024-02-22 openneuropet/PET2BIDS#236 @nbeliy
What about the `further testing`

> The Python code was developed to mirror the Matlab code, further testing data reading (i.e., which file bits are read according to the PET data frames) and writing, relying here on Nibabel [@brett_nipynibabel_2023].

I propose:

The Python code for PET conversion mirrors the Matlab code, except writing of Nifti files, which are delegated to Nibabel [@brett_nipynibabel_2023].

Or something similar. You mention the tests of conversion directly after, so no need `further data testing`.
