### 2024-02-22 openneuropet/PET2BIDS#236 @CPernet
done
