+++ scripts/python_conversions.sh
@@ -425,8 +425,10 @@ BodyPart='Phantom' \
 Units='Bq/mL' \
 TracerName='FDG' \
 TracerRadionuclide='F18' \
-InjectedRadioactivity=81.24 \
-SpecificRadioactivity=1.3019e+04 \
+InjectedRadioactivity=0.788 \
+InjectedRadioactivityUnits='mCi' \
+SpecificRadioactivity='n/a' \
+SpecificRadioactivityUnits='n/a' \
 ModeOfAdministration='infusion' \
 AcquisitionMode='list mode' \
 ImageDecayCorrected=true \
@@ -434,12 +436,14 @@ ImageDecayCorrectionTime=0 \
 ReconFilterType='none' \
 ReconFilterSize=0 \
 AttenuationCorrection='transmission scan with a 137Cs point source' \
+ScatterCorrectionMethod='Single-scatter simulation' \
 ScanStart=0 \
-InjectionStart=0 \
-ScatterFraction=[0] \
-PromptRate=[0] \
-SinglesRate=[0] \
-RandomRate=[0]
+InjectionStart=-2183 \
+ReconMethodParameterLabels='["subsets", "iterations"]' \
+ReconMethodParameterUnits='["none", "none"]' \
+ReconMethodParameterValues='[16, 2]' \
+ReconFilterType='Gaussian' \
+ReconFilterSize=2
 
 # General Electric Medical Systems Advance
 # -----------------------------------------
@@ -451,24 +455,22 @@ BodyPart='Phantom' \
 Units='Bq/mL' \
 TracerName='FDG' \
 TracerRadionuclide='F18' \
-InjectedRadioactivity=75.8500 \
-SpecificRadioactivity=418713.8 \
+InjectedRadioactivity=0.788 \
+InjectedRadioactivityUnits='mCi' \
+SpecificRadioactivity='n/a' \
+SpecificRadioactivityUnits='n/a' \
 ModeOfAdministration='infusion' \
-FrameTimesStart=[0] \
+ScanStart=0 \
+InjectionStart=-5336 \
+FrameTimesStart="[0]" \
 AcquisitionMode='list mode' \
 ImageDecayCorrected='true' \
 ImageDecayCorrectionTime=0 \
-ScatterCorrectionMethod='Gaussian Fit' \
-FrameDuration=[98000] \
-FrameTimesStart=[0] \
-ReconMethodName='filtered backprojection with ramp filter' \
-ReconMethodParameterLabels=["none"] \
-ReconMethodParameterUnits=["none"] \
-ReconMethodParameterValues="[0]" \
+ScatterCorrectionMethod='Single-scatter simulation' \
+FrameTimesStart="[0]" \
 ReconFilterType="none" \
-ReconFilterSize=1 \
-AttenuationCorrection='2D-acquired transmission scan with a 68Ge rod source' \
-
+ReconFilterSize=0 \
+AttenuationCorrection='2D-acquired transmission scan with a 68Ge pin'
 
 
 
@@ -486,16 +488,19 @@ BodyPart='Phantom' \
 Units='Bq/mL' \
 TracerName='FDG' \
 TracerRadionuclide='F18' \
-InjectedRadioactivity=32.19 \
-SpecificRadioactivity=32190000 \
+InjectedRadioactivity=0.87 \
+InjectedRadioactivityUnits='mCi' \
+SpecificRadioactivity='n/a' \
+SpecificRadioactivityUnits='n/a' \
 ModeOfAdministration='infusion' \
-FrameTimesStart=[0] \
+ScanStart=0 \
+InjectionStart=-2312 \
+FrameTimesStart="[0, 300, 600, 900]" \
 AcquisitionMode='list mode' \
[…]
