### 2024-02-26 06c2b13
add 3DRP but also remove optional fields
