### 2024-02-26 de5ae00
set filtersize to 0 until validator is fixed
