### 2024-02-26 48abce0
allows to not specify FilterSize when type is none
