### 2024-02-26 9beee53
bringing back case 6: int16=>int16
