+++ matlab/readECAT7.m
@@ -178,8 +178,7 @@ for m = 1:nmat
                     warning('old matlab version, using int16 to read')
                     data{m} = int16(fread(fid, [sz(1)*sz(2) sz(3)],'int16'));
                 else
-                    % REALLY MAKE IT A UINT16
-                    data{m} = fread(fid, [sz(1)*sz(2) sz(3)], 'uint16');
+                    data{m} = fread(fid, [sz(1)*sz(2) sz(3)],'int16=>int16');
                 end
             otherwise
                 warning('readECAT7: unrecognized data type');
