### 2024-02-26 cc18eaa
Update SeriesDescription to use ProctocolName as primary field
