+++ pypet2bids/pypet2bids/ecat.py
@@ -367,6 +367,7 @@ class Ecat:
         # this is mostly for ezBIDS, but it helps us to make better use of the series description that
         # dcm2niix generates by default for PET imaging, here we mirror the dcm2niix output for ecats
         collect_these_fields = {
+            'ProtocolName': '',
             'SeriesDescription': '',
             'TracerName': 'trc',
             'InjectedRadioactivity': '',
