### 2024-02-26 d6762b6
Update SeriesDescription to use ProctocolName as primary field (#279)
Removes BIDS entity label from SeriesDescription adds ProtocolName

No reason  to put BIDS entities in a non-BIDS field so removed (mostly
to prevent ezBIDS from aggressively collecting them). ProtocolName 
also seemed like a better fit for browsing through scans via the ezBIDS
interface.
