### 2024-03-04 f6d4993
More phantom conversion fixes
