+++ scripts/matlab_conversions.m
@@ -471,17 +471,23 @@ try
     meta.Units                      = 'Bq/mL';
     meta.TracerName                 = 'FDG';
     meta.TracerRadionuclide         = 'F18';
-    meta.InjectedRadioactivity      = 81.24; % Mbq
-    meta.SpecificRadioactivity      = 1.3019e+04; % ~ 81240000 Bq/ 6240 g
+    meta.InjectedRadioactivity      = 0.788;
+    meta.InjectedRadioactivityUnits = 'mCi';
+    meta.SpecificRadioactivity      = 'n/a';
+    meta.SpecificRadioactivityUnits = 'n/a';
     meta.ModeOfAdministration       = 'infusion';   
     meta.AcquisitionMode            = 'list mode';
     meta.ImageDecayCorrected        = true; % when passing this as string it fails validation
     meta.ImageDecayCorrectionTime   = 0;
-    meta.ReconFilterType            = 'none';
-    meta.ReconFilterSize            = 0;
+    meta.ReconFilterType            = 'Gaussian';
+    meta.ReconFilterSize            = 2;
     meta.AttenuationCorrection      = 'transmission scan with a 137Cs point source';
-    % meta.FrameDuration            = 1200; % not needed, encoded in ecat even for 4D images
-    % meta.FrameTimesStart          = 0;
+    meta.ScatterCorrectionMethod    = 'Single-scatter simulation';
+    meta.ScanStart                  = 0;
+    meta.InjectionStart             = -2183;
+    meta.ReconMethodParameterLabels = ["subsets" "iterations"];
+    meta.ReconMethodParameterLabels = ["none" "none"];
+    meta.ReconMethodParameterValues = [16 2];
     
     out = ecat2nii(fullfile(source,['SiemensHRRT-JHU' filesep 'Hoffman.v']),...
         meta,'gz',true,'FileListOut',fullfile(destination,['sub-SiemensHRRTJHU'  filesep 'pet' filesep 'sub-SiemensHRRTJHU.nii']));
@@ -503,23 +509,25 @@ try
     meta.Units                      = 'Bq/mL';
     meta.TracerName                 = 'FDG';
     meta.TracerRadionuclide         = 'F18';
-    meta.InjectedRadioactivity      = 75.8500; % Mbq
-    meta.SpecificRadioactivity      = 418713.8; % ~ 75850000 Bq/ 6240 g
+    meta.InjectedRadioactivity      = 0.788;
+    meta.InjectedRadioactivityUnits = 'mCi';
+    meta.SpecificRadioactivity      = 'n/a';
+    meta.SpecificRadioactivityUnits = 'n/a';
     meta.ModeOfAdministration       = 'infusion';
+    meta.ScanStart                  = 0;
+    meta.InjectionStart             = -5336;
     meta.FrameTimesStart            = 0;
     meta.AcquisitionMode            = 'list mode';
     meta.ImageDecayCorrected        = 'true';
     meta.ImageDecayCorrectionTime   = 0;
-    meta.ScatterCorrectionMethod    = 'Gaussian Fit';
-    meta.FrameDuration              = 98000;
-    meta.FrameTimesStart            = 0;
-    meta.ReconMethodName            = 'filtered backprojection with ramp filter';
+    meta.ScatterCorrectionMethod    = 'Single-scatter simulation';
+    meta.ReconMethodName            = '3D Reprojection';
     meta.ReconMethodParameterLabels = ["none"];
     meta.ReconParameterUnits        = ["none"];
     meta.ReconMethodParameterValues = [0];
     meta.ReconFilterType            = "none";
-    meta.ReconFilterSize            = 1;
[…]
