+++ pypet2bids/pypet2bids/multiple_spreadsheets.py
@@ -7,9 +7,10 @@ except ModuleNotFoundError:
 
 
 def read_multi_subject_spreadsheets(
-        general_metadata_spreadsheet: pathlib.Path,
-        multiple_subject_spreadsheet: pathlib.Path,
-        **kwargs) -> dict:
+    general_metadata_spreadsheet: pathlib.Path,
+    multiple_subject_spreadsheet: pathlib.Path,
+    **kwargs,
+) -> dict:
     """
     Reads in two spreadsheets as formatted in PET2BIDS/spreadsheet_conversion/many_subject_sheets,
     generic (scanner or subject independent data is supplied via the first argument and subject
@@ -27,7 +28,7 @@ def read_multi_subject_spreadsheets(
     :return: dictionary of subject data extracted from each spreadsheet along with any additional
         kwargs supplied
     :rtype: dict
-    
+
     Anthony Galassi
     Copyright Open NeuroPET team
 
@@ -35,12 +36,17 @@ def read_multi_subject_spreadsheets(
 
     required_fields = load_pet_bids_requirements_json()
 
-    if general_metadata_spreadsheet.is_file() and multiple_subject_spreadsheet.is_file():
+    if (
+        general_metadata_spreadsheet.is_file()
+        and multiple_subject_spreadsheet.is_file()
+    ):
         general_metadata = single_spreadsheet_reader(general_metadata_spreadsheet)
         multiple_subject_metadata = open_meta_data(multiple_subject_spreadsheet)
         multiple_subject_metadata
 
-        column_set = set(list(general_metadata.keys()) + list(multiple_subject_metadata.columns))
+        column_set = set(
+            list(general_metadata.keys()) + list(multiple_subject_metadata.columns)
+        )
 
         column_set = list(column_set)
 
@@ -48,26 +54,39 @@ def read_multi_subject_spreadsheets(
         for k in required_fields.keys():
             for field in required_fields[k]:
                 field_exists = field in column_set
-                if k == 'mandatory' and not field_exists:
-                    logging.warning(f"Input spreadsheet(s) {general_metadata_spreadsheet} and "
-                                    f"{multiple_subject_spreadsheet} are missing required column {field}")
-                elif k == 'recommended' and not field_exists:
-                    logging.info(f"Input spreadsheet(s) {general_metadata_spreadsheet} and "
-                                 f"{multiple_subject_spreadsheet} are missing recommended column {field}")
-                elif k == 'optional':
-                    logging.info(f"Input spreadsheet(s) {general_metadata_spreadsheet} and "
-                                 f"{multiple_subject_spreadsheet} are missing optional column {field}")
+                if k == "mandatory" and not field_exists:
+                    logging.warning(
+                        f"Input spreadsheet(s) {general_metadata_spreadsheet} and "
+                        f"{multiple_subject_spreadsheet} are missing required column {field}"
+                    )
+                elif k == "recommended" and not field_exists:
+                    logging.info(
+                        f"Input spreadsheet(s) {general_metadata_spreadsheet} and "
+                        f"{multiple_subject_spreadsheet} are missing recommended column {field}"
+                    )
+                elif k == "optional":
+                    logging.info(
+                        f"Input spreadsheet(s) {general_metadata_spreadsheet} and "
+                        f"{multiple_subject_spreadsheet} are missing optional column {field}"
+                    )
 
         # check to see if there's a subject column in the multi subject data
-        accepted_column_names = ['participant_id', 'participant', 'subject', 'subject_id']
+        accepted_column_names = [
+            "participant_id",
+            "participant",
+            "subject",
+            "subject_id",
+        ]
         found_column_names = []
         for acceptable in accepted_column_names:
[…]
