### 2024-03-04 openneuropet/PET2BIDS#276 @bendhouseart
> Repository don't have tags/releases

Is tagging in git a requirement? We haven't used tagging as pypi and our pyproject.toml to keep track of version and minor fixes.

>  Repository do not follow PEP8, flake8 finds > 1000 errors. Outside purely stylistic ones, I saw few  F541 f-string is missing placeholders, and few F405 'pathlib' may be undefined, that looked dangerous.

Is this a requirement? It's easy enough to fix, but I don't want to apply it until after this review completes and I've closed out some existing PR's as it changes every file and makes PR's difficult to evaluate. See here -> https://github.com/openneuropet/PET2BIDS/pull/283/files

Would you be satisfied for now if I added formatting to the main Makefile and later updated the CI to run the command as well as a precommit hook or similar?

```make
# Quickly run black on all python files in this repository, local version of the pre-commit hook
black:
	@for file in `find . -name "*.py"`; do \
		black $$file; \
	done
```
