### 2024-03-04 openneuropet/PET2BIDS#276 @bendhouseart
> I saw few F541 f-string is missing placeholders

flake8 doesn't seem to like the following:

```python
 logging.warning(
                            f"Autosampled .bld input for {auto['name']} has {auto['sample_length']} rows"
                            f" and Manually sampled input has {manual['sample_length']} rows. Autosampled blood "
                            f"files should have more rows than manually sampled input files. Check .bld inputs."
                        )
```

But I would argue that it's being a bit pedantic as that's what my linter in Pycharm suggests. I guess the alternative would be:

``` logging.warning(
                            f"Autosampled .bld input for {auto['name']} has {auto['sample_length']} rows"
                            f" and Manually sampled input has {manual['sample_length']} rows. Autosampled blood "
                            "files should have more rows than manually sampled input files. Check .bld inputs."
                        )
```

Which looks worse to me.
