### 2024-03-04 openneuropet/PET2BIDS#276 @nbeliy
Hi @bendhouseart ,

The tags are not a requirement, but you will be asked to create a release for the publication of JOSS. It also would facilitate the reference of the `pet2bids`. But I will not push this point.

For the `flake8`, I didn't checked the f-string issues, just saw that there are. So if you confirm that none of these "malformed" f-strings will cause problems (for example if your linter do not complains), I'll accept it.

I'm more worried for the undefined references in the code. This looks much more scary. But again, if you confirm that these references are not causing problems, I'll accept it. Well, maybe I'll cross-check, just in case)
