### 2024-03-04 openneuropet/PET2BIDS#276 @bendhouseart
I'm on the fence about whether we should tag more as we end up making so many small changes and fixes as new data comes in. Pushing directly to PyPi has made the most sense for us in distributing this to our users that we've just stuck with it. That said, happy to tag with `JOSS` or whatever you ask pending publication.

> I'm more worried for the undefined references in the code. This looks much more scary. But again, if you confirm that these references are not causing problems, I'll accept it. Well, maybe I'll cross-check, just in case)

Yeah, let's fix these as they _might_ be bad, they're probably okay, but they're popping up in an older (not as heavily used if at all) sections of the code.

> Lack of command description for online help (here example of dcm2niix4pet, but issue is present in other commands):

Maintaining web links at the CLI level isn't something I particularly want to delve into at this moment. However, would updating the PyPi page with links to the readthedocs page as well as a rendering of the `README.md` be a good compromise?  See:

Then anyone with a search engine _should_ be able to locate the documentation and source by reaching either PyPi or the github repo?

I'll update the other bits of documentation as you pointed out in this issue and make a PR that you can sign off on.
