### 2024-03-04 d43f8cb
added symlink to README.md updated pypi links
- added links to documention and repository on pypi page
- added symlink from README.md to pypet2bids/README.md for rendering
  on pypi project page.
