+++ pypet2bids/pyproject.toml
@@ -1,13 +1,17 @@
 [tool.poetry]
 name = "pypet2bids"
-version = "1.3.7"
-description = "A python implementation of an ECAT to BIDS converter."
+version = "1.3.75"
+description = "A python library for converting PET imaging and blood data to BIDS."
 authors = ["anthony galassi <28850131+bendhouseart@users.noreply.github.com>"]
 license = "MIT"
 include = [
     'pypet2bids/metadata/*',
     'pypet2bids/pyproject.toml',
+    'pypet2bids/README.md',
     ]
+documentation = "https://pypet2bids.readthedocs.io/en/latest/"
+repository = "https://github.com/openneuropet/pet2bids/"
+readme = "README.md"
 
 [tool.poetry.dependencies]
 python = ">=3.8,<=3.11.6"
