### 2024-03-04 1f71e90
Update pypet2bids/pypet2bids/dcm2niix4pet.py
remove spaces
