### 2024-03-04 4f6f4ba
actually using the one in the config if it's present
