### 2024-03-05 openneuropet/PET2BIDS#276 @nbeliy
> Maintaining web links at the CLI level isn't something I particularly want to delve into at this moment. However, would updating the PyPi page with links to the readthedocs page as well as a rendering of the README.md be a good compromise? See:

Sorry, used old vocabulary. When I said online help, I meant the help you get using `--help` option. When you initialize the argument parser, you just need to add one line of what the given program does. For example:
```python
parser = argparse.ArgumentParser(formatter_class=argparse.RawDescriptionHelpFormatter,
                                                                description="Converts DICOM PET images to Nifti and extract metadata following BIDS",
                                                                epilog=epilog)
```
