### 2024-03-05 272dc58
update error messages (#282)
* update error messages
* Update pypet2bids/pypet2bids/dcm2niix4pet.py
* Use dcm2niix specified in pet2bidsconfig file instead of dcm2niix found on path
