### 2024-03-05 2921d2d
update docstrings and cli descriptions
