+++ pypet2bids/pypet2bids/convert_pmod_to_blood.py
@@ -117,7 +117,9 @@ def cli():
     :rtype: argparse.namespace
 
     """
-    parser = argparse.ArgumentParser(formatter_class=argparse.RawDescriptionHelpFormatter)
+    parser = argparse.ArgumentParser(formatter_class=argparse.RawDescriptionHelpFormatter,
+                                     description="Converts PMOD blood files to PET BIDS compliant tsv's and json "
+                                                 "_blood.* files")
     parser.add_argument(
         "--whole-blood-path",
         '-w',
