### 2024-03-05 openneuropet/PET2BIDS#276 @bendhouseart
Alright, added the update pet stuff to the read the docs site and descriptions for all of the cli's. I think I just need to touch up the readme slightly and make the import's happy with flake*

> I also stumbled on undefined variable in dicom_convert.py, line 240. zfill used here alone, and few lines above is used as nifti_run_number.zfill.

I think I need to delete this file as it was the original converter and hasn't been used forever.
