### 2024-03-05 e9373be
add formatting to makefile, fixed flake8 F405
Fixed flake8 complaint's about * imports, added a make black command
to stop flake yelling so much in the future (will probably ignore the 79
char limit though).
