### 2024-03-05 40ae2ae
276 joss review documentation review (#284)
* added symlink to README.md updated pypi links
* added links to documention and repository on pypi page
* added symlink from README.md to pypet2bids/README.md for rendering
  on pypi project page.
* bump version
* update docstrings and cli descriptions
* add formatting to makefile, fixed flake8 F405

Fixed flake8 complaint's about * imports, added a make black command
to stop flake yelling so much in the future (will probably ignore the 79
char limit though).
