### 2024-03-06 #6067 @nbeliy
@britta-wstnr Authors addressed my [issues](https://github.com/openneuropet/PET2BIDS/issues/276) on code and documentation. So for me paper is good to go.
