### 2024-03-12 9dfc2b0
These these modifications are helping?
It seems like as I more closely approach the matlab code in read_ecat
and ecat2nii the image quality of the python images continues to
decrease.
