+++ pypet2bids/pypet2bids/ecat2nii.py
@@ -5,6 +5,7 @@ Contains class ecat2ii, used to write ecat information out to a nifti file, depe
 | *Authors: Anthony Galassi*
 | *Copyright OpenNeuroPET team*
 """
+
 import datetime
 import nibabel
 import numpy
@@ -20,23 +21,26 @@ except ModuleNotFoundError:
     import pypet2bids.helper_functions as helper_functions
 
 
-#logger = helper_functions.logger('pypet2bids')
+# logger = helper_functions.logger('pypet2bids')
 
-logger = logging.getLogger('pypet2bids')
+logger = logging.getLogger("pypet2bids")
 
-def ecat2nii(ecat_main_header=None,
-             ecat_subheaders=None,
-             ecat_pixel_data=None,
-             ecat_file=None,
-             nifti_file: str = '',
-             sif_out=False,
-             affine=None,
-             save_binary=False,
-             **kwargs):
+
+def ecat2nii(
+    ecat_main_header=None,
+    ecat_subheaders=None,
+    ecat_pixel_data=None,
+    ecat_file=None,
+    nifti_file: str = "",
+    sif_out=False,
+    affine=None,
+    save_binary=False,
+    **kwargs,
+):
     """
     Converts an ECAT file into a nifti and a sidecar json, used in conjunction with
     :func:`pypet2bids.read_ecat.read_ecat`
-    
+
     :param ecat_main_header: the main header of an ECAT file
     :param ecat_subheaders: the subheaders for each frame of the ECAT file
     :param ecat_pixel_data: the imaging/pixel data from the ECAT file
@@ -65,32 +69,51 @@ def ecat2nii(ecat_main_header=None,
     nifti_file_w_out_extension = os.path.splitext(str(pathlib.Path(nifti_file).name))[0]
 
     # if already read nifti file skip re-reading
-    if ecat_main_header is None and ecat_subheaders is None and ecat_pixel_data is None and ecat_file:
+    if (
+        ecat_main_header is None
+        and ecat_subheaders is None
+        and ecat_pixel_data is None
+        and ecat_file
+    ):
         # collect ecat_file
         main_header, sub_headers, data = read_ecat(ecat_file=ecat_file)
-    elif ecat_file is None and type(ecat_main_header) is dict and type(ecat_subheaders) is list and type(
-            ecat_pixel_data) is numpy.ndarray:
-        main_header, sub_headers, data = ecat_main_header, ecat_subheaders, ecat_pixel_data
+    elif (
+        ecat_file is None
+        and type(ecat_main_header) is dict
+        and type(ecat_subheaders) is list
+        and type(ecat_pixel_data) is numpy.ndarray
+    ):
+        main_header, sub_headers, data = (
+            ecat_main_header,
+            ecat_subheaders,
+            ecat_pixel_data,
+        )
     else:
-        raise Exception("Must pass in filepath for ECAT file or "
-                        "(ecat_main_header, ecat_subheaders, and ecat_pixel data "
-                        f"got ecat_file={ecat_file}, type(ecat_main_header)={type(ecat_main_header)}, "
-                        f"type(ecat_subheaders)={type(ecat_subheaders)}, "
-                        f"type(ecat_pixel_data)={type(ecat_pixel_data)} instead.")
+        raise Exception(
+            "Must pass in filepath for ECAT file or "
+            "(ecat_main_header, ecat_subheaders, and ecat_pixel data "
[…]
