+++ pypet2bids/pypet2bids/update_json_pet_file.py
@@ -21,35 +21,52 @@ except ModuleNotFoundError:
 logger = helper_functions.logger("pypet2bids")
 
 # load module and metadata_json paths from helper_functions
-module_folder, metadata_folder = helper_functions.module_folder, helper_functions.metadata_folder
+module_folder, metadata_folder = (
+    helper_functions.module_folder,
+    helper_functions.metadata_folder,
+)
 
 try:
     # collect metadata jsons in dev mode
-    metadata_jsons = \
-        [Path(join(metadata_folder, metadata_json)) for metadata_json
-         in listdir(metadata_folder) if '.json' in metadata_json]
+    metadata_jsons = [
+        Path(join(metadata_folder, metadata_json))
+        for metadata_json in listdir(metadata_folder)
+        if ".json" in metadata_json
+    ]
 except FileNotFoundError:
-    metadata_jsons = \
-        [Path(join(module_folder, 'metadata', metadata_json)) for metadata_json
-         in listdir(join(module_folder, 'metadata')) if '.json' in metadata_json]
+    metadata_jsons = [
+        Path(join(module_folder, "metadata", metadata_json))
+        for metadata_json in listdir(join(module_folder, "metadata"))
+        if ".json" in metadata_json
+    ]
 
 # create a dictionary to house the PET metadata files
 metadata_dictionaries = {}
 
 for metadata_json in metadata_jsons:
     try:
-        with open(metadata_json, 'r') as infile:
+        with open(metadata_json, "r") as infile:
             dictionary = json.load(infile)
 
         metadata_dictionaries[metadata_json.name] = dictionary
     except FileNotFoundError as err:
-        raise Exception(f"Missing pet metadata file {metadata_json} in {metadata_folder}, unable to validate metadata.")
+        raise Exception(
+            f"Missing pet metadata file {metadata_json} in {metadata_folder}, unable to validate metadata."
+        )
     except json.decoder.JSONDecodeError as err:
         raise IOError(f"Unable to read from {metadata_json}")
 
 
-def check_json(path_to_json, items_to_check=None, silent=False, spreadsheet_metadata={}, mandatory=True,
-               recommended=True, logger_name='pypet2bids', **additional_arguments):
+def check_json(
+    path_to_json,
+    items_to_check=None,
+    silent=False,
+    spreadsheet_metadata={},
+    mandatory=True,
+    recommended=True,
+    logger_name="pypet2bids",
+    **additional_arguments,
+):
     """
     This method opens a json and checks to see if a set of mandatory values is present within that json, optionally it
     also checks for recommended key value pairs. If fields are not present a warning is raised to the user.
@@ -78,67 +95,85 @@ def check_json(path_to_json, items_to_check=None, silent=False, spreadsheet_meta
 
     # check for default argument for dictionary of items to check
     if items_to_check is None:
-        items_to_check = metadata_dictionaries['PET_metadata.json']
+        items_to_check = metadata_dictionaries["PET_metadata.json"]
         # remove blood tsv data from items to check
-        if items_to_check.get('blood_recording_fields', None):
-            items_to_check.pop('blood_recording_fields')
+        if items_to_check.get("blood_recording_fields", None):
+            items_to_check.pop("blood_recording_fields")
 
     # open the json
-    with open(path_to_json, 'r') as in_file:
+    with open(path_to_json, "r") as in_file:
         json_to_check = json.load(in_file)
 
     # initialize warning colors and warning storage dictionary
     storage = {}
     flattened_spreadsheet_metadata = {}
-    flattened_spreadsheet_metadata.update(spreadsheet_metadata.get('nifti_json', {}))
[…]
