### 2024-03-19 a0d95ac
Update is_pet.py and reformat w/ black (#287)
Sped up is_pet.py when running on folders by providing the optional argument --skim.
Using --skim will instruct is_pet.py to only examine one of each file with a unique
extension contained within a folder/subfolder.

Reformatted code with black and updated `make black` to only examine `pypet2bids`.

Bumped version to 1.3.9 to denote minor fix.
