+++ pypet2bids/pypet2bids/ecat_cli.py
@@ -4,6 +4,7 @@ Simple command line tool to extract header and pixel information from ecat files
 | *Authors: Anthony Galassi*
 | *Copyright OpenNeuroPET team*
 """
+
 import argparse
 import os
 import pathlib
@@ -20,7 +21,8 @@ except ModuleNotFoundError:
     from pypet2bids.ecat import Ecat
     from pypet2bids.update_json_pet_file import check_json, check_meta_radio_inputs
 
-epilog = textwrap.dedent('''
+epilog = textwrap.dedent(
+    """
     
     example usage:
     
@@ -31,7 +33,8 @@ epilog = textwrap.dedent('''
 params
     
     For additional (highly verbose) example usage call this program with the --show-examples flag.
-''')
+"""
+)
 
 
 def cli():
@@ -67,58 +70,125 @@ def cli():
 
     :return: argparse.ArgumentParser.args for later use in executing conversions or ECAT methods
     """
-    parser = argparse.ArgumentParser(formatter_class=argparse.RawDescriptionHelpFormatter, epilog=epilog,
-                                     description="Extracts information from an ECAT file and/or converts ECAT imaging"
-                                                 " and PET metadata files to BIDS compliant nifti, json, and tsv")
+    parser = argparse.ArgumentParser(
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+        epilog=epilog,
+        description="Extracts information from an ECAT file and/or converts ECAT imaging"
+        " and PET metadata files to BIDS compliant nifti, json, and tsv",
+    )
     update_or_convert = parser.add_mutually_exclusive_group()
-    parser.add_argument("ecat", nargs='?', metavar="ecat_file", help="Ecat image to collect info from.")
-    parser.add_argument("--affine", "-a", help="Show affine matrix", action="store_true", default=False)
-    update_or_convert.add_argument("--convert", "-c", required=False, action='store_true',
-                                   help="If supplied will attempt conversion.")
-    parser.add_argument("--dump", "-d", help="Dump information in Header", action="store_true", default=False)
-    parser.add_argument("--json", "-j", action="store_true", default=False, help="""
-        Output header and subheader info as JSON to stdout, overrides all other options""")
-    parser.add_argument("--nifti", "-n", metavar="file_name", help="Name of nifti output file", required=False,
-                        default=None)
-    parser.add_argument("--subheader", '-s', help="Display subheaders", action="store_true", default=False)
-    parser.add_argument("--sidecar", action="store_true", help="Output a bids formatted sidecar for pairing with"
-                                                               "a nifti.")
-    parser.add_argument('--kwargs', '-k', nargs='*', action=helper_functions.ParseKwargs, default={},
-                        help="Include additional values in the nifti sidecar json or override values extracted from "
-                             "the supplied nifti. e.g. including `--kwargs TimeZero=\"12:12:12\"` would override the "
-                             "calculated TimeZero. Any number of additional arguments can be supplied after --kwargs "
-                             "e.g. `--kwargs BidsVariable1=1 BidsVariable2=2` etc etc."
-                             "Note: the value portion of the argument (right side of the equal's sign) should always"
-                             "be surrounded by double quotes BidsVarQuoted=\"[0, 1 , 3]\"")
-    parser.add_argument('--scannerparams', nargs='*',
-                        help="Loads saved scanner params from a configuration file following "
-                             "--scanner-params/-s if this option is used without an argument "
-                             "this cli will look for any scanner parameters file in the "
-                             "directory with the name *parameters.txt from which this cli is "
-                             "called.")
[…]
