+++ pypet2bids/pypet2bids/single_spreadsheet.py
@@ -10,12 +10,13 @@ try:
 except ModuleNotFoundError:
     import pypet2bids.helper_functions as helper_functions
 
-#from pypet2bids.helper_functions import single_spreadsheet_reader, \
+# from pypet2bids.helper_functions import single_spreadsheet_reader, \
 #    collect_bids_part, open_meta_data, load_pet_bids_requirements_json, ParseKwargs
 
+
 def read_single_subject_spreadsheets(
-        general_metadata_spreadsheet: pathlib.Path,
-        **kwargs) -> dict:
+    general_metadata_spreadsheet: pathlib.Path, **kwargs
+) -> dict:
     """
     Reads in spreadsheet as formatted in PET2BIDS/spreadsheet_conversion/single_subject_sheet,
     extra arguments are supplied in key pair form via kwargs as is convention.
@@ -30,7 +31,7 @@ def read_single_subject_spreadsheets(
     :return: dictionary of subject data extracted from each spreadsheet along with any additional
     kwargs supplied
     :rtype: dict
-    
+
     Anthony Galassi
     -----------------------------
     Copyright Open NeuroPET team
@@ -38,64 +39,87 @@ def read_single_subject_spreadsheets(
 
     required_fields = helper_functions.load_pet_bids_requirements_json()
 
-    subject_id = kwargs.get('subject_id', None)
-    session_id = kwargs.get('session_id', None)
+    subject_id = kwargs.get("subject_id", None)
+    session_id = kwargs.get("session_id", None)
 
     if general_metadata_spreadsheet.is_file():
-        general_metadata = helper_functions.single_spreadsheet_reader(general_metadata_spreadsheet)
+        general_metadata = helper_functions.single_spreadsheet_reader(
+            general_metadata_spreadsheet
+        )
         if not subject_id:
-            subject_id = helper_functions.collect_bids_part('sub', str(general_metadata_spreadsheet))
+            subject_id = helper_functions.collect_bids_part(
+                "sub", str(general_metadata_spreadsheet)
+            )
             if subject_id:
-                general_metadata['subject_id'] = subject_id
+                general_metadata["subject_id"] = subject_id
         else:
-            general_metadata['subject_id'] = subject_id
+            general_metadata["subject_id"] = subject_id
 
         if not session_id:
-            session_id = helper_functions.collect_bids_part('ses', str(general_metadata_spreadsheet))
+            session_id = helper_functions.collect_bids_part(
+                "ses", str(general_metadata_spreadsheet)
+            )
             if session_id:
-                general_metadata['session_id'] = session_id
+                general_metadata["session_id"] = session_id
         else:
-            general_metadata['session_id'] = session_id
+            general_metadata["session_id"] = session_id
 
         column_set = list(general_metadata.keys())
         # assert required columns are in input spreadsheets
         for k in required_fields.keys():
             for field in required_fields[k]:
                 field_exists = field in column_set
-                if k == 'mandatory' and not field_exists:
-                    logging.warning(f"Input spreadsheet {general_metadata_spreadsheet} is missing required "
-                                    f"column {field}")
-                elif k == 'recommended' and not field_exists:
-                    logging.info(f"Input spreadsheet(s) {general_metadata_spreadsheet} and is missing "
-                                 f"recommended column {field}")
-                elif k == 'optional':
-                    logging.info(f"Input spreadsheet(s) {general_metadata_spreadsheet} is missing "
-                                 f"optional column {field}")
+                if k == "mandatory" and not field_exists:
+                    logging.warning(
+                        f"Input spreadsheet {general_metadata_spreadsheet} is missing required "
+                        f"column {field}"
+                    )
+                elif k == "recommended" and not field_exists:
[…]

[…]
