+++ ecat_testing/ints/README.md
@@ -1,0 +1,62 @@
+# matlab_numpy_ints
+
+This is a very simple example of testing matlab vs python reading and writing of ecat files scaling up to a more
+complex set of testing that reads real ecat data and writes it to niftis.
+
+## Dead Simple Testing / Proofing 
+
+Then simplest test cases that use "imaging" data made of small 2 x 1 arrays of integer data can found in:
+
+`ints.m` and `ints.py`
+
+## Outline of Testing on Real Data (e.g. steps required to read and write ecat into Nifti)
+
+A part of this debugging process is to capture the state of inputs and outputs through each iteration of the python
+and matlab code. These ouputs are cataloged and saved in this directory (`ecat_testing/ints/`) within subdirectory 
+`steps/`. The naming, order, and description of each step are laid out in the table below for easy reference. At each
+of these steps the outputs are saved in the `steps/` directory with the filestem (and the python or matlab code used to 
+make them) as described. These temporary files are written if and only if the environment variable `ECAT_SAVE_STEPS=1` 
+is set.
+
+| Step | Description                     | Matlab        | Python         | FileStem                      |
+|------|---------------------------------|---------------|----------------|-------------------------------|
+| 1    | Read ecat file                  | `read_ecat.m` | `read_ecat.py` | `1_read_ecat*`                |
+| 2    | Read main header                | `read_ecat.m` | `read_ecat.py` | `2_read_mh_ecat*`             |
+| 3    | Read subheaders                 | `read_ecat.m` | `read_ecat.py` | `3_read_sh_ecat*`             |
+| 4    | Determine file/data type        | `read_ecat.m` | `read_ecat.py` | `4_determine_data_type*`      |
+| 5    | Read image data                 | `read_ecat.m` | `read_ecat.py` | `5_read_img_ecat*`            |
+| 6    | scale if calibrated             | `read_ecat.m` | `read_ecat.py` | `6_scale_img_ecat*`           |
+| 7    | Pass Data to ecat2nii           | `ecat2nii.m`  | `ecat2nii.py`  | `7_ecat2nii*`                 |
+| 8    | Flip ECAT data into Nifti space | `ecat2nii.m`  | `ecat2nii.py`  | `8_flip_ecat2nii*`            |
+| 9    | Rescale to 16 bits              | `ecat2nii.m`  | `ecat2nii.py`  | `9_rescale_to_16_ecat2nii*`   |
+| 10   | Calibration Units Scaling       | `ecat2nii.m`  | `ecat2nii.py`  | `10_scal_cal_units_ecat2nii*` |
+| 11   | Save to Nifti                   | `ecat2nii.m`  | `ecat2nii.py`  | `11_save_nii_ecat2nii*`       |
+| 12   | Additional Steps                | TBD           | TBD            | TBD                           |
+
+1. Read ecat file: opening the ecat file, outputs are difficult to compare at this point but generally take the form of
+   mainheader, subheaders, and image data, this step might be useless
+2. Read main header: this is the first step in reading the ecat file, the main header contains information about the
+   file, the subheaders, and the image data. These will be saved as jsons for comparison.
+3. Read subheaders: the subheaders contain information about the image data, these will be saved as jsons for comparison
+   as well.
+4. Determine file/data type: this step is to determine the type of data in the image data, this will be saved a json or 
+   a text file with a single line specifying the datatype and endendianess, but probably a json.
+5. Read image data: this is the final step in reading the ecat file, the image data is read in and will be examined at 
+   3 different time points (if available). E.g. the first frame, the middle frame, and the final frame. Only a single 2D
+   slice will be saved from each of the time points, and it too will be taken from the "middle" of its 3D volume. We're
+   only attempting to compare whether python and matlab have done a decent job of reading the data as recorded in.
[…]
