### 2024-03-27 b0fd706
remove redundant step 1
