+++ ecat_testing/ints/README.md
@@ -12,51 +12,49 @@ Then simplest test cases that use "imaging" data made of small 2 x 1 arrays of i
 ## Outline of Testing on Real Data (e.g. steps required to read and write ecat into Nifti)
 
 A part of this debugging process is to capture the state of inputs and outputs through each iteration of the python
-and matlab code. These ouputs are cataloged and saved in this directory (`ecat_testing/ints/`) within subdirectory 
+and matlab code. These outputs are cataloged and saved in this directory (`ecat_testing/ints/`) within subdirectory 
 `steps/`. The naming, order, and description of each step are laid out in the table below for easy reference. At each
 of these steps the outputs are saved in the `steps/` directory with the filestem (and the python or matlab code used to 
 make them) as described. These temporary files are written if and only if the environment variable `ECAT_SAVE_STEPS=1` 
 is set.
 
-| Step | Description                     | Matlab        | Python         | FileStem                      |
-|------|---------------------------------|---------------|----------------|-------------------------------|
-| 1    | Read ecat file                  | `read_ecat.m` | `read_ecat.py` | `1_read_ecat*`                |
-| 2    | Read main header                | `read_ecat.m` | `read_ecat.py` | `2_read_mh_ecat*`             |
-| 3    | Read subheaders                 | `read_ecat.m` | `read_ecat.py` | `3_read_sh_ecat*`             |
-| 4    | Determine file/data type        | `read_ecat.m` | `read_ecat.py` | `4_determine_data_type*`      |
-| 5    | Read image data                 | `read_ecat.m` | `read_ecat.py` | `5_read_img_ecat*`            |
-| 6    | scale if calibrated             | `read_ecat.m` | `read_ecat.py` | `6_scale_img_ecat*`           |
-| 7    | Pass Data to ecat2nii           | `ecat2nii.m`  | `ecat2nii.py`  | `7_ecat2nii*`                 |
-| 8    | Flip ECAT data into Nifti space | `ecat2nii.m`  | `ecat2nii.py`  | `8_flip_ecat2nii*`            |
-| 9    | Rescale to 16 bits              | `ecat2nii.m`  | `ecat2nii.py`  | `9_rescale_to_16_ecat2nii*`   |
-| 10   | Calibration Units Scaling       | `ecat2nii.m`  | `ecat2nii.py`  | `10_scal_cal_units_ecat2nii*` |
-| 11   | Save to Nifti                   | `ecat2nii.m`  | `ecat2nii.py`  | `11_save_nii_ecat2nii*`       |
-| 12   | Additional Steps                | TBD           | TBD            | TBD                           |
+| Step | Description                     | Matlab        | Python         | FileStem                     |
+|------|---------------------------------|---------------|----------------|------------------------------|
+| 1    | Read main header                | `read_ecat.m` | `read_ecat.py` | `1_read_mh_ecat*`            |
+| 2    | Read subheaders                 | `read_ecat.m` | `read_ecat.py` | `2_read_sh_ecat*`            |
+| 3    | Determine file/data type        | `read_ecat.m` | `read_ecat.py` | `3_determine_data_type*`     |
+| 4    | Read image data                 | `read_ecat.m` | `read_ecat.py` | `4_read_img_ecat*`           |
+| 5    | scale if calibrated             | `read_ecat.m` | `read_ecat.py` | `5_scale_img_ecat*`          |
+| 6    | Pass Data to ecat2nii           | `ecat2nii.m`  | `ecat2nii.py`  | `6_ecat2nii*`                |
[…]
