### 2024-03-30 #6067 @pjtoussaint
> > except @pjtoussaint mac issue, everything else is in order i think -- but this is likely specific to a version of mac, so as much as we will try to resolve it once we know more as requested, I do think that it would be fair to move forward if no other issues.
>
> Apologies, @CPernet et al: I had to tend to other pressing matters in recent weeks. I am resuming work on this now.

Hello again: it turns out that there was a clash between bash and zsh generated at the time of migration. All is good!
