+++ ecat_testing/ints/README.md
@@ -47,13 +47,13 @@ is set.
 6. Save objects for comparison (as best as one can) before they are passed to the ecat2nii function. This will include
    the mainheader, subheaders, and image data.
 7. Return the transformed data to the nifti space from ecat. This follows the 3 flip dimension steps performed across 
-   the 3D image data. This output will use the same frames as step 5 and 6.
+   the 3D image data. This output will use the same frames as step 4 and 5.
 8. Rescale the data to 16 bits: this should only occur if the data is 12bit as is sometimes the case with ecat data. As 
    a note to self, attempting these steps in Python will start to lead to wildly different values when compared to 
    matlab. It's most likely not necessary to do this step as the data is handled in numpy, but this writer won't promise
    to eat his hat if it turns out to be necessary.
 9. Calibration Units: (Search for 'calibration_units == 1' to locate this in code) Here we can potentially alter the 
-   data again by scaling, rounding, and converting from int to float. As in steps 5 and 6 we will save the first, 
+   data again by scaling, rounding, and converting from int to float. As in steps 4 and 5 we will save the first, 
    middle, and last frames of the data as 2D slices in the middle of their respective 3D volumes for comparison.
 10. Save to nifti: the data is saved to a nifti file and the output is saved for comparison as .nii files named in line
    with the FileStem column above.
