+++ pypet2bids/pypet2bids/dcm2niix4pet.py
@@ -20,7 +20,7 @@ from platform import system
 import subprocess
 import pandas as pd
 from os.path import join
-from os import listdir, walk
+from os import listdir, walk, environ
 from pathlib import Path
 import json
 import pydicom
@@ -85,6 +85,9 @@ else:
     helper_functions.modify_config_file(
         "DEFAULT_METADATA_JSON", module_folder / "template_json.json"
     )
+    # set the default metadata json to the template json included in this library in our environment so
+    # we don't trip up retrival of the default metadata json later
+    environ["DEFAULT_METADATA_JSON"] = str(module_folder / "template_json.json")
 
 
 def dicom_datetime_to_dcm2niix_time(dicom=None, date="", time=""):
@@ -159,6 +162,7 @@ class Dcm2niix4PET:
         file_format="%p_%i_%t_%s",
         silent=False,
         tempdir_location=None,
+        default_metadata_json=default_metadata_json,
     ):
         """
         This class is a simple wrapper for dcm2niix and contains methods to do the following in order:
