+++ pypet2bids/pypet2bids/dcm2niix4pet.py
@@ -1062,6 +1062,10 @@ def cli():
         default="",
         help="Provide a reconstruction method to be used in the output file name",
     )
+    parser.add_argument(
+        "--version", '-v', action='version',
+        version=f"{helper_functions.get_version()}",
+    )
 
     return parser
 
