### 2024-04-03 ae362cf
Changed pypet2bids to check env for vars (#292)
* Changed pypet2bids to check env for vars

Added logic so that if a .pet2bidsconfig file isn't found (or
able to be written) check_pet2bids_config will return whatever value
it is able to locate from environment variables. This might just be a
new footgun, but it's got the virtue of being new.

* remove unused var

* add version to help

* loosen python req's

* update lock file
