### 2024-04-24 #6067 @pjtoussaint
> @pjtoussaint is everything ok for you to finalize your checklist? that's needed for the bot to run I believe tag @britta-wstnr

Hello @CPernet : I just checked the last box!! Functionality was tested and working as described (my github was muted, as @britta-wstnr guessed - apologies).
