### 2024-04-24 2c2d89e
need to make a new output function for first_middle_last_frames_to_text for arrays since they're different than cell arrays.
