+++ matlab/ecat2nii.m
@@ -48,6 +48,16 @@ sifout  = false; % 0/1 to indicate if sif file are also created
 gz      = true;  % compress nifti
 savemat = false; % save ecat data as .mat
 
+%% debebugging
+% ------------
+ecat_save_steps = getenv('ECAT_SAVE_STEPS');
+ecat_save_steps_dir = mfilename('fullpath');
+if contains(ecat_save_steps_dir, 'LiveEditorEvaluationHelper')
+    ecat_save_steps_dir = matlab.desktop.editor.getActiveFilename;
+end
+parts = strsplit(ecat_save_steps_dir, filesep);
+ecat_save_steps_dir = strjoin([parts(1:end-2), 'ecat_testing', 'steps'], filesep);
+
 %% check inputs
 % ------------
 
@@ -122,7 +132,7 @@ end
 % --------------------
 for j=1:length(FileListIn)
     
-    try
+    %try
         fprintf('Conversion of file: %s\n',FileListIn{j});
         
         % quickly ensure we have the TimeZero - key to all analyzes!
@@ -143,7 +153,7 @@ for j=1:length(FileListIn)
         end
         
         pet_file = [pet_file ext];
-        [mh,sh]  = readECAT7([pet_path filesep pet_file]); % loading the whole file here and iterating to flipdim below only minuimally improves time (0.6sec on NRU server)
+        [mh,sh, data]  = readECAT7([pet_path filesep pet_file]); % loading the whole file here and iterating to flipdim below only minuimally improves time (0.6sec on NRU server)
         if sh{1}.data_type ~= 6
             error('Conversion for 16 bit data only (type 6 in ecat file) - error loading ecat file');
         end
@@ -166,7 +176,13 @@ for j=1:length(FileListIn)
                 Randoms(i)    = 0;
             end
         end
-        
+
+        % save debugging steps 6 and 7
+        if (ecat_save_steps == '1')
+            first_middle_last_frames_to_text(data,ecat_save_steps_dir, '6_ecat2nii_matlab');
+            first_middle_last_frames_to_text(img_temp,ecat_save_steps_dir, '7_flip_ecat2nii_matlab');
+        end
+
         % rescale to 16 bits
         rg = max(img_temp(:))-min(img_temp(:));
         if rg ~= 32767 % same as range(img_temp(:)) but no need of stats toolbox
@@ -179,7 +195,12 @@ for j=1:length(FileListIn)
                 Sca      = Sca*MinImg/(-32768);
             end
         end
-        
+
+        % save debugging step 8 - rescale to 16 bits
+        if (ecat_save_steps == '1')
+            first_middle_last_frames_to_text(img_temp,ecat_save_steps_dir, '8_rescale_to_16_ecat2nii_matlab');
+        end
+
         % check names
         if ~exist('FileListOut','var')
             [~,pet_filename]           = fileparts(pet_file);
@@ -215,7 +236,7 @@ for j=1:length(FileListIn)
         end
         
         % save raw data
-        if savemat
+        if savemat or ecat_save_steps == '1'
             if mh.calibration_units == 1 % see line 337
                 ecat = img_temp.*Sca;
             else
@@ -336,6 +357,9 @@ for j=1:length(FileListIn)
         
         if mh.calibration_units == 1 % do calibrate
             img_temp                          = single(round(img_temp).*(Sca*mh.ecat_calibration_factor)); % scale and dose calibrated
+            if save_ecat_steps == '1'
+                first_middle_last_frames_to_text(img_temp,ecat_save_steps_dir, '9_scal_cal_units_ecat2nii_matlab');
+            end
             warning('it looks like the source data are not dose calibrated - ecat2nii is thus scaling the data')
[…]
