### 2024-04-24 b8b21fd
things diverge after step 7
