### 2024-04-25 #6067 @britta-wstnr
A big thanks already to @nbeliy, @adswa, and @pjtoussaint for their reviewing work! JOSS and I appreciate the time and work you put into this! 💙

@CPernet, the next step is for me to read the paper again. I might get back to you with some things (like typos) to fix.
After that, I will have some more things to do for you, so keep an eye out for notifications 🙂
In the meanwhile, can you please double check the DOIs that the editorialbot reported as missing or invalid [here](https://github.com/openjournals/joss-reviews/issues/6067#issuecomment-2076734028)? 🙏
