### 2024-04-26 #6067 @britta-wstnr
Hi @CPernet,

I looked over the paper and there is a couple of things I noted:
- line 27: "e.g., timing information such as ~e.g.~ 'time zero' ..."  ⬅️  I think the second e.g. is not needed, since you already say "such as"
- line 36: I think it's BIDSCoin, no "s"
- line 43: The reference `B. T. Christian (1998)` is not formatted as a proper reference and does not show up in the bibliography
- line 44: The reference `R.F. Muzic (2002)` is not formatted as a proper reference and does not show up in the bibliography
- line 52: "... PET scanner data which is why ..." ⬅️ this is a long sentence, you can consider placing a comma after "why" to make it easier to parse
- generally: you could consider putting all function names (e.g., `ecat2nii`) and file names (e.g., `update_json_pet_file.json`) in monospaced font to make the text easier to parse
- as mentioned above, could you please check the DOIs that the editorialbot as missing or invalid [here](https://github.com/openjournals/joss-reviews/issues/6067#issuecomment-2076734028)?

Thanks a lot!
