### 2024-04-26 09b5c34
fix type in ecat.py
