### 2024-04-26 06113ae
dcm2niix isn't on path?
