### 2024-04-26 3651260
bug fix with ignoring 0 values
