### 2024-04-26 9fd7779
fix logic and not or
