### 2024-04-26 74db9f1
Fixes for issues 293 and 288 (#295)
Resolves issues #293 and #288
* fix typo in ecat.py
* moving times to be relative to timezero
* fixed error message about dcm2niix not being on path
