### 2024-04-27 d9831b3
step 8 now matches between python and matlab
