### 2024-04-29 openneuropet/PET2BIDS#301 @bendhouseart
Grammar/Formatting/Tweaks
Per @britta-wstnr's comment [here](https://github.com/openjournals/joss-reviews/issues/6067#issuecomment-2079182471)

- [x] line 27: "e.g., timing information such as e.g. 'time zero' ..." ⬅️ I think the second e.g. is not needed, since you already say "such as"
- [x] line 36: I think it's BIDSCoin, no "s"
- [x]  line 43: The reference B. T. Christian (1998) is not formatted as a proper reference and does not show up in the bibliography
- [x] line 44: The reference R.F. Muzic (2002) is not formatted as a proper reference and does not show up in the bibliography
- [x] line 52: "... PET scanner data which is why ..." ⬅️ this is a long sentence, you can consider placing a comma after "why" to make it easier to parse
- [x] generally: you could consider putting all function names (e.g., ecat2nii) and file names (e.g., update_json_pet_file.json) in monospaced font to make the text easier to parse
 - [x]  as mentioned above, could you please check the DOIs that the editorialbot as missing or invalid https://github.com/openjournals/joss-reviews/issues/6067#issuecomment-2076734028?
