### 2024-04-29 #6067 @bendhouseart
@CPernet is pretty busy this week, so I'll be doing my best to take over. I've broken down the Post-Review Checklist below

> - Double check authors and affiliations (including ORCIDs)
>  - Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
>  - Archive the release on Zenodo/figshare/etc and post the DOI here.
>  - Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
>   - Make sure that the license listed for the archive is the same as the software license.

into a series of issues [here](https://github.com/openneuropet/PET2BIDS/labels/JOSS). As well as your comments @britta-wstnr into this issue -> [#301](https://github.com/openneuropet/PET2BIDS/issues/301#issue-2269502829).

Will tag/update you all as I work on it this week.
