### 2024-04-29 7b41bb6
it's the nifti's that are the issue..
