+++ pypet2bids/pypet2bids/ecat_cli.py
@@ -185,7 +185,9 @@ def cli():
         "path/to/metadata.xlsx",
     )
     parser.add_argument(
-        "--version", '-v', action='version',
+        "--version",
+        "-v",
+        action="version",
         version=f"{helper_functions.get_version()}",
     )
 
