### 2024-05-01 3e32edd
update paper.md and paper.bib where able
