+++ paper.bib
@@ -13,8 +13,8 @@
 @article{Hayashi_2023,
 	title = {brainlife.io: A decentralized and open source cloud platform to support 
       neuroscience research.},
-	url = {https://arxiv.org/abs/2306.02183v3},
-	doi = {arXiv:2306.02183v3},
+	url = {https://arxiv.org/abs/2306.02183},
+	doi = {arXiv:2306.02183},
 	journaltitle = {{ArXiv [Preprint]}},
 	shortjournal = {{ArXiv}},
 	author = {Hayashi S, Caron BA, Heinsfeld AS, Vinci-Booher S, McPherson B, Bullock DN, Bertò G, Niso G, Hanekamp S, Levitas D, Ray K, MacKenzie A, Kitchell L, Leong JK, Nascimento-Silva F, Koudoro S, Willis H, Jolly JK, Pisner D, Zuidema TR, Kurzawski JW, Mikellidou K, Bussalb A, Rorden C, Victory C, Bhatia D, Baran Aydogan D, Yeh FF, Delogu F, Guaje J, Veraart J, Bollman S, Stewart A, Fischer J, Faskowitz J, Chaumon M, Fabrega R, Hunt D, McKee S, Brown ST, Heyman S, Iacovella V, Mejia AF, Marinazzo D, Craddock RC, Olivetti E, Hanson JL, Avesani P, Garyfallidis E, Stanzione D, Carson J, Henschel R, Hancock DY, Stewart CA, Schnyer D, Eke DO, Poldrack RA, George N, Bridge H, Sani I, Freiwald WA, Puce A, Port NL, Pestilli F},
@@ -124,32 +124,35 @@ Publisher: Nature Publishing Group},
 }
 
 @article{Burger1997,
-author = {Burger, Cyril and Buck, Alfred},
-issn = {01615505},
-journal = {Journal of Nuclear Medicine},
-keywords = {Coupled fitting,Modeling software,Tracer kinetic modeling},
-number = {11},
-pages = {1818--1823},
-pmid = {9374364},
-title = {{Requirements and implementation of a flexible kinetic modeling tool}},
-volume = {38},
-year = {1997}
+	author = {Burger, Cyril and Buck, Alfred},
+	issn = {01615505},
+	journal = {Journal of Nuclear Medicine},
+	keywords = {Coupled fitting,Modeling software,Tracer kinetic modeling},
+	number = {11},
+	pages = {1818--1823},
+	pmid = {9374364},
+	title = {{Requirements and implementation of a flexible kinetic modeling tool}},
+	volume = {38},
+	year = {1997}
 }
 
 @article{ezBIDS2024,
-author = {Levitas, Daniel and Hayashi, Soichi and Vinci-Booher, Sophia and Heinsfeld, Anibal, and Bhatia, Dheeraj and Lee, Nicholas and Galassi, Anthony and Niso, Guiomar and Pestilli, Franco},
-title = {{ezBIDS: Guided standardization of neuroimaging data interoperable with major data archives and platforms.}},
-doi = {10.1038/s41597-024-02959-0},
-journal = {Scientifc Data},
-volume = {11},
-year = {2024}
+	author = {Levitas, Daniel and Hayashi, Soichi and Vinci-Booher, Sophia and Heinsfeld, Anibal, and Bhatia, Dheeraj and Lee, Nicholas and Galassi, Anthony and Niso, Guiomar and Pestilli, Franco},
+	title = {{ezBIDS: Guided standardization of neuroimaging data interoperable with major data archives and platforms.}},
+	doi = {10.1038/s41597-024-02959-0},
+	journal = {Scientifc Data},
+	volume = {11},
+	year = {2024}
 } 
 
[…]
