+++ paper.md
@@ -7,61 +7,66 @@ tags:
   - positron emission tomography (PET)
   - brain imaging data structure (BIDS)
 authors:
-- name: Anthony Galassi
-  orcid: 0000-0001-6550-4574
-  affiliation: 1
-- name: Cyrus Eierud
-  orcid: 0000-0002-9942-676X
-  affiliation: 2
-- name: Martin Norgaard
-  orcid: 0000-0003-2131-5688
-  affiliation: "3, 4, 5"
-- name: Adam G. Thomas
-  orcid: 0000-0002-2850-1419
-  affiliation: 1
-- name: Gabriel Gonzalez-Escamilla
-  orcid: 0000-0002-7209-1736
-  affiliation: 6
-- name: Claus Svarer
-  orcid: 0000-0001-7811-1825
-  affiliation: 3
-- name: Chris Rorden
-  orcid: 0000-0002-7554-6142
-  affiliation: 7
-- name: Granville J. Matheson
-  orcid: 0000-0002-5646-4547
-  affiliation: "8, 9"
-- name: Gitte M. Knudsen
-  orcid: 0000-0003-1508-6866
-  affiliation: 3
-- name: Robert B. Innis
-  orcid: 0000-0003-1238-7209
-  affiliation: 1 
-- name: Melanie Ganz
-  orcid: 0000-0002-9120-8098
-  affiliation: "3, 4"
-- name: Cyril Pernet
-  orcid: 0000-0003-4010-4632
-  affiliation: 3
+  - name: Anthony Galassi
+    orcid: 0000-0001-6550-4574
+    affiliation: 1
+  - name: Cyrus Eierud
+    orcid: 0000-0002-9942-676X
+    affiliation: 2
+  - name: Martin Norgaard
+    orcid: 0000-0003-2131-5688
+    affiliation: "3, 4, 5"
+  - name: Adam G. Thomas
+    orcid: 0000-0002-2850-1419
+    affiliation: 1
+  - name: Gabriel Gonzalez-Escamilla
+    orcid: 0000-0002-7209-1736
+    affiliation: 6
+  - name: Claus Svarer
+    orcid: 0000-0001-7811-1825
+    affiliation: 3
+  - name: Chris Rorden
+    orcid: 0000-0002-7554-6142
+    affiliation: 7
+  - name: Granville J. Matheson
+    orcid: 0000-0002-5646-4547
+    affiliation: "8, 9"
+  - name: Gitte M. Knudsen
+    orcid: 0000-0003-1508-6866
+    affiliation: 3
+  - name: Robert B. Innis
+    orcid: 0000-0003-1238-7209
+    affiliation: 1
+  - name: Melanie Ganz
+    orcid: 0000-0002-9120-8098
+    affiliation: "3, 4"
+  - name: Murat Bilgel
+    orcid: 0000-0001-5042-7422
+    affiliation: 10
+  - name: Cyril Pernet
+    orcid: 0000-0003-4010-4632
+    affiliation: 3
 affiliations:
- - name: National Institutes of Health, Bethesda, MD, USA
-   index: 1
- - name: TReNDS Center, Georgia State University, Atlanta, GA, USA
-   index: 2
- - name: Neurobiology Research Unit, Rigshospitalet, Copenhagen, Denmark
-   index: 3
- - name: Department of Computer Science, University of Copenhagen, Copenhagen, Denmark
-   index: 4
- - name: Department of Psychology, Stanford University, CA, USA
-   index: 5
- - name: University Medical Center of the Johannes Gutenberg University Mainz, Mainz, Germany
-   index: 6
- - name: Department of Psychology, University of South Carolina, Columbia, SC, USA
-   index: 7
- - name: Mailman school of Public Health, Columbia University, New York, NY, USA
-   index: 8
[…]
