### 2024-05-01 d7130cd
Update links in paper.md
