+++ paper.md
@@ -76,7 +76,7 @@ bibliography: paper.bib
 
 The Brain Imaging Data Structure [@gorgolewski_brain_2016] is a standard for organizing and naming neuroimaging data,
 which has quickly become successful and popular in the community with adoption by brain imaging repositories
-(e.g., OpenNeuro [@noauthor_openneuro], PublicnEUro [@noauthor_public]),
+(e.g., OpenNeuro [@openneuro], PublicnEUro [@noauthor_public]),
 data management tools (e.g., COINS [@landis_coins_2016], XNAT [@marcus_extensible_2007]) and computational platforms
 (e.g. BrainLife [@Hayashi_2023]). BIDS allows data to be shared much more efficiently, enables the development of
 automated data analysis pipelines, and improves reproducibility.
@@ -101,7 +101,7 @@ _File conversion_: The conversion for PET data stored in DICOM format to NIfTI i
 functions, which are wrappers around dcm2niix [@rorden_dcm2nii; @li_first_2016]. Those functions can extend the image
 sidecar JSON file generated by dcm2niix with user-provided metadata, making them BIDS-compliant. The conversion of PET
 data stored in ECAT format is performed using the dedicated ecat2nii functions. The Matlab code relies on the
-`readECAT7.m` [@readECAT7] function written by B.T. Christian (1998) and revised by R.F. Muzic (2002) to read the data
+`readECAT7.m` [@christian_readECAT7] function written by B.T. Christian (1998) and revised by R.F. Muzic (2002) to read the data
 while writing Nifti files relies on `nii_tool` [@Li_2016]. The Python code mirrors the Matlab code, except writing of
 Nifti files,
 which are delegated to Nibabel [@brett_nipynibabel_2023]. For both Matlab and Python, the data conversion and writing
