+++ paper.bib
@@ -13,12 +13,12 @@
 @article{Hayashi_2023,
 	title = {brainlife.io: A decentralized and open source cloud platform to support 
       neuroscience research.},
-	url = {https://arxiv.org/abs/2306.02183},
-	doi = {arXiv:2306.02183},
-	journaltitle = {{ArXiv [Preprint]}},
-	shortjournal = {{ArXiv}},
+	url = {https://doi.org/10.1038/s41592-024-02237-2},
+	doi = {10.1038/s41592-024-02237-2},
+	journaltitle = {Nature Methods},
+	shortjournal = {Nat Methods},
 	author = {Hayashi S, Caron BA, Heinsfeld AS, Vinci-Booher S, McPherson B, Bullock DN, Bertò G, Niso G, Hanekamp S, Levitas D, Ray K, MacKenzie A, Kitchell L, Leong JK, Nascimento-Silva F, Koudoro S, Willis H, Jolly JK, Pisner D, Zuidema TR, Kurzawski JW, Mikellidou K, Bussalb A, Rorden C, Victory C, Bhatia D, Baran Aydogan D, Yeh FF, Delogu F, Guaje J, Veraart J, Bollman S, Stewart A, Fischer J, Faskowitz J, Chaumon M, Fabrega R, Hunt D, McKee S, Brown ST, Heyman S, Iacovella V, Mejia AF, Marinazzo D, Craddock RC, Olivetti E, Hanson JL, Avesani P, Garyfallidis E, Stanzione D, Carson J, Henschel R, Hancock DY, Stewart CA, Schnyer D, Eke DO, Poldrack RA, George N, Bridge H, Sani I, Freiwald WA, Puce A, Port NL, Pestilli F},
-	urldate = {2024-02-06},
+	urldate = {2024-05-01},
 	date = {2023-08-11},
 	langid = {english},
 }
