### 2024-05-01 42fead9
update paper.md and paper.bib where able (#304)
* update paper.md and paper.bib where able
* Update links in paper.md
* update brainlife source
