### 2024-05-01 #6067 @bendhouseart
@britta-wstnr addressed the corrections, considered tho comma for line 52, monospaced functions, and fixed doi/invalid links where able (many of these don't have doi's or urls) see [here](https://github.com/openneuropet/PET2BIDS/pull/304).

Double checking everything else.
