+++ JOSS_metadata/persons_info.txt
@@ -21,3 +21,5 @@
 0000-0002-9120-8098  Melanie Ganz-Benjaminse  Department of Computer Science, University of Copenhagen, Copenhagen, Denmark
 
 0000-0003-4010-4632  Cyril Pernet  Neurobiology Research Unit, Rigshospitalet, Copenhagen, Denmark
+
+0000-0001-5042-7422  Murat Bilgel  National Institute on Aging Intramural Research Program, Baltimore, MD, USA
