+++ CITATION.cff
@@ -34,6 +34,9 @@ authors:
 - family-names: "Bilgel"
   given-names: "Murat"
   orcid: "https://orcid.org/0000-0001-5042-7422"
+- family-names: "Eierud"
+  given-names: "Cyrus"
+  orcid: "https://orcid.org/0000-0002-9942-676X"
 - family-names: "Pernet"
   given-names: "Cyril"
   orcid: "https://orcid.org/0000-0003-4010-4632"
