+++ CITATION.cff
@@ -41,7 +41,7 @@ authors:
   given-names: "Cyril"
   orcid: "https://orcid.org/0000-0003-4010-4632"
 title: "PET2BIDS"
-version: 1.0.0
+version: 1.3.20240502
 url: "https://github.com/openneuropet/PET2BIDS"
 date-released: 2022-10-01
 license: MIT
