### 2024-05-02 openneuropet/PET2BIDS#301 @bendhouseart
Done
