### 2024-05-02 #6067 @bendhouseart
> Double check authors and affiliations (including ORCIDs)

Authors orcid ID's match in paper, citations.cff and zenodo.org

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

Tagged and released as v1.3.20240502, most updates had to do with aligning the various files that contained author info. This author is going to write a script to update these in the future as there are too many to keep track of.

 > Archive the release on Zenodo/figshare/etc and post the DOI here.

Frozen forever at [zenodo.org](https://zenodo.org/records/11105167) with DOI:

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.11105167.svg)](https://doi.org/10.5281/zenodo.11105167)

Will see about automating this in the future so it's no so odious.

 > Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Indexing for affiliations, but everything else looks right.

> Make sure that the license listed for the archive is the same as the software license.

Both are MIT.
