### 2024-05-06 #6067 @britta-wstnr
@bendhouseart the difference in indexing comes from your change [here](https://github.com/openneuropet/PET2BIDS/commit/1e9d5dad37e7a86712634c2ddd064b01bd65c1ff) where you changed the author order.
Would you mind motivating this late change in author order for us? Thanks!
CC @CPernet
