### 2024-05-07 63b5f4e
Change main header frame number from exception to warning, fixed image time dimension to number of subheaders.
