### 2024-05-08 #6067 @britta-wstnr
Hi @mnoergaard and @CPernet,
thanks for explaining and sending the abstract. There are other author discrepancies between the abstract and this paper, I assume that is intentional?
Regarding the affiliations, I would suggest re-ordering them in your  `paper.md` for a clean look (such that M. Norgaard has affiliation numbers 2, 3, 4 instead of 3, 4, 5). CC @bendhouseart
