### 2024-05-09 161ca42
fixed min/max, need to check difference between outputs before final :thumbs up:
