### 2024-05-09 e923ed3
Update ecat2nii.py
* Use more idiomatic way of constructing an affine
* Remove unnecessary header zeroing, use helper methods where appropriate
