### 2024-05-09 2392575
Update matlab/nii_tool.m remove print
