+++ matlab/nii_tool.m
@@ -453,7 +453,6 @@ elseif strcmpi(cmd, 'save')
     % Write nii image
     fwrite(fid, nii.img, fmt);
     fclose(fid); % all written
-    disp(fmt)
 
     % gzip if asked
     if do_gzip
