### 2024-05-09 6dd9a78
Update matlab/ecat2nii.m uncomment commented out catch
