### 2024-05-09 f946848
put try/catch back in
