+++ ecat_testing/ints/nii_tool.m
@@ -1,1249 +1,0 @@
-function varargout = nii_tool(cmd, varargin)
-%
-% Copyright (c) 2016, Xiangrui Li https://github.com/xiangruili/dicm2nii
-% BSD-2-Clause License :: commit Mar 8, 2021
-% Basic function to create, load and save NIfTI file.
-% 
-% :format: - rst = nii_tool('cmd', para);
-% 
-% % .. note:: To list all command, type 'nii_tool ?'
-% 
-%     To get help information for each command, include '?' in cmd, for example:
-%     nii_tool init?
-%     nii_tool('init?')
-% 
-%     Here is a list of all command:
-% 
-%     nii_tool('default', 'version', 1, 'rgb_dim', 1);
-%     nii = nii_tool('init', img);
-%     nii = nii_tool('update', nii, mat);
-%     nii_tool('save', nii, filename, force_3D);
-%     hdr = nii_tool('hdr', filename);
-%     img = nii_tool('img', filename_or_hdr);
-%     ext = nii_tool('ext', filename_or_hdr);
-%     nii = nii_tool('load', filename_or_hdr);
-%     nii = nii_tool('cat3D', filenames);
-%     nii_tool('RGBStyle', 'afni');
-% 
-%     Detail for each command is described below.
-% 
-%     oldVal = nii_tool('default', 'version', 1, 'rgb_dim', 1);
-%     oldVal = nii_tool('default', struct('version', 1, 'rgb_dim', 1));
-% 
-%     - Set/query default NIfTI version and/or rgb_dim. To check the setting, run
-%      nii_tool('default') without other input. The input for 'default' command can
-%      be either a struct with fields of 'version' and/or 'rgb_dim', or
-%      parameter/value pairs. See nii_tool('RGBstyle') for meaning of rgb_dim.
-%      
-%      Note that the setting will be saved for future use. If one wants to change the
-%      settting temporarily, it is better to return the oldVal, and to restore it
-%      after done:
-%      
-%      oldVal = nii_tool('default', 'version', 2); % set version 2 as default
-%      % 'init' and 'save' NIfTI using above version
-%      nii_tool('default', 'version', oldVal); % restore default setting
-%     
-%     The default version setting affects 'init' command only. If you 'load' a NIfTI
-%     file, modify it, and then 'save' it, the version will be the same as the
-%     original file, unless it is changed explicitly (see help for 'save' command).
-%     All 'load' command ('load', 'hdr', 'ext', 'img') will read any version
-%     correctly, regardless of version setting.
-%     
-%     nii = nii_tool('init', img, RGB_dim);
-%     
-%     - Initialize nii struct based on img, normally 3D or 4D array. Most fields in
-%     the returned nii.hdr contain default values, and need to be updated based on
-%     dicom or other information. Important ones include pixdim, s/qform_code and
-%     related parameters.
-%     
-%     The NIfTI datatype will depend on data type of img. Most Matlab data types are
-%     supported, including 8/16/32/64 bit signed and unsigned integers, single and
-%     double floating numbers. Single/double complex and logical array are also
-%     supported.
-%     
-%     The optional third input, RGB_dim, is needed only if img contains RGB/RGBA
-%     data. It specifies which dimension in img encodes RGB or RGBA. In other words,
-%     if a non-empty RGB_dim is provided, img will be interpreted as RGB/RGBA data.
-%     
-%     Another way to signify RGB/RGBA data is to permute color dim to 8th-dim of img
[…]
