### 2024-05-09 a029106
add ad-hoc final test script
