+++ pypet2bids/pypet2bids/read_ecat.py
@@ -13,6 +13,7 @@ glutton or twisted in that sort of way.
 :Copyright: Open NeuroPET team
 
 """
+
 import json
 import os.path
 import struct
@@ -29,22 +30,24 @@ data_dir = code_dir.parent
 
 # debug variable
 # save steps for debugging for more info see ecat_testing/README.md
-ecat_save_steps = os.environ.get('ECAT_SAVE_STEPS', 0)
-if ecat_save_steps == '1':
+ecat_save_steps = os.environ.get("ECAT_SAVE_STEPS", 0)
+if ecat_save_steps == "1":
     # check to see if the code directory is available, if it's not create it and
     # the steps dir to save outputs created if ecat_save_steps is set to 1
-    steps_dir = code_dir.parent / 'ecat_testing' / 'steps'
+    steps_dir = code_dir.parent / "ecat_testing" / "steps"
     if not steps_dir.is_dir():
-        os.makedirs(code_dir.parent / 'ecat_testing' / 'steps', exist_ok=True)
+        os.makedirs(code_dir.parent / "ecat_testing" / "steps", exist_ok=True)
 
 
 # collect ecat header maps, this program will not work without these as ECAT data varies in the byte location of its
 # data depending on the version of ECAT it was formatted with.
 try:
-    with open(join(parent_dir, 'ecat_headers.json'), 'r') as infile:
+    with open(join(parent_dir, "ecat_headers.json"), "r") as infile:
         ecat_header_maps = json.load(infile)
 except FileNotFoundError:
-    raise Exception("Unable to load header definitions and map from ecat_headers.json. Aborting.")
+    raise Exception(
+        "Unable to load header definitions and map from ecat_headers.json. Aborting."
+    )
 
 
 # noinspection PyShadowingNames
@@ -59,7 +62,7 @@ def get_ecat_bytes(path_to_ecat: str):
     """
     # check if file exists
     if path.isfile(path_to_ecat):
-        with open(path_to_ecat, 'rb') as infile:
+        with open(path_to_ecat, "rb") as infile:
             ecat_bytes = infile.read()
     else:
         raise Exception(f"No such file found at {path_to_ecat}")
@@ -80,7 +83,7 @@ def read_bytes(path_to_bytes: str, byte_start: int, byte_stop: int = -1):
         raise Exception(f"{path_to_bytes} is not a valid file.")
 
     # open that file
-    bytes_to_read = open(path_to_bytes, 'rb')
+    bytes_to_read = open(path_to_bytes, "rb")
 
     # move to start byte
     bytes_to_read.seek(byte_start, 0)
@@ -93,7 +96,9 @@ def read_bytes(path_to_bytes: str, byte_start: int, byte_stop: int = -1):
     return sought_bytes
 
 
-def collect_specific_bytes(bytes_object: bytes, start_position: int = 0, width: int = 0):
+def collect_specific_bytes(
+    bytes_object: bytes, start_position: int = 0, width: int = 0
+):
     """
     Collects specific bytes within a bytes object.
 
@@ -104,7 +109,7 @@ def collect_specific_bytes(bytes_object: bytes, start_position: int = 0, width: 
     :return: the bytes starting at position
     """
     # navigate to byte position
-    content = bytes_object[start_position: start_position + width]
+    content = bytes_object[start_position : start_position + width]
     return {"content": content, "new_position": start_position + width}
 
 
@@ -117,17 +122,19 @@ def get_buffer_size(data_type: str, variable_name: str):
     :param variable_name:
     :return: the number of bytes to expand a buffer to
     """
-    first_split = variable_name.split('(')
+    first_split = variable_name.split("(")
[…]
