+++ pypet2bids/pypet2bids/ecat2nii.py
@@ -10,7 +10,10 @@ import datetime
 import nibabel
 import numpy
 import pathlib
-from pypet2bids.read_ecat import read_ecat, code_dir  # we use code_dir for ecat debugging
+from pypet2bids.read_ecat import (
+    read_ecat,
+    code_dir,
+)  # we use code_dir for ecat debugging
 import os
 import pickle
 import logging
@@ -23,30 +26,32 @@ except ModuleNotFoundError:
 
 # debug variable
 # save steps for debugging, for mor infor see ecat_testing/README.md
-ecat_save_steps = os.environ.get('ECAT_SAVE_STEPS', 0)
-if ecat_save_steps == '1':
+ecat_save_steps = os.environ.get("ECAT_SAVE_STEPS", 0)
+if ecat_save_steps == "1":
     # check to see if the code directory is available, if it's not create it and
     # the steps dir to save outputs created if ecat_save_steps is set to 1
-    steps_dir = code_dir.parent / 'ecat_testing' / 'steps'
+    steps_dir = code_dir.parent / "ecat_testing" / "steps"
     if not steps_dir.is_dir():
-        os.makedirs(code_dir.parent / 'ecat_testing' / 'steps', exist_ok=True)
+        os.makedirs(code_dir.parent / "ecat_testing" / "steps", exist_ok=True)
 
 logger = logging.getLogger("pypet2bids")
 
 
-def ecat2nii(ecat_main_header=None,
-             ecat_subheaders=None,
-             ecat_pixel_data=None,
-             ecat_file=None,
-             nifti_file: str = '',
-             sif_out=False,
-             affine=None,
-             save_binary=False,
-             **kwargs):
+def ecat2nii(
+    ecat_main_header=None,
+    ecat_subheaders=None,
+    ecat_pixel_data=None,
+    ecat_file=None,
+    nifti_file: str = "",
+    sif_out=False,
+    affine=None,
+    save_binary=False,
+    **kwargs,
+):
     """
     Converts an ECAT file into a nifti and a sidecar json, used in conjunction with
     :func:`pypet2bids.read_ecat.read_ecat`
-    
+
     :param ecat_main_header: the main header of an ECAT file
     :param ecat_subheaders: the subheaders for each frame of the ECAT file
     :param ecat_pixel_data: the imaging/pixel data from the ECAT file
@@ -75,21 +80,36 @@ def ecat2nii(ecat_main_header=None,
     nifti_file_w_out_extension = os.path.splitext(str(pathlib.Path(nifti_file).name))[0]
 
     # if already read nifti file skip re-reading
-    if ecat_main_header is None and ecat_subheaders is None and ecat_pixel_data is None and ecat_file:
+    if (
+        ecat_main_header is None
+        and ecat_subheaders is None
+        and ecat_pixel_data is None
+        and ecat_file
+    ):
         # collect ecat_file
         main_header, sub_headers, data = read_ecat(ecat_file=ecat_file)
-    elif ecat_file is None and type(ecat_main_header) is dict and type(ecat_subheaders) is list and type(
-            ecat_pixel_data) is numpy.ndarray:
-        main_header, sub_headers, data = ecat_main_header, ecat_subheaders, ecat_pixel_data
+    elif (
+        ecat_file is None
+        and type(ecat_main_header) is dict
+        and type(ecat_subheaders) is list
+        and type(ecat_pixel_data) is numpy.ndarray
+    ):
+        main_header, sub_headers, data = (
+            ecat_main_header,
+            ecat_subheaders,
+            ecat_pixel_data,
+        )
     else:
-        raise Exception("Must pass in filepath for ECAT file or "
-                        "(ecat_main_header, ecat_subheaders, and ecat_pixel data "
[…]
