### 2024-05-10 0e3422c
Apply suggestions from code review
fixed % and added rms

Co-authored-by: Chris Markiewicz <effigies@gmail.com>
