### 2024-05-10 ad7368f
Update ecat_testing/read_matlab_nii.py
Co-authored-by: Chris Markiewicz <effigies@gmail.com>
