### 2024-05-10 eef3fd4
fixed int scaling/display/rollover error observed with ecat2nii.py and ecat2nii.m
* update python to go along with matlab changes in 9beee53
* moved int testing to this repo
* added function for first_middle_last_frames_to_text for arrays since they're different than cell arrays.
* fixed min/max, need to check difference between outputs before final
* Update ecat2nii.py (#308)
* Use more idiomatic way of constructing an affine
* Remove unnecessary header zeroing, use helper methods where appropriate
* add ad-hoc final test script
* run black, add conditional for testing
* Apply suggestions fro
[…]
