+++ matlab/ecat2nii.m
@@ -48,6 +48,16 @@ sifout  = false; % 0/1 to indicate if sif file are also created
 gz      = true;  % compress nifti
 savemat = false; % save ecat data as .mat
 
+%% debebugging
+% ------------
+ecat_save_steps = getenv('ECAT_SAVE_STEPS');
+ecat_save_steps_dir = mfilename('fullpath');
+if contains(ecat_save_steps_dir, 'LiveEditorEvaluationHelper')
+    ecat_save_steps_dir = matlab.desktop.editor.getActiveFilename;
+end
+parts = strsplit(ecat_save_steps_dir, filesep);
+ecat_save_steps_dir = strjoin([parts(1:end-2), 'ecat_testing', 'steps'], filesep);
+
 %% check inputs
 % ------------
 
@@ -143,7 +153,10 @@ for j=1:length(FileListIn)
         end
         
         pet_file = [pet_file ext];
-        [mh,sh]  = readECAT7([pet_path filesep pet_file]); % loading the whole file here and iterating to flipdim below only minuimally improves time (0.6sec on NRU server)
+        [mh,sh,data]  = readECAT7([pet_path filesep pet_file]); % loading the whole file here and iterating to flipdim below only minuimally improves time (0.6sec on NRU server)
+        if (ecat_save_steps == '1')
+            first_middle_last_frames_to_text_cell(data,ecat_save_steps_dir, '6_ecat2nii_matlab');
+        end
         if sh{1}.data_type ~= 6
             error('Conversion for 16 bit data only (type 6 in ecat file) - error loading ecat file');
         end
@@ -166,7 +179,13 @@ for j=1:length(FileListIn)
                 Randoms(i)    = 0;
             end
         end
-        
+
+        % save debugging steps 6 and 7
+        if (ecat_save_steps == '1')
+            first_middle_last_frames_to_text_cell(data,ecat_save_steps_dir, '6_ecat2nii_matlab');
+            first_middle_last_frames_to_text(img_temp,ecat_save_steps_dir, '7_flip_ecat2nii_matlab');
+        end
+
         % rescale to 16 bits
         rg = max(img_temp(:))-min(img_temp(:));
         if rg ~= 32767 % same as range(img_temp(:)) but no need of stats toolbox
@@ -178,8 +197,22 @@ for j=1:length(FileListIn)
                 img_temp = img_temp/MinImg*(-32768);
                 Sca      = Sca*MinImg/(-32768);
             end
+
+            % save scaling factor to file located at ecat_save_steps_dir/8.5_sca_matlab.txt
+            fid = fopen([ecat_save_steps_dir filesep '8.5_sca_matlab.txt'],'w');
+            fprintf(fid,'Scaling factor: %10e\n',Sca);
+            x = mh.ecat_calibration_factor * Sca;
+            fprintf(fid,'Scaling factor * ECAT Cal Factor: %10.10f\n',x);
+            fclose(fid);
         end
-        
+
+
+
+        % save debugging step 8 - rescale to 16 bits
+        if (ecat_save_steps == '1')
+            first_middle_last_frames_to_text(img_temp,ecat_save_steps_dir, '8_rescale_to_16_ecat2nii_matlab');
+        end
+
         % check names
         if ~exist('FileListOut','var')
             [~,pet_filename]           = fileparts(pet_file);
@@ -215,7 +248,7 @@ for j=1:length(FileListIn)
         end
         
         % save raw data
-        if savemat
+        if savemat or ecat_save_steps == '1'
             if mh.calibration_units == 1 % see line 337
                 ecat = img_temp.*Sca;
             else
@@ -336,6 +369,9 @@ for j=1:length(FileListIn)
         
         if mh.calibration_units == 1 % do calibrate
[…]

[…]
