### 2024-05-15 f5d7bba
add case ecat annotation (method) is missing but user provides it
