### 2024-06-04 e49524e
Update JHU and Canon phantom conversions (#281)
* update python conversion script for newer phantoms
* Add 3DRP recon; fix phantom conversions


---------

Co-authored-by: Anthony Galassi <28850131+bendhouseart@users.noreply.github.com>
