+++ scripts/python_conversions.sh
@@ -299,7 +299,9 @@ FrameTimesStart=[0] \
 FrameDuration=[300] \
 AttenuationCorrection="MR-corrected" \
 RandomsCorrectionMethod="DLYD" \
-ReconFilterSize=1 
+ReconFilterSize=1 \
+--silent
+#DecayCorrectionFactor="[1]" \
 
 # General Electric Medical Systems Signa PET-MR
 # ----------------------------------------------
@@ -327,7 +329,8 @@ ReconFilterType="n/a" \
 ReconFilterSize=1 \
 ReconMethodParameterLabels="[none, none]" \
 ReconMethodParameterUnits="[none, none]" \
-ReconMethodParameterValues="[0, 0]"
+ReconMethodParameterValues="[0, 0]" \
+--silent
 
 
 # General Electric Medical Systems Advance
@@ -350,11 +353,15 @@ ImageDecayCorrected='true' \
 AcquisitionMode='list mode' \
 ImageDecayCorrectionTime="0" \
 ScatterCorrectionMethod="Convolution subtraction" \
-FrameDuration=[98000] \
+FrameDuration="[98000]" \
 ScanStart="0" \
 ReconMethodParameterLabels="[none]" \
 ReconMethodParameterUnits="[none]" \
-ReconMethodParameterValues="[0, 0]"
+ReconMethodParameterValues="[0, 0]" \
+--silent
+#TimeZero="13:39:41" \
+#AttenuationCorrection='n/a' \
+
 
 echo "${SOURCE_FOLDER}/GeneralElectricAdvance-NIMH"
 dcm2niix4pet $SOURCE_FOLDER/GeneralElectricAdvance-NIMH/long_trans --destination-path $DESTINATION/sub-GeneralElectricAdvanceLongNIMH/pet --kwargs \
@@ -374,38 +381,108 @@ ImageDecayCorrected='false' \
 AttenuationCorrection='measured' \
 AcquisitionMode='list mode' \
 ImageDecayCorrectionTime="0" \
-FrameDuration=[98000] \
+FrameDuration="[98000]" \
 ScatterCorrectionMethod="Gaussian Fit" \
 ScanStart="0" \
 ReconMethodParameterLabels="[none, none]" \
 ReconMethodParameterUnits="[none, none]" \
-ReconMethodParameterValues="[0, 0]"
+ReconMethodParameterValues="[0, 0]" \
+--silent
+
+# Johns Hopkins University
+# ------------------------
 
 # Siemens HRRT
 # ------------
-echo "${SOURCE_FOLDER}/SiemensHRRT-JHU/Hoffman.v"
-ecatpet2bids $SOURCE_FOLDER/SiemensHRRT-JHU/Hoffman.v --nifti $DESTINATION/sub-SiemensHRRTJHU/pet/sub-SiemensHRRTJHU_pet.nii --convert --kwargs \
-Manufacturer=Siemens \
-ManufacturersModelName=HRRT \
-InstitutionName="JHU" \
-BodyPart="Phantom" \
-Units="Bq/mL" \
-TracerName=FDG \
-TracerRadionuclide=F18 \
-InjectedRadioactivity=81.24 \
-SpecificRadioactivity="1.3019e+04" \
-ModeOfAdministration=infusion \
-AcquisitionMode="list mode" \
-ImageDecayCorrected="true" \
+echo "${SOURCE_FOLDER}/SiemensHRRT-JHU"
+ecatpet2bids $SOURCE_FOLDER/SiemensHRRT-JHU/Hoffman.v \
+--nifti $DESTINATION/sub-SiemensHRRTJHU/pet/sub-SiemensHRRTJHU_pet.nii.gz \
+--convert \
+--kwargs \
+Manufacturer='Siemens' \
+ManufacturersModelName='HRRT' \
[…]
