+++ pypet2bids/pypet2bids/helper_functions.py
@@ -725,8 +725,6 @@ def get_recon_method(ReconstructionMethodString: str) -> dict:
     contents = ReconstructionMethodString.replace(" ", "")
     subsets = None
     iterations = None
-    ReconMethodParameterUnits = ["none", "none"]
-    ReconMethodParameterLabels = ["subsets", "iterations"]
 
     # determine order of recon iterations and subsets, this is not  a surefire way to determine this...
     iter_sub_combos = {
@@ -836,11 +834,20 @@ def get_recon_method(ReconstructionMethodString: str) -> dict:
         ReconMethodName = dimension + " " * len(dimension) + expanded_name.rstrip()
         ReconMethodName = " ".join(ReconMethodName.split())
 
+    if ReconMethodName in ["Filtered Back Projection", "3D Reprojection"]:
+        ReconMethodParameterLabels = []
+        ReconMethodParameterUnits = []
+        ReconMethodParameterValues = []
+    else:  # assume it is OSEM or a variant
+        ReconMethodParameterLabels = ["subsets", "iterations"]
+        ReconMethodParameterUnits = ["none", "none"]
+        ReconMethodParameterValues = [subsets, iterations]
+
     reconstruction_dict = {
         "ReconMethodName": ReconMethodName,
         "ReconMethodParameterUnits": ReconMethodParameterUnits,
         "ReconMethodParameterLabels": ReconMethodParameterLabels,
-        "ReconMethodParameterValues": [subsets, iterations],
+        "ReconMethodParameterValues": ReconMethodParameterValues,
     }
 
     if None in reconstruction_dict["ReconMethodParameterValues"]:
@@ -952,11 +959,11 @@ def ad_hoc_checks(
     if items_that_should_be_checked is None:
         items_that_should_be_checked = {}
     hardcoded_items = {
-        "InjectedRadioactivityUnits": "MBq",
-        "SpecificRadioactivityUnits": ["Bq/g", "MBq/ug"],
-        "InjectedMassUnits": "ug",
-        "MolarActivityUnits": "GBq/umolug",
-        "MolecularWeightUnits": "g/mol",
+        'InjectedRadioactivityUnits': ['MBq', 'mCi'],
+        'SpecificRadioactivityUnits': ['Bq/g', 'MBq/ug'],
+        'InjectedMassUnits': 'ug',
+        'MolarActivityUnits': 'GBq/umolug',
+        'MolecularWeightUnits': 'g/mol'
     }
 
     # if none are
