+++ scripts/matlab_conversions.m
@@ -2,11 +2,11 @@
 % note that usually we use the *parameter.txt file facilitating the metadata
 % information but here it is done fully for completeness
 % 
-% cyril pernet - may 2022
+% cyril pernet - Nov 2023
 
-%% set paths where the repo is
-%source      = 'D:\BIDS\ONP\OpenNeuroPET-Phantoms\sourcedata\';
-%destination = 'D:\BIDS\ONP\OpenNeuroPET-Phantoms\';
+%% set paths where the repo is to bypass the path/env parts below
+% source      = 'D:\BIDS\ONP\OpenNeuroPET-Phantoms\sourcedata\';
+% destination = 'D:\BIDS\ONP\OpenNeuroPET-Phantoms\';
 
 %% get the path to this script
 matlab_conversions_script_path = mfilename('fullpath');
@@ -20,7 +20,9 @@ phantoms_folder_path = strjoin(parts(1:end-2),filesep);
 source      = strjoin({phantoms_folder_path, 'sourcedata'}, filesep);
 destination = strjoin({phantoms_folder_path, 'matlab'}, filesep);
 
-mkdir(destination)
+if ~exist(destination,'dir')
+    mkdir(destination)
+end
 copyfile(strjoin({phantoms_folder_path, 'dataset_description.json'}, filesep), strjoin({destination, 'dataset_description.json'}, filesep))
 
 %% sets the environment variable so that matlab can reach dcm2niix on posix
@@ -48,6 +50,7 @@ else
 end
 
 message = 'Failed to convert subject, moving onto next.';
+
 
 %%% Neurobiology Research Unit - Copenhagen
 %% ----------------------------------------
@@ -436,13 +439,14 @@ catch
 end
 
 try
-meta.AttenuationCorrection = 'measured' ; % some how the field is not there
-meta.ReconMethodParameterLabels = ["none", "none"];
-meta.ReconMethodParameterUnits = ["none", "none"];
-meta.ReconMethodParameterValues = [0, 0];
-meta.FrameDuration = 98000
-dcm2niix4pet(fullfile(source,['GeneralElectricAdvance-NIMH' filesep ...
-    'long_trans']),meta,'o',fullfile(destination,['sub-GeneralElectricAdvanceLongNIMH' filesep 'pet'])); 
+    meta.AttenuationCorrection = 'measured' ; % some how the field is not there
+    meta.ReconMethodParameterLabels = ["none", "none"];
+    meta.ReconMethodParameterUnits = ["none", "none"];
+    meta.ReconMethodParameterValues = [0, 0];
+    meta.FrameDuration = 98000;
+
+    dcm2niix4pet(fullfile(source,['GeneralElectricAdvance-NIMH' filesep ...
+        'long_trans']),meta,'o',fullfile(destination,['sub-GeneralElectricAdvanceLongNIMH' filesep 'pet']));
 catch
     disp(message);
 end
@@ -476,6 +480,123 @@ try
     out = ecat2nii(fullfile(source,['SiemensHRRT-JHU' filesep 'Hoffman.v']),...
         meta,'gz',true,'FileListOut',fullfile(destination,['sub-SiemensHRRTJHU'  filesep 'pet' filesep 'sub-SiemensHRRTJHU.nii']));
 
+
+%% Johns Hopkins University
+% ----------------------------------------------
+
+% Siemens HRRT
+% ------------
+try
+    clear meta
+    meta.TimeZero                   = 'ScanStart';
+    meta.Manufacturer               = 'Siemens';
+    meta.ManufacturersModelName     = 'HRRT';
+    meta.InstitutionName            = 'Johns Hopkins University, USA';
+    meta.BodyPart                   = 'Phantom';
+    meta.Units                      = 'Bq/mL';
[…]
