+++ .github/workflows/matlab.yaml
@@ -51,7 +51,7 @@ jobs:
             addpath('${{ env.MATLAB_CODE_PATH }}');
             addpath('${{ env.MATLAB_CODE_PATH }}/unit_tests');
             addpath('${{ env.MATLAB_CODE_PATH }}/template_txt');
-            addpath('${{ env.TEST_ECAT_PATH }}');
+            addpath('${{ env.MATLAB_CODE_PATH }}/ecat_validation');
             addpath('${{ env.METADATA_PATH }}');
             ecat2nii_test
         if: matrix.os == 'ubuntu-latest'
