+++ .github/workflows/matlab.yaml
@@ -48,10 +48,10 @@ jobs:
         uses: matlab-actions/run-command@v2
         with:
           command: |
-            addpath('${{ env.MATLAB_CODE_PATH }}');
-            addpath('${{ env.MATLAB_CODE_PATH }}/unit_tests');
-            addpath('${{ env.MATLAB_CODE_PATH }}/template_txt');
-            addpath('${{ env.MATLAB_CODE_PATH }}/ecat_validation');
-            addpath('${{ env.METADATA_PATH }}');
-            ecat2nii_test('${{ env.TEST_ECAT_PATH }}');
+            addpath('${{ env.MATLAB_CODE_PATH }}')
+            addpath('${{ env.MATLAB_CODE_PATH }}/unit_tests')
+            addpath('${{ env.MATLAB_CODE_PATH }}/template_txt')
+            addpath('${{ env.MATLAB_CODE_PATH }}/ecat_validation')
+            addpath('${{ env.METADATA_PATH }}')
+            ecat2nii_test('${{ env.TEST_ECAT_PATH }}')
         if: matrix.os == 'ubuntu-latest'
