+++ matlab/unit_tests/ecat2nii_test.m
@@ -25,7 +25,6 @@ else
     ecatfile = varargin{1};
 end
 
-groundtruth = [ecatfile(1:end-5) '.mat'];
 meta.info = 'just running a test';
 meta.TimeZero = datestr(now,'hh:mm:ss');
 cd(fileparts(ecatfile))
@@ -65,68 +64,76 @@ else
     delete(fullfile(filepath,[filename(1:end-2) '.ecat.mat']))
 end
 
-figure
-A = min(Diff_ecat_nifti(:));
-B = max(Diff_ecat_nifti(:));
-if exist('groundtruth','var')
-    subplot(1,3,1);
-    plot(img,sort(img_reread(:)),'*');
-    M = max(max([img(:) img_reread(:)]));
-    axis([0 numel(img(:))*2 0 M]); hold on;
-    plot([0 numel(img(:))*2],[0 M],'r','LineWidth',1); 
+% if this is running in CI don't plot things
+
+CI = getenv('CI');
+
+if strcmp(CI, 'true')
+    disp("Running in CI, not attempting to plot.")
 else
-    subplot(2,2,1);
-    plot(img(:),img_reread(:),'*');
-    M = max(max([img(:) img_reread(:)])); 
-    axis([0 M 0 M]); hold on;
-    plot([0 M],[0 M],'r','LineWidth',1); 
-end
-xlabel('Original'); ylabel('Nifti'); 
-grid on; title('Read vs Written');
-
-[simg,index] = sort(img(:)); 
-if exist('groundtruth','var')
-    subplot(1,3,2);
-else
-    subplot(2,2,3);
-end
-plot(img(:),(simg-img_reread(index)),'*'); grid on
-xlabel('Original'); ylabel('Difference'); 
-title(sprintf('Average error: %f\n',meandiff));
-
-if exist('groundtruth','var')
-    subplot(1,3,3);
-else
-    subplot(2,2,4);
-end
-histogram(Diff_ecat_nifti(:)); grid on
-axis([A B 0 10000]); xlabel('error'); 
-title('Distribution of all errors')
-
-if ~exist('groundtruth','var')
-    
-    subplot(2,2,2); plot(summary_diff(:,2),'LineWidth',2); grid on; xlabel('frames');
-    ylabel('Difference'); title('Avg error per frame')
-    saveas(gcf, [filename '.jpg'],'jpg');
-
-    vidObj = VideoWriter([filename '_error']);
-    open(vidObj); figure('Name','error per frames');
-    set(gcf,'Color','w','InvertHardCopy','off', 'units','normalized', 'outerposition',[0 0 1 1]);
-    for f=1:size(Diff_ecat_nifti,4)
-        for rep = 1:8 % use to slow down the video
-            clf
-            subplot(3,2,[1 2 3 4]);
-            all = frame{f}(:); all(all==0)=[];
-            histogram(all); axis([-13 13 -0.5 50])
-            xlabel('error values'); ylabel('freq.');
-            title(sprintf('errors frame %g max %g%%',f,max(all(:))/max(max(max(img(:,:,:,f))))*100));
-            
-            subplot(3,2,[5 6]); plot(signal,'LineWidth',2); axis tight
-            hold on; plot(f,signal(f),'rO','LineWidth',4); axis tight;
-            xlabel('frames'); ylabel('all brain TAC'); grid on; hold off
-            
-            CF = getframe(gcf); writeVideo(vidObj,CF);
-        end
+    figure
+    A = min(Diff_ecat_nifti(:));
+    B = max(Diff_ecat_nifti(:));
+    if exist('groundtruth','var')
+        subplot(1,3,1);
+        plot(img,sort(img_reread(:)),'*');
[…]
