### 2024-06-06 4e82396
move ground truth to get initialized
