+++ matlab/unit_tests/ecat2nii_test.m
@@ -20,10 +20,10 @@ function ecat2nii_test(varargin)
 if nargin ==0 || isempty(varargin{1})
     ecatfile = fullfile(fileparts(fileparts(fileparts(which('ecat2nii_test.m')))),...
         ['ecat_validation' filesep 'synthetic_ecat_integer_16x16x16x4.v.gz']);
-    groundtruth = [ecatfile(1:end-5) '.mat'];
 else
     ecatfile = varargin{1};
 end
+groundtruth = [ecatfile(1:end-5) '.mat'];
 
 meta.info = 'just running a test';
 meta.TimeZero = datestr(now,'hh:mm:ss');
