### 2024-06-07 b16dd70
check if file instead of if var exists
