### 2024-06-07 ec11ac0
check installed matlab toolboxes
