+++ .github/workflows/matlab.yaml
@@ -19,7 +19,7 @@ jobs:
     runs-on: ${{ matrix.os }}
 
     env:
-      TEST_ECAT_PATH: ${{ github.workspace }}/ecat_validation/synthetic_ecat_integer.v.gz
+      TEST_ECAT_PATH: ${{ github.workspace }}/ecat_validation/synthetic_ecat_integer16x16x16x4.v.gz
       MATLAB_CODE_PATH: ${{ github.workspace }}/matlab/
       METADATA_PATH: ${{ github.workspace }}/metadata/
 
@@ -42,6 +42,12 @@ jobs:
         timeout-minutes: 15
         with:
           limit-access-to-actor: true
+
+      - name: Get list of Matlab Toolboxes
+        uses: matlab-actions/run-command@v2
+        with:
+         command: |
+           ver
 
         # Again only running this on ubuntu
       - name: Add Matlab Code Folder to Path and Run Tests
