+++ .github/workflows/matlab.yaml
@@ -34,6 +34,109 @@ jobs:
         # matlab not supported on MacOS!! at this time
       - name: Perform setup-matlab
         uses: matlab-actions/setup-matlab@v2
+        with:
+          release: R2023a
+          cache: true
+          products: >
+            Simulink
+            Simulink_Test
+            5G_Toolbox                                            
+            AUTOSAR_Blockset                                      
+            Aerospace-Blockset                                    
+            Aerospace_Toolbox                                     
+            Antenna_Toolbox                                       
+            Audio_Toolbox                                         
+            Automated_Driving_Toolbox                             
+            Bioinformatics_Toolbox                                
+            Bluetooth_Toolbox                                     
+            Communications_Toolbox                                
+            Computer_Vision_Toolbox                               
+            Control_System_Toolbox                                
+            Curve_Fitting_Toolbox                                 
+            DDS_Blockset                                          
+            DSP_HDL_Toolbox                                       
+            DSP_System_Toolbox                                    
+            Database_Toolbox                                      
+            Datafeed_Toolbox                                      
+            Deep_Learning_Toolbox                                 
+            Econometrics_Toolbox                                  
+            Embedded_Coder                                        
+            Filter_Design_HDL_Coder                               
+            Financial_Instruments_Toolbox                         
+            Financial_Toolbox                                     
+            Fixed-Point_Designer                                  
+            Fuzzy_Logic_Toolbox                                   
+            Global_Optimization_Toolbox                           
+            HDL_Coder                                             
+            Image_Acquisition_Toolbox                             
+            Image_Processing_Toolbox                              
+            Industrial_Communication_Toolbox                      
+            Instrument_Control_Toolbox                            
+            LTE_Toolbox                                           
+            Lidar_Toolbox                                         
+            MATLAB_Coder                                          
+            MATLAB_Compiler                                       
+            MATLAB_Compiler SDK                                   
+            MATLAB_Report Generator                               
+            MATLAB_Test                                           
+            Mapping_Toolbox                                       
+            Medical_Imaging_Toolbox                               
+            Mixed-Signal_Blockset                                 
+            Model_Predictive_Control_Toolbox                      
+            Motor_Control_Blockset                                
+            Navigation_Toolbox                                    
+            Optimization_Toolbox                                  
+            Parallel_Computing_Toolbox                            
+            Partial_Differential_Equation_Toolbox                 
+            Phased_Array_System_Toolbox                           
+            Powertrain_Blockset                                   
+            Predictive_Maintenance_Toolbox                        
+            RF_Blockset                                           
+            RF_PCB_Toolbox                                        
+            RF_Toolbox                                            
+            ROS_Toolbox                                           
+            Radar_Toolbox                                         
+            Reinforcement_Learning_Toolbox                        
+            Requirements_Toolbox                                  
+            Risk_Management_Toolbox                               
+            Robotics_System_Toolbox                               
+            Robust_Control_Toolbox                                
+            Satellite_Communications_Toolbox                      
+            Sensor_Fusion_and_Tracking_Toolbox                    
+            SerDes_Toolbox                                        
+            Signal_Processing_Toolbox                             
+            SimBiology                                            
+            SimEvents                                             
+            Simscape                                              
+            Simscape_Battery                                      
+            Simscape_Driveline                                    
+            Simscape_Electrical                                   
+            Simscape_Fluids                                       
+            Simscape_Multibody                                    
+            Simulink_3D_Animation                                 
+            Simulink_Check                                        
+            Simulink_Coder                                        
+            Simulink_Compiler                                     
+            Simulink_Control_Design                               
+            Simulink_Coverage                                     
[…]
