### 2024-06-07 21175ea
update phantom conversion scripts
