+++ .github/workflows/phantoms.yaml
@@ -143,10 +143,10 @@ jobs:
 
       - name: SetupMatlab
         if: ${{ matrix.os == 'ubuntu-20.04' }}
-        uses: matlab-actions/setup-matlab@v2
-        with:
-          release: 'R2023a'
-          cache: true
+        uses: matlab-actions/setup-matlab@v1
+        #with:
+          #release: 'R2023a'
+          #cache: true
 
       - name: Run Matlab Conversion Script on Phantoms
         if: ${{ matrix.os == 'ubuntu-20.04' }}
