### 2024-06-07 c90be5e
Fix phantoms ci (#312)
* update phantom conversion scripts
