### 2024-06-10 9209e88
[DATALAD RUNCMD] run codespell throughout fixing typos automagically (but ignoring overall fail due to ambigous ones)
=== Do not change lines below ===
{
 "chain": [
  "ae0c26e93ede67f21c604113cbb8265efdc32c8f"
 ],
 "cmd": "codespell -w || :",
 "exit": 0,
 "extra_inputs": [],
 "inputs": [],
 "outputs": [],
 "pwd": "."
}
^^^ Do not change lines above ^^^
