+++ matlab/run_m2docgen_example.m
@@ -71,7 +71,7 @@ opts = struct(  'toolboxName',      "PET2BIDS", ...
                 'outputFolder',     ['/Users/galassiae/Projects/PET2BIDS/matlab/docs'], ...
                 'excludeFolder',	["m2docgen_documentation"], ...
                 'excludeFile',      ["/Users/galassiae/Projects/PET2BIDS/matlab/check_metaradioinputs.m"], ...
-                'htmlMetaFolder',   "ressources", ...
+                'htmlMetaFolder',   "resources", ...
                 'htmlTemplate',     "m2doc-standard", ...
                 'startPage',        ["m2docgen_PET2BIDS.html"], ...
                 'toc',              [], ...
