+++ paper.md
@@ -12,7 +12,7 @@ authors:
     affiliation: 1
   - name: Martin Norgaard
     orcid: 0000-0003-2131-5688
-    affiliation: "3, 4, 5"
+    affiliation: "2, 3, 4"
   - name: Adam G. Thomas
     orcid: 0000-0002-2850-1419
     affiliation: 1
@@ -21,7 +21,7 @@ authors:
     affiliation: 6
   - name: Claus Svarer
     orcid: 0000-0001-7811-1825
-    affiliation: 3
+    affiliation: 2
   - name: Chris Rorden
     orcid: 0000-0002-7554-6142
     affiliation: 7
@@ -30,32 +30,32 @@ authors:
     affiliation: "8, 9"
   - name: Gitte M. Knudsen
     orcid: 0000-0003-1508-6866
-    affiliation: 3
+    affiliation: 2
   - name: Robert B. Innis
     orcid: 0000-0003-1238-7209
     affiliation: 1
   - name: Melanie Ganz
     orcid: 0000-0002-9120-8098
-    affiliation: "3, 4"
+    affiliation: "2, 3"
   - name: Cyrus Eierud
     orcid: 0000-0002-9942-676X
-    affiliation: 2
+    affiliation: 5
   - name: Murat Bilgel
     orcid: 0000-0001-5042-7422
     affiliation: 10
   - name: Cyril Pernet
     orcid: 0000-0003-4010-4632
-    affiliation: 3
+    affiliation: 2
 affiliations:
   - name: National Institutes of Health, Bethesda, MD, USA
     index: 1
-  - name: TReNDS Center, Georgia State University, Atlanta, GA, USA
-    index: 2
   - name: Neurobiology Research Unit, Rigshospitalet, Copenhagen, Denmark
-    index: 3
+    index: 2
   - name: Department of Computer Science, University of Copenhagen, Copenhagen, Denmark
-    index: 4
+    index: 3
   - name: Department of Psychology, Stanford University, CA, USA
+    index: 4
+  - name: TReNDS Center, Georgia State University, Atlanta, GA, USA
     index: 5
   - name: University Medical Center of the Johannes Gutenberg University Mainz, Mainz, Germany
     index: 6
