### 2024-06-14 f6e221b
Add codespell support (config, workflow to detect/not fix) and make it fix few typos (#313)
* Add github action to codespell main on push and PRs

* Add rudimentary codespell config

* Ignore PROMIS

* [DATALAD RUNCMD] Some custom ambigous fixes

=== Do not change lines below ===
{
 "chain": [],
 "cmd": "git-sedi 'thes frame' 'the frame'; git-sedi labled labeled",
 "exit": 0,
 "extra_inputs": [],
 "inputs": [],
 "outputs": [],
 "pwd": "."
}
^^^ Do not change lines above ^^^

* [DATALAD RUNCMD] run codespell throughout fixing typos automagically (but ignoring overall fail due to ambigous ones)

[…]
