+++ pypet2bids/pypet2bids/convert_pmod_to_blood.py
@@ -47,7 +47,7 @@ epilog = textwrap.dedent(
 example1 = textwrap.dedent(
     """
 Usage examples are below that verbosely describe and show the input consumed an the output generated by this program.
-Additonal arguments/fields are passed via the kwargs flag in key value pairs.
+Additional arguments/fields are passed via the kwargs flag in key value pairs.
 
 Note: lines prepended with # denote comments/notes where as lines without # denote data or input arguments
 
@@ -205,7 +205,7 @@ def type_cast_cli_input(kwarg_arg):
         if type(var) in [dict, list, str, int, float, bool]:
             return var
     except (ValueError, SyntaxError):
-        # try truthy evals if the input doesn't evalute as a python type listed in the try statement
+        # try truthy evals if the input doesn't evaluate as a python type listed in the try statement
         if kwarg_arg.lower() in ["true", "t", "yes"]:
             return True
         elif kwarg_arg.lower() in ["false", "f", "no"]:
@@ -762,7 +762,7 @@ class PmodToBlood:
             elif str.lower(how) == "y":
                 self.data_collection[recording] = "manual"
                 warnings.warn(
-                    f"Received {how} as input, assuming input recieved from cli w/ '-y' option on bash/zsh etc, "
+                    f"Received {how} as input, assuming input received from cli w/ '-y' option on bash/zsh etc, "
                     f"defaulting to manual input"
                 )
                 break
@@ -779,7 +779,7 @@ class PmodToBlood:
         :return: None
         :rtype: None
         """
-        # first we combine the various blood datas into one or two dataframes, the autosampled data goes into a
+        # first we combine the various blood data into one or two dataframes, the autosampled data goes into a
         # recording_autosample, and the manually sampled data goes into a recording_manual if they exist
         if self.subject_id:
             file_path = join(self.output_path, self.subject_id + "_")

[…]
