### 2024-06-17 #6067 @bendhouseart
@britta-wstnr not sure what you prefer, either commit 0b790f470a or the author/affiliation order in fe0faa1be. My only requirements are that the author order conforms to:

1) Me
2) Martin
-1) Cyril

It's difficult to get a nice _neat_ numbering and ordering of authors w/ affiliations while keeping the above intact.

Thanks,

Anthony
