+++ paper.md
@@ -16,33 +16,33 @@ authors:
   - name: Adam G. Thomas
     orcid: 0000-0002-2850-1419
     affiliation: 1
-  - name: Robert B. Innis
-    orcid: 0000-0003-1238-7209
-    affiliation: 1
+  - name: Gabriel Gonzalez-Escamilla
+    orcid: 0000-0002-7209-1736
+    affiliation: 5    
   - name: Claus Svarer
     orcid: 0000-0001-7811-1825
     affiliation: 2
+  - name: Chris Rorden
+    orcid: 0000-0002-7554-6142
+    affiliation: 6
+  - name: Granville J. Matheson
+    orcid: 0000-0002-5646-4547
+    affiliation: "7, 8, 9"
   - name: Gitte M. Knudsen
     orcid: 0000-0003-1508-6866
     affiliation: 2
+  - name: Robert B. Innis
+    orcid: 0000-0003-1238-7209
+    affiliation: 1
   - name: Melanie Ganz
     orcid: 0000-0002-9120-8098
     affiliation: "2, 3"
-  - name: Gabriel Gonzalez-Escamilla
-    orcid: 0000-0002-7209-1736
-    affiliation: 5 
   - name: Cyrus Eierud
     orcid: 0000-0002-9942-676X
-    affiliation: 6
-  - name: Chris Rorden
-    orcid: 0000-0002-7554-6142
-    affiliation: 7
-  - name: Granville J. Matheson
-    orcid: 0000-0002-5646-4547
-    affiliation: "8, 9"
+    affiliation: 10
   - name: Murat Bilgel
     orcid: 0000-0001-5042-7422
-    affiliation: 10
+    affiliation: 11
   - name: Cyril Pernet
     orcid: 0000-0003-4010-4632
     affiliation: 2
@@ -57,17 +57,19 @@ affiliations:
     index: 4
   - name: University Medical Center of the Johannes Gutenberg University Mainz, Mainz, Germany
     index: 5
-  - name: TReNDS Center, Georgia State University, Atlanta, GA, USA
-    index: 6
   - name: Department of Psychology, University of South Carolina, Columbia, SC, USA
-    index: 7
+    index: 6
   - name: Mailman school of Public Health, Columbia University, New York, NY, USA
-    index: 8
+    index: 7
   - name: Department of Clinical Neuroscience, Karolinska Institutet and Stockholm County Council, Stockholm, Sweden
+    index: 8
+  - name: Karolinska Institutet
     index: 9
-  - name: National Institute on Aging Intramural Research Program, Baltimore, MD, USA
+  - name: TReNDS Center, Georgia State University, Atlanta, GA, USA
     index: 10
-date: 17 August 2024
+  - name: National Institute on Aging Intramural Research Program, Baltimore, MD, USA
+    index: 11
+date: 26 June 2024
 bibliography: paper.bib
 
 ---
