### 2024-06-26 #6067 @britta-wstnr
@bendhouseart, I opened a PR over at your repo (see link above), where I attempted to match the author list and affiliation order to what you have in your Zenodo archive. That is equivalent to the author order in `fe0faa1`. The affiliation order is different to both your commits but should match Zenodo now, I hope.
Please double check everything - I will look at the rendered paper again once it's up (i.e. after you merged) to make sure it renders as I hope it will based on my changes 🙂 🤞
