### 2024-07-18 c6c6df5
added telemetry call
