### 2024-07-19 #6067 @britta-wstnr
Hi @CPernet - sorry for the delay, I was first at a conference and then sick.
Paper looks good now, but can you please change the archive title to match the full paper title?
And can you confirm the version of the software, is it still `v1.3.20240502` as per [this comment](https://github.com/openjournals/joss-reviews/issues/6067#issuecomment-2091000496)?
