### 2024-07-19 5a84ebb
added file count to telemetery tracking for dcm2niix
