### 2024-08-01 b3bf9af
matlab is being weird and not finding dcm2niix on path
