+++ matlab/dcm2niix4pet.m
@@ -54,7 +54,7 @@ if ispc && ~exist(dcm2niixpath,'file')
 end
 
 if ~ispc % overwrite if not windowns (as it should be in the computer path)
-    dcm2niixpath = '/usr/local/bin/dcm2niix';
+    dcm2niixpath = 'dcm2niix';
 end
 
 % we rely on more recent version of dcm2niix, certain pet fields are unavailable in the sidecar jsons for versions
@@ -62,16 +62,16 @@ end
 
 minimum_version = 'v1.0.20220720';
 minimum_version_date = datetime(minimum_version(6:end), 'InputFormat', 'yyyyMMdd');
-%version_cmd = ['dcm2niix', ' -v']; % TODO fix this as it's not detecting dcm2niix running matlab on osx as of matlab v2023b.
-version_cmd = ['/usr/local/bin/dcm2niix', ' -v'];
+version_cmd = ['dcm2niix', ' -v'];
 
 [status, version_output_string] = system(version_cmd);
-version = regexp(version_output_string, 'v[0-9].[0-9].{8}[0-9]', 'match'); % TODO this returns an array with two of the same versions
+version = regexp(version_output_string, 'v[0-9].[0-9].{8}[0-9]', 'match');
+
 
 % initalize telemetry data fror later uploading
 telemetry_data = {};
 dcm2niix_data = {};
-dcm2niix_data.version = version;
+dcm2niix_data.version = version(1);
 dcm2niix_data.returncode = 0;
 telemetry_data.dcm2niix = dcm2niix_data;
 telemetry_data.description = "Matlab_dcm2niix4pet.m"
