### 2024-08-07 62791ec
added option to opt out of telemetry
