+++ pypet2bids/pypet2bids/telemetry.py
@@ -2,15 +2,65 @@ import requests
 import os
 import pathlib
 import subprocess
+import time
+import sys
+import select
 from dotenv import load_dotenv
 from typing import Union
 
 
-# collect pet2bids version
-from pypet2bids.helper_functions import get_version
+try:
+    from helper_functions import get_version, load_vars_from_config, modify_config_file
+except ModuleNotFoundError:
+    # collect pet2bids version
+    from pypet2bids.helper_functions import get_version, load_vars_from_config, modify_config_file
+
+pet2bids_config = load_vars_from_config()
+telemetry_default_url = pet2bids_config.get("TELEMETRY_URL", "http://52.87.154.236/telemetry/")
+telemetry_enabled = pet2bids_config.get("TELEMETRY_ENABLED", True)
+telemetry_notify_user = pet2bids_config.get("NOTIFY_USER_OF_TELEMETRY", False)
+
+
+def ask_user_to_opt_out(timeout=10):
+    """Waits for user input for a specified number of seconds.
+
+    Args:
+        timeout: Number of seconds to wait for input.
+
+    Returns:
+        The user input if provided within the timeout, otherwise None.
+    """
+
+    print("Do you want to opt out of telemetry? (yes/no): ".format(timeout), end="", flush=True)
+    sys.stdout.flush()
+
+    ready, _, _ = select.select([sys.stdin], [], [], timeout)
+    if ready:
+        response = sys.stdin.readline().rstrip('\n')
+        if response.lower() == "yes":
+            return True
+    else:
+        return False
+
+
+if telemetry_notify_user is False:
+    opt_out = ask_user_to_opt_out()
+    if opt_out:
+        # update the config file
+        modify_config_file("TELEMETRY_ENABLED", False)
+
+    modify_config_file("NOTIFY_USER_OF_TELEMETRY", True)
 
 
 def telemetry_enabled(config_path=None):
+    """
+    Check if telemetry is enabled, if it isn't disabled in the .pet2bidsconfig file
+    it will be considered enabled. One must opt out of tracking usage manually.
+    :param config_path: The path to the config file
+    :type config_path: Union[str, pathlib.Path]
+    :return: Whether telemetry is enabled or not
+    :rtype: bool
+    """
     # load dcm2niix file
     if not config_path:
         config_file = pathlib.Path.home() / ".pet2bidsconfig"
@@ -27,7 +77,17 @@ def telemetry_enabled(config_path=None):
         return True
 
 
-def send_telemetry(json_data: dict, url: str = "http://52.87.154.236/telemetry/"):
+def send_telemetry(json_data: dict, url: str = telemetry_default_url):
+    """
+    Send telemetry data to the telemetry server, by default this will first try
+    to load the telemetry server url from the config file, if it's not found it will
+    default to the hardcoded value in this module. This will always send, unless a user
+    has disabled the pet2bids telemetry in the .pet2bidsconfig file.
+    :param json_data: The dat to be sent to the telemetry server
+    :type json_data: dict
+    :param url: The url of the telemetry server
+    :type url: str
+    """
     if telemetry_enabled():
         # update data with version of pet2bids
         json_data["pypet2bids_version"] = get_version()
@@ -47,7 +107,13 @@ def send_telemetry(json_data: dict, url: str = "http://52.87.154.236/telemetry/"
 
 
 def count_input_files(input_file_path: Union[str, pathlib.Path]):
-    # Count the number of input files in the input file path
+    """
+    Count the number of input files in the input file path
+    :param input_file_path: location of the input files
+    :type input_file_path: Union[str, pathlib.Path]
+    :return: The number of input files and their size
+    :rtype: dict
+    """
     if isinstance(input_file_path, str):
         input_file_path = pathlib.Path(input_file_path)
 
[…]
