### 2024-08-08 #6067 @britta-wstnr
@bendhouseart @CPernet while checking the references again, I saw that the link for `dcm2nii` http://www.mccauslandcenter.sc.edu/mricro/mricron/dcm2nii.html now redirects to the very generic https://sc.edu/study/colleges_schools/artsandsciences/centers_and_institutes/mccausland_center/index.php which is not very helpful in the context of the references.
You might want to fix that?
