+++ paper.bib
@@ -179,7 +179,7 @@ Publisher: Nature Publishing Group},
 
 @software{rorden_dcm2nii,
 	title = {dcm2nii},
-	url = {http://www.mccauslandcenter.sc.edu/mricro/mricron/dcm2nii.html},
+	url = {https://github.com/rordenlab/dcm2niix},
 	author = {Rorden, C.},
 	date = {2023},
 }
