### 2024-08-12 #6067 @CPernet
thx @britta-wstnr indeed, now using the GitHub repo URL instead https://github.com/openneuropet/PET2BIDS/commit/3d431dbf708ac22fcc5912461c6121defc3bcfcb
