### 2024-08-12 e60aa57
added cli flag and environment variable disabling of telemetry
