### 2024-08-12 ba2557a
Bump Pandas Version  (#317)
* bumped pandas to a minimum version not a singular one
* fix spelling errors to appease codespell
