### 2024-08-12 400fec2
update tracking on matlab to set env to not track if provided via arguments
