+++ matlab/ecat2nii.m
@@ -115,8 +115,7 @@ for v=1:length(varargin)
     elseif strcmpi(varargin{v},'savemat')
         savemat = varargin{v+1};
     elseif strcmpi(varargin{v},'notrack')
-        notrack = varargin{v+1};
-        if notrack; setenv('TELEMETRY_ENABLED', 'False'); end
+        setenv('TELEMETRY_ENABLED', 'False')
     end
 end
 
