### 2024-08-14 4b4928a
fix error with loading config
