### 2024-08-15 878d4ac
add try catch for .env file check on telemetry.m
