+++ pypet2bids/pypet2bids/helper_functions.py
@@ -1149,3 +1149,37 @@ def first_middle_last_frames_to_text(
             delimiter="\t",
             fmt="%s",
         )
+
+def reorder_isotope(isotope: str) -> str:
+    """
+    Reorders the isotope string to be in the format of "isotope""element name"
+    :param isotope: isotope string
+    :type isotope: str
+    :return: reordered isotope string
+    :rtype: str
+    """
+    # remove all non-alphanumeric characters from isotope
+    isotope = re.findall(r"[a-zA-Z0-9]+", isotope)
+    # combine all elements in isotope into one string
+    isotope = "".join(isotope)
+    # collect the isotope number from the isotope string
+    isotope_num = re.findall(r"\d+", isotope)
+    # collect the element name from the isotope string
+    element_name = re.findall(r"[a-zA-Z]+", isotope)
+
+    # capitalize the first letter of the element name if the element name's length is <= 2
+    if 1 < len(element_name[0]) <= 2:
+        e = element_name[0][0].capitalize()
+        if len(element_name[0]) == 2:
+            e += element_name[0][1].lower()
+    elif len(element_name[0]) == 1:
+        e = element_name[0].capitalize()
+
+    # special case for 52mMn
+    if "".join(element_name).lower() == "mmn" or "".join(element_name).lower() == "mnm":
+        e = "mMn"
+
+    # reorder to put the number before the element name
+    isotope = f"{isotope_num[0]}{e}"
+
+    return isotope
