### 2024-08-20 #6067 @britta-wstnr
@CPernet great, thanks!
I just checked the paper again and things look good to me. Double-checked the missing DOIs again, and they indeed seem to be "legit", no DOI given.

Time to recommend acception of the paper -  from here, the EiC will take over.
Thanks again for your review work, @nbeliy @adswa and @pjtoussaint ! We value and appreciate your work ❤️

@CPernet and @bendhouseart et al., if you are not signed up yet, please consider to sign up as a reviewer for JOSS: https://reviewers.joss.theoj.org/join . You can of course also recommend this to your co-authors 😉

Thanks everyone for working on this!
