+++ README.md
@@ -1,0 +1,3 @@
+[![DOI](https://joss.theoj.org/papers/10.21105/joss.06067/status.svg)](https://doi.org/10.21105/joss.06067)
+
+Galassi, A., Norgaard, M., Gonzalez-Escamilla, G., Svarer, C., Rorden, C., Matheson, G.J:, Knudsen, G.M, Innis, R.B., Ganz, M., Eierud, C., Bilgel., M. & Pernet, C. (2024). PET2BIDS: a library for converting Positron Emission Tomography data to BIDS. Journal of Open Source Software, 9(100), 6067, https://doi.org/10.21105/joss.06067
